Av1CondContIC(ROptRegTS)	R Documentation

Generating function for Av1CondContIC-class

Description

Generates an object of class "Av1CondContIC"; i.e., an influence curves eta of the form

eta = (A Lambda - a)min(1, b/|A Lambda - a|)

with clipping bound b, centering function a and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable parametric family which can be created via CallL2Fam.

Usage


Av1CondContIC(name, CallL2Fam = call("L2RegTypeFamily"),
       Curve = EuclRandVarList(RealRandVariable(Map = list(function(x){x[1]*x[2]}), 
                                                Domain = EuclideanSpace(dimension = 2))), 
       Risks, Infos, clip = Inf, stand = as.matrix(1), 
       cent = EuclRandVarList(RealRandVariable(Map = list(function(x){numeric(length(x))}),
                                               Domain = EuclideanSpace(dimension = 2))), 
       lowerCase = NULL, neighborRadius = 0)

Arguments

name 	 object of class "character".

CallL2Fam 	 object of class "call": creates an object of the underlying L2-differentiable regression type family.

Curve 	 object of class "EuclRandVarList"

Risks 	 object of class "list": list of risks; cf. RiskType-class.

Infos 	 matrix of characters with two columns named method and message: additional informations.

clip 	 positive real: clipping bound.

cent 	 object of class "EuclRandVarList": centering function.

stand 	 matrix: standardizing matrix.

lowerCase 	 optional constant for lower case solution.

neighborRadius 	 radius of the corresponding (unconditional) contamination neighborhood.

Value

Object of class "Av1CondContIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class, Av1CondContIC-class

Examples


IC1 <- Av1CondContIC()
IC1

[Package ROptRegTS version 0.6.1 Index]