Av1CondTotalVarIC(ROptRegTS)	R Documentation

Generating function for Av1CondTotalVarIC-class

Description

Generates an object of class "Av1CondContIC"; i.e., an influence curves eta of the form

eta = A x Lambda_f min(1, max(c(x)/(|Ax|Lambda_f), (c(x) + b)/(|Ax|Lambda_f)))

with lower clipping function c, standardized bias b and standardizing matrix A. Lambda_f stands for the L2 derivative of the corresponding error distribution.

Usage


Av1CondTotalVarIC(name, CallL2Fam = call("L2RegTypeFamily"), 
        Curve = EuclRandVarList(RealRandVariable(Map = list(function(x) {x[1] * x[2]}), 
                                                 Domain = EuclideanSpace(dimension = 1))), 
        Risks, Infos, clipUp = Inf, stand = as.matrix(1), 
        clipLo = RealRandVariable(Map = list(function(x) {-Inf}, 
                                  Domain = EuclideanSpace(dimension = 1)), 
        Domain = EuclideanSpace(dimension = 2)), lowerCase = NULL, neighborRadius = 0)

Arguments

name 	 object of class "character".

CallL2Fam 	 object of class "call": creates an object of the underlying L2-differentiable regression type family.

Curve 	 object of class "EuclRandVarList"

Risks 	 object of class "list": list of risks; cf. RiskType-class.

Infos 	 matrix of characters with two columns named method and message: additional informations.

clipUp 	 positive real: standardized bias.

clipLo 	 object of class "RealRandVariable": lower clipping function.

stand 	 matrix: standardizing matrix.

lowerCase 	 optional constant for lower case solution.

neighborRadius 	 radius of the corresponding (unconditional) contamination neighborhood.

Value

Object of class "Av1CondTotalVarIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

,

Examples


IC1 <- Av1CondTotalVarIC()
IC1

[Package ROptRegTS version 0.6.1 Index]