Av1CondTotalVarIC-class(ROptRegTS)	R Documentation

Conditionally centered influence curve of total variaton type

Description

Class of conditionally centered (partial) influence curves of contamination type for average conditional total variation neighborhoods; i.e., influence curves eta of the form

eta = A x Lambda_f min(1, max(c(x)/(|Ax|Lambda_f), (c(x) + b)/(|Ax|Lambda_f)))

with lower clipping function c, standardized bias b and standardizing matrix A. Lambda_f stands for the L2 derivative of the corresponding error distribution.

Objects from the Class

Objects can be created by calls of the form new("Av1CondTotalVarIC", ...). More frequently they are created via the generating function Av1CondTotalVarIC, respectively via the method generateIC.

Slots

CallL2Fam:

object of class "call": creates an object of the underlying L2-differentiable regression type family.

name:

object of class "character"

Curve:

object of class "EuclRandVarList"

Risks:

object of class "list": list of risks; cf. RiskType-class.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

clipUp:

object of class "numeric": standardized bias.

clipLo:

object of class "RealRandVariable": lower clipping function.

stand:

object of class "matrix": standardizing matrix.

lowerCase:

object of class "OptionalNumeric": optional constant for lower case solution.

neighborRadius:

object of class "numeric": radius of the corresponding average conditional contamination neighborhood.

Extends

Class "CondIC", directly. Class "IC", by class "CondIC". Class "InfluenceCurve", by class "CondIC".

Methods

CallL2Fam<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot CallL2Fam.

clipLo

signature(object = "Av1CondTotalVarIC"): accessor function for slot clipLo.

clipLo<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot clipLo.

clipUp

signature(object = "Av1CondTotalVarIC"): accessor function for slot clipUp.

clipUp<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot clipUp.

stand

signature(object = "Av1CondTotalVarIC"): accessor function for slot stand.

stand<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot stand.

lowerCase

signature(object = "Av1CondTotalVarIC"): accessor function for slot lowerCase.

lowerCase<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot lowerCase.

neighborRadius

signature(object = "Av1CondTotalVarIC"): accessor function for slot neighborRadius.

neighborRadius<-

signature(object = "Av1CondTotalVarIC"): replacement function for slot neighborRadius.

generateIC

signature(neighbor = "Av1CondTotalVarNeighborhood", L2Fam = "L2RegTypeFamily"): generate an object of class "Av1CondTotalVarIC". Rarely called directly.

show

signature(object = "Av1CondTotalVarIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class, Av1CondTotalVarIC

Examples


IC1 <- new("Av1CondTotalVarIC")
IC1

[Package ROptRegTS version 0.6.1 Index]