Av2CondContIC(ROptRegTS)	R Documentation

Generating function for Av2CondContIC-class

Description

Generates an object of class "Av2CondContIC"; i.e., an influence curves eta of the form

eta = A K^(-1) x(Lambda_f - z)min(1, c/|Lambda_f - a|)

with K=E x t(x), clipping bound c, centering constant z and standardizing constant A. Lambda_f stands for the L2 derivative of the corresponding error distribution.

Usage


Av2CondContIC(name, CallL2Fam = call("L2RegTypeFamily"), 
        Curve = EuclRandVarList(RealRandVariable(Map = list(function(x) {x[1] * x[2]}),    
                                                 Domain = EuclideanSpace(dimension = 2))), 
        Risks, Infos, clip = Inf, stand = 1, cent = 0, lowerCase = NULL, neighborRadius = 0)

Arguments

name 	 object of class "character".

CallL2Fam 	 object of class "call": creates an object of the underlying L2-differentiable regression type family.

Curve 	 object of class "EuclRandVarList"

Risks 	 object of class "list": list of risks; cf. RiskType-class.

Infos 	 matrix of characters with two columns named method and message: additional informations.

clip 	 positive real: clipping bound.

cent 	 real: centering constant

stand 	 real: standardizing constant

lowerCase 	 optional constant for lower case solution.

neighborRadius 	 radius of the corresponding (unconditional) contamination neighborhood.

Value

Object of class "Av2CondContIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class, Av2CondContIC-class

Examples


IC1 <- Av2CondContIC()
IC1

[Package ROptRegTS version 0.6.1 Index]