Av2CondContIC-class(ROptRegTS)	R Documentation

Conditionally centered influence curve of contamination type

Description

Class of conditionally centered (partial) influence curves of contamination type for average square conditional contamination neighborhoods; i.e., influence curves eta of the form

eta = A K^(-1) x(Lambda_f - z)min(1, c/|Lambda_f - a|)

with K=E x t(x), clipping bound c, centering constant z and standardizing constant A. Lambda_f stands for the L2 derivative of the corresponding error distribution.

Objects from the Class

Objects can be created by calls of the form new("Av2CondContIC", ...). More frequently they are created via the generating function Av2CondContIC, respectively via the method generateIC.

Slots

CallL2Fam:

object of class "call": creates an object of the underlying L2-differentiable regression type family.

name:

object of class "character"

Curve:

object of class "EuclRandVarList"

Risks:

object of class "list": list of risks; cf. RiskType-class.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

clip:

object of class "numeric": clipping bound.

cent:

object of class "numeric": centering constant.

stand:

object of class "numeric": standardizing constant.

lowerCase:

object of class "OptionalNumeric": optional constant for lower case solution.

neighborRadius:

object of class "numeric": radius of the corresponding average conditional contamination neighborhood.

Extends

Class "CondIC", directly. Class "IC", by class "CondIC". Class "InfluenceCurve", by class "CondIC".

Methods

CallL2Fam<-

signature(object = "Av2CondContIC"): replacement function for slot CallL2Fam.

cent

signature(object = "Av2CondContIC"): accessor function for slot cent.

cent<-

signature(object = "Av2CondContIC"): replacement function for slot cent.

clip

signature(object = "Av2CondContIC"): accessor function for slot clip.

clip<-

signature(object = "Av2CondContIC"): replacement function for slot clip.

stand

signature(object = "Av2CondContIC"): accessor function for slot stand.

stand<-

signature(object = "Av2CondContIC"): replacement function for slot stand.

lowerCase

signature(object = "Av2CondContIC"): accessor function for slot lowerCase.

lowerCase<-

signature(object = "Av2CondContIC"): replacement function for slot lowerCase.

neighborRadius

signature(object = "Av2CondContIC"): accessor function for slot neighborRadius.

neighborRadius<-

signature(object = "Av2CondContIC"): replacement function for slot neighborRadius.

generateIC

signature(neighbor = "Av2CondContNeighborhood", L2Fam = "L2RegTypeFamily"): generate an object of class "Av2CondContIC". Rarely called directly.

show

signature(object = "Av2CondContIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class, Av2CondContIC

Examples


IC1 <- new("Av2CondContIC")
IC1

[Package ROptRegTS version 0.6.1 Index]