Av2CondContNeighborhood-class(ROptRegTS)	R Documentation

Average square conditional contamination neighborhood

Description

Class of average square conditional contamination neighborhoods (exponent == 2); i.e. only radius curves varepsilon with ||varepsilon||_2 <= 1.

Objects from the Class

Objects can be created by calls of the form new("Av2CondContNeighborhood", ...). More frequently they are created via the generating function Av2CondContNeighborhood.

Slots

type:

Object of class "character": “average square conditional convex contamination neighborhood”.

radius:

Object of class "numeric": neighborhood radius.

radiusCurve:

Object of class "function": radius curve with L2 norm <= 1.

exponent:

equal to 2.

Extends

Class "Av2CondNeighborhood", directly.
Class "AvCondNeighborhood", by class "Av2CondNeighborhood".
Class "CondNeighborhood", by class "Av2CondNeighborhood".
Class "Neighborhood", by class "Av2CondNeighborhood".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

Examples


new("Av2CondContNeighborhood")

[Package ROptRegTS version 0.6.1 Index]