CondContIC-class(ROptRegTS)	R Documentation

Conditionally centered influence curve of contamination type

Description

Class of conditionally centered (partial) influence curves of contamination type for conditional contamination neighborhoods; i.e., influence curves eta of the form

eta = (A Lambda - a)min(1, b/|A Lambda - a|)

with clipping function b, centering function a and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable regression type family created via the call in the slot CallL2Fam.

Objects from the Class

Objects can be created by calls of the form new("CondContIC", ...). More frequently they are created via the generating function CondContIC, respectively via the method generateIC.

Slots

CallL2Fam:

object of class "call": creates an object of the underlying L2-differentiable regression type family.

name:

object of class "character"

Curve:

object of class "EuclRandVarList"

Risks:

object of class "list": list of risks; cf. RiskType-class.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

clip:

object of class "RealRandVariable": clipping function.

cent:

object of class "EuclRandVarList": centering function.

stand:

object of class "matrix": standardizing matrix.

lowerCase:

object of class "OptionalNumeric": optional constant for lower case solution.

neighborRadius:

object of class "numeric": radius of the corresponding conditional contamination neighborhood.

neighborRadiusCurve:

object of class "function": radius curve of the corresponding conditional contamination neighborhood.

Extends

Class "CondIC", directly.
Class "IC", by class "CondIC".
Class "InfluenceCurve", by class "CondIC".

Methods

CallL2Fam<-

signature(object = "CondContIC"): replacement function for slot CallL2Fam.

cent

signature(object = "CondContIC"): accessor function for slot cent.

cent<-

signature(object = "CondContIC"): replacement function for slot cent.

clip

signature(object = "CondContIC"): accessor function for slot clip.

clip<-

signature(object = "CondContIC"): replacement function for slot clip.

stand

signature(object = "CondContIC"): accessor function for slot stand.

stand<-

signature(object = "CondContIC"): replacement function for slot stand.

lowerCase

signature(object = "CondContIC"): accessor function for slot lowerCase.

lowerCase<-

signature(object = "CondContIC"): replacement function for slot lowerCase.

neighborRadius

signature(object = "CondContIC"): accessor function for slot neighborRadius.

neighborRadius<-

signature(object = "CondContIC"): replacement function for slot neighborRadius.

neighborRadiusCurve

signature(object = "CondContIC"): accessor function for slot neighborRadiusCurve.

neighborRadiusCurve<-

signature(object = "CondContIC"): replacement function for slot neighborRadiusCurve.

generateIC

signature(neighbor = "CondContNeighborhood", L2Fam = "L2RegTypeFamily"): generate an object of class "CondContIC". Rarely called directly.

show

signature(object = "CondContIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class, CondContIC

Examples


IC1 <- new("CondContIC")
IC1

[Package ROptRegTS version 0.6.1 Index]