CondIC-class(ROptRegTS)	R Documentation

Conditionally centered partial influence curve

Description

Class of conditionally centered partial influence curves.

Objects from the Class

Objects can be created by calls of the form new("CondIC", ...). More frequently they are created via the generating function CondIC.

Slots

CallL2Fam:

Object of class "call": creates an object of class "L2RegTypeFamily".

name:

Object of class "character": name

Curve:

Object of class "EuclRandVariable": curve.

Risks:

Object of class "list": list of risks; cf. RiskType-class.

Infos:

Object of class "matrix" with two columns named method and message: additional informations.

Extends

Class "IC", directly.
Class "InfluenceCurve", by class "IC".

Methods

CallL2Fam<-

signature(object = "IC"): replacement function for slot CallL2Fam.

checkIC

signature(IC = "CondIC", L2Fam = "missing"): check conditional centering and Fisher consistency of IC assuming the L2-differentiable regression-type family which can be created via the slot CallL2Fam of IC.

checkIC

signature(IC = "CondIC", L2Fam = "L2RegTypeFamily"): check conditional centering and Fisher consistency of IC assuming the L2-differentiable regression-type family L2Fam.

show

signature(object = "CondIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, IC-class

Examples


new("CondIC")

[Package ROptRegTS version 0.6.1 Index]