CondTotalVarNeighborhood-class(ROptRegTS)	R Documentation

Conditional total variation neighborhood

Description

Class of conditional (error-free-variables) total variation neighborhoods.

Objects from the Class

Objects can be created by calls of the form new("CondTotalVarNeighborhood", ...). More frequently they are created via the generating function CondTotalVarNeighborhood.

Slots

type:

Object of class "character": “conditional total variation neighborhood”.

radius:

Object of class "numeric": neighborhood radius.

radiusCurve:

Object of class "function": radius curve

Extends

Class "CondNeighborhood", directly.
Class "Neighborhood", by class "CondNeighborhood".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

,

Examples


new("CondTotalVarNeighborhood")

[Package ROptRegTS version 0.6.1 Index]