ContIC(RobAStBase)	R Documentation

Generating function for ContIC-class

Description

Generates an object of class "ContIC"; i.e., an influence curves eta of the form

eta = (A Lambda - a)min(1, b/|A Lambda - a|)

with clipping bound b, centering constant a and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable parametric family which can be created via CallL2Fam.

Usage


ContIC(name, CallL2Fam = call("L2ParamFamily"), 
       Curve = EuclRandVarList(RealRandVariable(Map = c(function(x){x}), 
                                                Domain = Reals())), 
       Risks, Infos, clip = Inf, cent = 0, stand = as.matrix(1), 
       lowerCase = NULL, neighborRadius = 0, w = new("HampelWeight"),
       normtype = NormType(), biastype = symmetricBias())

Arguments

name 	 object of class "character".

CallL2Fam 	 object of class "call": creates an object of the underlying L2-differentiable parametric family.

Curve 	 object of class "EuclRandVarList"

Risks 	 object of class "list": list of risks; cf. RiskType-class.

Infos 	 matrix of characters with two columns named method and message: additional informations.

clip 	 positive real: clipping bound.

cent 	 real: centering constant

stand 	 matrix: standardizing matrix

w 	 HampelWeight: weight object

lowerCase 	 optional constant for lower case solution.

neighborRadius 	 radius of the corresponding (unconditional) contamination neighborhood.

biastype 	 BiasType: type of the bias

normtype 	 NormType: type of the norm

Value

Object of class "ContIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class, ContIC

Examples


IC1 <- ContIC()
plot(IC1)

[Package RobAStBase version 0.1.0 Index]