ContIC-class(RobAStBase)	R Documentation

Influence curve of contamination type

Description

Class of (partial) influence curves of contamination type; i.e., influence curves eta of the form

eta = (A Lambda - a)min(1, b/|A Lambda - a|)

with clipping bound b, centering constant a and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable parametric family created via the call in the slot CallL2Fam.

Objects from the Class

Objects can be created by calls of the form new("ContIC", ...). More frequently they are created via the generating function ContIC, respectively via the method generateIC.

Slots

CallL2Fam:

object of class "call": creates an object of the underlying L2-differentiable parametric family.

name:

object of class "character"

Curve:

object of class "EuclRandVarList"

modifyIC:

Object of class "OptionalFunction": function of two arguments, which are an L2 parametric family and an optional influence curve. Returns an object of class "IC". This slot is mainly used for internal computations!

Risks:

object of class "list": list of risks; cf. RiskType-class.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

clip:

object of class "numeric": clipping bound.

cent:

object of class "numeric": centering constant.

stand:

object of class "matrix": standardizing matrix.

weight:

object of class "HampelWeight": weight function

biastype:

object of class "BiasType": bias type (symmetric/onsided/asymmetric)

normtype:

object of class "NormType": norm type (Euclidean, information/self-standardized)

lowerCase:

object of class "OptionalNumeric": optional constant for lower case solution.

neighborRadius:

object of class "numeric": radius of the corresponding (unconditional) contamination neighborhood.

Extends

Class "HampIC", directly.
Class "IC", by class "HampIC".
Class "InfluenceCurve", by class "IC".

Methods

CallL2Fam<-

signature(object = "ContIC"): replacement function for slot CallL2Fam.

cent

signature(object = "ContIC"): accessor function for slot cent.

cent<-

signature(object = "ContIC"): replacement function for slot cent.

clip

signature(x1 = "ContIC"): accessor function for slot clip.

clip<-

signature(object = "ContIC"): replacement function for slot clip.

stand<-

signature(object = "ContIC"): replacement function for slot stand.

lowerCase<-

signature(object = "ContIC"): replacement function for slot lowerCase.

neighbor

signature(object = "ContIC"): generates an object of class "ContNeighborhood" with radius given in slot neighborRadius.

generateIC

signature(neighbor = "ContNeighborhood", L2Fam = "L2ParamFamily"): generate an object of class "ContIC". Rarely called directly.

show

signature(object = "ContIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

,

Examples


IC1 <- new("ContIC")
plot(IC1)

[Package RobAStBase version 0.7 Index]