DistrSymmList-class(ROptEstOld)	R Documentation

List of Symmetries for a List of Distributions

Description

Create a list of symmetries for a list of distributions

Objects from the Class

Objects can be created by calls of the form new("DistrSymmList", ...). More frequently they are created via the generating function DistrSymmList.

Slots

.Data:

Object of class "list". A list of objects of class "DistributionSymmetry".

Extends

Class "list", from data part.
Class "vector", by class "list".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

DistributionSymmetry-class

Examples


new("DistrSymmList", list(NoSymmetry(), SphericalSymmetry(SymmCenter = 1), 
                          EllipticalSymmetry(SymmCenter = 2)))

[Package ROptEstOld version 0.5.2 Index]