EllipticalSymmetry-class(ROptEstOld)	R Documentation

Class for Elliptically Symmetric Distributions

Description

Class for elliptically symmetric distributions.

Objects from the Class

Objects can be created by calls of the form new("EllipticalSymmetry"). More frequently they are created via the generating function EllipticalSymmetry. Elliptical symmetry for instance leads to a simplification for the computation of optimally robust influence curves.

Slots

type:

Object of class "character": contains “elliptical symmetric distribution”

SymmCenter:

Object of class "numeric": center of symmetry

Extends

Class "DistributionSymmetry", directly.
Class "Symmetry", by class "DistributionSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

EllipticalSymmetry, DistributionSymmetry-class

Examples


new("EllipticalSymmetry")

[Package ROptEstOld version 0.5.2 Index]