EvenSymmetric-class(ROptEst)	R Documentation

Class for Even Functions

Description

Class for even functions.

Objects from the Class

Objects can be created by calls of the form new("EvenSymmetric"). More frequently they are created via the generating function EvenSymmetric.

Slots

type:

Object of class "character": contains “even function”

SymmCenter:

Object of class "numeric": center of symmetry

Extends

Class "FunctionSymmetry", directly.
Class "Symmetry", by class "FunctionSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

EvenSymmetric, FunctionSymmetry-class

Examples


new("EvenSymmetric")

[Package ROptEst version 0.5.0 Index]