FixRobModel-class(ROptEstOld)	R Documentation

Robust model with fixed (unconditional) neighborhood

Description

Class of robust models with fixed (unconditional) neighborhoods.

Objects from the Class

Objects can be created by calls of the form new("FixRobModel", ...). More frequently they are created via the generating function FixRobModel.

Slots

center:

Object of class "ProbFamily".

neighbor:

Object of class "UncondNeighborhood".

Extends

Class "RobModel", directly.

Methods

neighbor<-

signature(object = "FixRobModel"): replacement function for slot neighbor<-

show

signature(object = "FixRobModel")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ProbFamily-class, UncondNeighborhood-class, FixRobModel

Examples


new("FixRobModel")

[Package ROptEstOld version 0.5.2 Index]