FixRobRegTypeModel-class(ROptRegTS)	R Documentation

Robust regression-type model with fixed neighborhood

Description

Class of robust regression-type models with fixed (conditional or unconditional) neighborhoods.

Objects from the Class

Objects can be created by calls of the form new("FixRobRegTypeModel", ...). More frequently they are created via the generating function FixRobRegTypeModel.

Slots

center:

Object of class "RegTypeFamily".

neighbor:

Object of class "Neighborhood".

Extends

Class "RobModel", directly.

Methods

neighbor<-

signature(object = "FixRobRegTypeModel") replacement function for slot neighbor.

show

signature(object = "FixRobRegTypeModel")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

RegTypeFamily-class, Neighborhood-class, FixRobRegTypeModel

Examples


new("FixRobRegTypeModel")

[Package ROptRegTS version 0.6.1 Index]