FunSymmList-class(ROptEstOld)	R Documentation

List of Symmetries for a List of Functions

Description

Create a list of symmetries for a list of functions

Objects from the Class

Objects can be created by calls of the form new("FunSymmList", ...). More frequently they are created via the generating function FunSymmList.

Slots

.Data:

Object of class "list". A list of objects of class "FunctionSymmetry".

Extends

Class "list", from data part.
Class "vector", by class "list".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

FunctionSymmetry-class

Examples


new("FunSymmList", list(NonSymmetric(), EvenSymmetric(SymmCenter = 1), 
                        OddSymmetric(SymmCenter = 2)))

[Package ROptEstOld version 0.5.2 Index]