HampelWeight-class(RobAStBase)	R Documentation

Robust Weight classes for weights of Hampel type

Description

Classes for weights of Hampel type.

Objects from the Class

Objects can be created by calls of the form new("HampelWeight", ...); to fill slot weight, you will use the generating functions getweight and minbiasweight.

Slots

name:

Object of class "character"; inherited from class RobWeight.

weight:

Object of class "function" — the weight function; inherited from class RobWeight.

clip:

Object of class "numeric" — clipping bound(s); inherited from class BoundedWeight.

stand:

Object of class "matrix" — standardization; inherited from class BdStWeight.

cent:

Object of class "numeric" — centering.

Extends

Class "RobWeight", via class "BoundedWeight". Class "BoundedWeight", via class "BdStWeight". Class "BdStWeight", directly.

Methods

cent

signature(object = "HampelWeight"): accessor function for slot cent.

cent<-

signature(object = "HampelWeight", value = "matrix"): replacement function for slot cent. This replacement method should be used with great care, as the slot weight is not simultaneously updated and hence, this may lead to inconsistent objects.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BdStWeight-class, BoundedWeight-class, RobWeight-class, IC, InfluenceCurve-class

Examples


## prototype
new("HampelWeight")

[Package RobAStBase version 0.1.0 Index]