IC-class(ROptEstOld)	R Documentation

Influence curve

Description

Class of (partial) influence curves.

Objects from the Class

Objects can be created by calls of the form new("IC", ...). More frequently they are created via the generating function IC.

Slots

CallL2Fam:

Object of class "call": creates an object of the underlying L2-differentiable parametric family.

name:

Object of class "character".

Curve:

Object of class "EuclRandVarList".

Risks:

Object of class "list": list of risks; cf. RiskType-class.

Infos:

Object of class "matrix" with two columns named method and message: additional informations.

Extends

Class "InfluenceCurve", directly.

Methods

CallL2Fam

signature(object = "IC"): accessor function for slot CallL2Fam.

CallL2Fam<-

signature(object = "IC"): replacement function for slot CallL2Fam.

checkIC

signature(IC = "IC", L2Fam = "missing"): check centering and Fisher consistency of IC assuming the L2-differentiable parametric family which can be generated via the slot CallL2Fam of IC.

checkIC

signature(IC = "IC", L2Fam = "L2ParamFamily"): check centering and Fisher consistency of IC assuming the L2-differentiable parametric family L2Fam.

evalIC

signature(IC = "IC", x = "numeric"): evaluate IC at x.

evalIC

signature(IC = "IC", x = "matrix"): evaluate IC at the rows of x.

infoPlot

signature(object = "IC"): Plot absolute and relative information of IC.

plot

signature(x = "IC")

show

signature(object = "IC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, IC

Examples


IC1 <- new("IC")
plot(IC1)

[Package ROptEstOld version 0.5.2 Index]