InfRobModel-class(ROptEst)	R Documentation

Robust model with infinitesimal (unconditional) neighborhood

Description

Class of robust models with infinitesimal (unconditional) neighborhoods; i.e., the neighborhood is shrinking at a rate of sqrt(n).

Objects from the Class

Objects can be created by calls of the form new("InfRobModel", ...). More frequently they are created via the generating function InfRobModel.

Slots

center:

Object of class "ProbFamily".

neighbor:

Object of class "UncondNeighborhood".

Extends

Class "RobModel", directly.

Methods

neighbor<-

signature(object = "InfRobModel"): replacement function for slot neighbor<-

show

signature(object = "InfRobModel")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ProbFamily-class, UncondNeighborhood-class, InfRobModel

Examples


new("InfRobModel")

[Package ROptEst version 0.5.0 Index]