InfluenceCurve-class(RobAStBase)	R Documentation

Influence curve

Description

Class of influence curves (functions).

Objects from the Class

Objects can be created by calls of the form new("InfluenceCurve", ...). More frequently they are created via the generating function InfluenceCurve.

Slots

name:

object of class "character"

Curve:

object of class "EuclRandVarList"

Risks:

object of class "list": list of risks; cf. RiskType-class.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

Methods

name

signature(object = "InfluenceCurve"): accessor function for slot name.

name<-

signature(object = "InfluenceCurve"): replacement function for slot name.

Curve

signature(object = "InfluenceCurve"): accessor function for slot Curve.

Map

signature(object = "InfluenceCurve"): accessor function for slot Map of slot Curve.

Domain

signature(object = "InfluenceCurve"): accessor function for slot Domain of slot Curve.

Range

signature(object = "InfluenceCurve"): accessor function for slot Range of slot Curve.

Infos

signature(object = "InfluenceCurve"): accessor function for slot Infos.

Infos<-

signature(object = "InfluenceCurve"): replacement function for slot Infos.

addInfo<-

signature(object = "InfluenceCurve"): function to add an information to slot Infos.

Risks

signature(object = "InfluenceCurve"): accessor function for slot Risks.

Risks<-

signature(object = "InfluenceCurve"): replacement function for slot Risks.

addRisk<-

signature(object = "InfluenceCurve"): function to add a risk to slot Risks.

show

signature(object = "InfluenceCurve")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve, RiskType-class

Examples


new("InfluenceCurve")

[Package RobAStBase version 0.2 Index]