L2ParamFamily-class(ROptEst)	R Documentation

L2 differentiable parametric family

Description

Class of L2 differentiable parametric families.

Objects from the Class

Objects can be created by calls of the form new("L2ParamFamily", ...). More frequently they are created via the generating function L2ParamFamily.

Slots

name:

object of class "character": name of the family.

distribution:

object of class "Distribution": member of the family.

distrSymm:

Object of class "DistributionSymmetry": symmetry of distribution.

param:

object of class "ParamFamParameter": parameter of the family.

props:

object of class "character": properties of the family.

L2deriv:

object of class "EuclRandVariable": L2 derivative of the family.

L2derivSymm:

object of class "FunSymmList": symmetry of the maps included in L2deriv.

L2derivDistr:

object of class "UnivarDistrList": list which includes the distribution of L2deriv.

L2derivDistrSymm:

object of class "DistrSymmList": symmetry of the distributions included in L2derivDistr.

FisherInfo:

object of class "PosDefSymmMatrix": Fisher information of the family.

Extends

Class "ParamFamily", directly.
Class "ProbFamily", by class "ParamFamily".

Methods

L2deriv

signature(object = "L2ParamFamily"): accessor function for L2deriv.

L2derivSymm

signature(object = "L2ParamFamily"): accessor function for L2derivSymm.

L2derivDistr

signature(object = "L2ParamFamily"): accessor function for L2derivDistr.

L2derivDistrSymm

signature(object = "L2ParamFamily"): accessor function for L2derivDistrSymm.

FisherInfo

signature(object = "L2ParamFamily"): accessor function for FisherInfo.

checkL2deriv

signature(object = "L2ParamFamily"): check centering of L2deriv and compute precision of Fisher information.

E

signature(object = "L2ParamFamily", fun = "EuclRandVariable", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandMatrix", cond = "missing"): expectation of fun under the distribution of object.

E

signature(object = "L2ParamFamily", fun = "EuclRandVarList", cond = "missing"): expectation of fun under the distribution of object.

plot

signature(x = "L2ParamFamily"): plot of distribution and L2deriv.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily, ParamFamily-class

Examples


F1 <- new("L2ParamFamily")
plot(F1)

[Package ROptEst version 0.5.0 Index]