L2RegTypeFamily(ROptRegTS)	R Documentation

Generating function for L2RegTypeFamily-class

Description

Generates an object of class "RegTypeFamily".

Usage


L2RegTypeFamily(name, distribution = LMCondDistribution(), distrSymm, 
        main = 0, nuisance, trafo, param, props = character(0), 
        L2deriv = EuclRandVarList(EuclRandVariable(Map = list(function(x) {x[1] * x[2]}), 
                                                   Domain = EuclideanSpace(dimension = 2), 
                                                   dimension = 1)), 
        ErrorDistr = Norm(), ErrorSymm, RegDistr = Norm(), RegSymm, 
        Regressor = RealRandVariable(Map = list(function(x) {x}), Domain = Reals()), 
        ErrorL2deriv = EuclRandVarList(RealRandVariable(Map = list(function(x) {x}), 
                                                        Domain = Reals())), 
        ErrorL2derivSymm, ErrorL2derivDistr, ErrorL2derivDistrSymm, FisherInfo)

Arguments

name 	 name of the family

distribution 	 conditional distribution (given the regressor)

distrSymm 	 symmetry of distribution

ErrorDistr 	 error distribution

ErrorSymm 	 object of class "DistributionSymmetry": symmetry of ErrorDistr

main 	 main parameter

nuisance 	 optional nuisance parameter

trafo 	 matrix: optional transformation of the parameter

param 	 parameter of the family

props 	 properties of the family

RegDistr 	 regressor distribution

RegSymm 	 object of class "DistributionSymmetry": symmetry of RegDistr

Regressor 	 regressor

L2deriv 	 object of class "EuclRandVariable": L2 derivative

ErrorL2deriv 	 object of class "EuclRandVariable": L2 derivative of ErrorDistr

ErrorL2derivDistr 	 distribution of ErrorL2deriv

ErrorL2derivSymm 	 object of class "FunSymmList": symmetry of ErrorL2deriv

ErrorL2derivDistrSymm 	 object of class "DistrSymmList": symmetry of ErrorL2derivDistr

FisherInfo 	 Fisher information matrix

Details

If name is missing, the default “L2 differentiable regression type family” is used. If param is missing, the parameter is created via main, nuisance and trafo as described in ParamFamParameter. In case distrSymm, ErrorSymm, RegSymm is missing, they are set to NoSymmetry(). If FisherInfo is missing, it is computed via numerical integration.

Value

Object of class "L2RegTypeFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

Examples


L2RegTypeFamily()

[Package ROptRegTS version 0.6.1 Index]