L2RegTypeFamily-class(ROptRegTS)	R Documentation

L2 differentiable parametric regression-type family

Description

Class for L2 differentiable parametric regression-type families.

Objects from the Class

Objects can be created by calls of the form new("L2RegTypeFamily", ...). More frequently they are created via the generating function L2RegTypeFamily.

Slots

L2deriv:

Object of class "EuclRandVarList": L2 derivative.

ErrorL2deriv:

Object of class "EuclRandVarList": L2 derivative of ErrorDistr.

ErrorL2derivSymm:

Object of class "FunSymmList": symmetry of ErrorL2deriv.

ErrorL2derivDistr:

Object of class "DistrList": distribution of ErrorL2deriv.

ErrorL2derivDistrSymm:

Object of class "DistrSymmList": symmetry of ErrorL2derivDistr.

FisherInfo:

Object of class "PosDefSymmMatrix": Fisher information.

ErrorDistr:

Object of class "Distribution": error distribution.

ErrorSymm:

Object of class "DistributionSymmetry": symmetry of ErrorDistr.

RegDistr:

Object of class "Distribution": regressor distribution.

RegSymm:

Object of class "DistributionSymmetry": symmetry of RegDistr.

Regressor:

Object of class "EuclRandVariable": regressor.

param:

Object of class "ParamFamParameter": parameter of the family.

props:

Object of class "character": properties of the family.

name:

Object of class "character": name of the family.

distribution:

Object of class "CondDistribution": conditional distribution given the regressor.

distrSymm:

Object of class "DistributionSymmetry": symmetry of distribution.

Extends

Class "RegTypeFamily", directly.
Class "ParamFamily", by class "RegTypeFamily".
Class "ProbFamily", by class "RegTypeFamily".

Methods

L2deriv

signature(object = "L2RegTypeFamily"): acessor function for slot L2deriv

FisherInfo

signature(object = "L2RegTypeFamily"): acessor function for slot FisherInfo

ErrorL2deriv

signature(object = "L2RegTypeFamily"): acessor function for slot ErrorL2deriv.

ErrorL2derivDistr

signature(object = "L2RegTypeFamily"): acessor function for slot ErrorL2derivDistr.

ErrorL2derivSymm

signature(object = "L2RegTypeFamily"): acessor function for slot ErrorL2derivSymm

ErrorL2derivDistrSymm

signature(object = "L2RegTypeFamily"): acessor function for slot ErrorL2derivDistrSymm

checkL2deriv

signature(object = "L2RegTypeFamily"): check centering of L2deriv and compute precision of Fisher information.

checkIC

signature(IC = "IC", L2Fam = "missing"): check centering and Fisher consistency of IC assuming the L2-differentiable regression-type family which can be created via the slot CallL2Fam of IC.

checkIC

signature(IC = "IC", L2Fam = "L2RegTypeFamily"): check centering and Fisher consistency of IC assuming the L2-differentiable regression-type family L2Fam.

E

signature(object = "L2RegTypeFamily", fun = "EuclRandVariable", cond = "missing"): expectation of fun under object.

E

signature(object = "L2RegTypeFamily", fun = "EuclRandMatrix", cond = "missing"): expectation of fun under object.

E

signature(object = "L2RegTypeFamily", fun = "EuclRandVarList", cond = "missing"): expectation of fun under object.

show

signature(object = "L2RegTypeFamily")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

RegTypeFamily-class

Examples


new("L2RegTypeFamily")

[Package ROptRegTS version 0.6.1 Index]