MEstimate-class(RobAStBase)	R Documentation

MEstimate-class.

Description

Class of asymptotically linear estimates.

Objects from the Class

Objects can be created by calls of the form new("MEstimate", ...). More frequently they are created via the generating function locMEstimator.

Slots

name:

Object of class "character": name of the estimator.

estimate:

Object of class "ANY": estimate.

samplesize:

Object of class "numeric": sample size.

asvar:

Optional object of class "matrix": asymptotic variance.

asbias:

Optional object of class "numeric": asymptotic bias.

pIC:

Optional object of class InfluenceCurve: influence curve.

nuis.idx:

object of class "OptionalNumeric": indices of estimate belonging to the nuisance part.

Mroot:

Object of class "numeric": value of the M equation at the estimate.

Infos:

object of class "matrix" with two columns named method and message: additional informations.

Extends

Class "ALEstimate", directly.
Class "Estimate", by class "ALEstimate".

Methods

Mroot

signature(object = "MEstimate"): accessor function for slot Mroot.

show

signature(object = "MEstimate")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

ALEstimate-class

Examples


## prototype
new("MEstimate")

[Package RobAStBase version 0.1.0 Index]