NoSymmetry-class(ROptEst)	R Documentation

Class for Non-symmetric Distributions

Description

Class for non-symmetric distributions.

Objects from the Class

Objects can be created by calls of the form new("NoSymmetry"). More frequently they are created via the generating function NoSymmetry.

Slots

type:

Object of class "character": contains “non-symmetric distribution”

SymmCenter:

Object of class "NULL"

Extends

Class "DistributionSymmetry", directly.
Class "Symmetry", by class "DistributionSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

NoSymmetry, Distribution-class

Examples


new("NoSymmetry")

[Package ROptEst version 0.5.0 Index]