NonSymmetric-class(ROptEstOld)	R Documentation

Class for Non-symmetric Functions

Description

Class for non-symmetric functions.

Objects from the Class

Objects can be created by calls of the form new("NonSymmetric"). More frequently they are created via the generating function NonSymmetric.

Slots

type:

Object of class "character": contains “non-symmetric function”

SymmCenter:

Object of class "NULL"

Extends

Class "FunctionSymmetry", directly.
Class "Symmetry", by class "FunctionSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

NonSymmetric

Examples


new("NonSymmetric")

[Package ROptEstOld version 0.5.2 Index]