NormLinRegInterceptFamily(ROptRegTS)	R Documentation

Generating Function for Linear Regression Family with Unknown Intercept

Description

Generates an object of class "L2RegTypeFamily" which represents a linear regression family with standard normal distributed errors and random regressor where the intercept is unknown.

Usage


NormLinRegInterceptFamily(theta, intercept = 0, trafo, RegDistr = Norm(), 
                          RegSymm, Reg2Mom, nuisance = FALSE)

Arguments

theta 	 linear regression parameter

intercept 	 intercept parameter

trafo 	 matrix: transformation of the parameter

RegDistr 	 regressor distribution

RegSymm 	 symmetry of the regressor distribution

Reg2Mom 	 second moment matrix of regressor

nuisance 	 logical: is intercept nuisance parameter

Details

In case theta is missing, it is set to 0. If Reg2Mom is missing, it is computed via E.

Value

Object of class "L2RegTypeFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

Examples


(LM1 <- NormLinRegInterceptFamily(Reg2Mom = matrix(1)))
Map(L2deriv(LM1)[[1]])
FisherInfo(LM1)
checkL2deriv(LM1)

[Package ROptRegTS version 0.6.1 Index]