OddSymmetric-class(ROptEst)	R Documentation

Class for Odd Functions

Description

Class for odd functions.

Objects from the Class

Objects can be created by calls of the form new("OddSymmetric"). More frequently they are created via the generating function OddSymmetric.

Slots

type:

Object of class "character": contains “odd function”

SymmCenter:

Object of class "numeric": center of symmetry

Extends

Class "FunctionSymmetry", directly.
Class "Symmetry", by class "FunctionSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

OddSymmetric, FunctionSymmetry-class

Examples


new("OddSymmetric")

[Package ROptEst version 0.5.0 Index]