ParamFamParameter(ROptEstOld)	R Documentation

Generating function for ParamFamParameter-class

Description

Generates an object of class "ParamFamParameter".

Usage


ParamFamParameter(name, main = numeric(0), nuisance, trafo)

Arguments

name 	 character string: name of parameter

main 	 numeric vector: main parameter

nuisance 	 numeric vector: nuisance paramter

trafo 	 matrix: transformation of the parameter

Details

If name is missing, the default “"parameter of a parametric family of probability measures"” is used. If nuisance is missing, the nuisance parameter is set to NULL. The number of columns of trafo have to be equal and the number of rows have to be not larger than the sum of the lengths of main and nuisance. If trafo is missing, no transformation to the parameter is applied; i.e., trafo is set to an identity matrix.

Value

Object of class "ParamFamParameter"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

ParamFamParameter-class

Examples


ParamFamParameter(main = 0, nuisance = 1, trafo = diag(c(1,2)))

## The function is currently defined as
function(name, main = numeric(0), nuisance, trafo){
    if(missing(name))
        name <- "parameter of a parametric family of probability measures"
    if(missing(nuisance))
        nuisance <- NULL
    if(missing(trafo))
        trafo <- diag(length(main)+length(nuisance))

    return(new("ParamFamParameter", name = name, main = main, 
               nuisance = nuisance, trafo = trafo))
}

[Package ROptEstOld version 0.5.2 Index]