ParamFamParameter-class(ROptEst)	R Documentation

Parameter of a parametric family of probability measures

Description

Class of the parameter of parametric families of probability measures.

Objects from the Class

Objects can be created by calls of the form new("ParamFamParameter", ...). More frequently they are created via the generating function ParamFamParameter.

Slots

main:

Object of class "numeric": main parameter.

nuisance:

Object of class "OptionalNumeric": optional nuisance parameter.

trafo:

Object of class "matrix": transformation of the parameter.

name:

Object of class "character": name of the parameter.

Extends

Class "Parameter", directly.
Class "OptionalParameter", by class "Parameter".

Methods

main

signature(object = "ParamFamParameter"): accessor function for slot main.

main<-

signature(object = "ParamFamParameter"): replacement function for slot main.

nuisance

signature(object = "ParamFamParameter"): accessor function for slot nuisance.

nuisance<-

signature(object = "ParamFamParameter"): replacement function for slot nuisance.

trafo

signature(object = "ParamFamParameter"): accessor function for slot trafo.

trafo<-

signature(object = "ParamFamParameter"): replacement function for slot trafo.

length

signature(x = "ParamFamParameter"): sum of the lengths of main and nuisance.

show

signature(object = "ParamFamParameter")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Parameter-class

Examples


new("ParamFamParameter")

[Package ROptEst version 0.5.0 Index]