ParamFamily(ROptEstOld)	R Documentation

Generating function for ParamFamily-class

Description

Generates an object of class "ParamFamily".

Usage


ParamFamily(name, distribution = Norm(), distrSymm, main = 0, 
            nuisance, trafo, param, props = character(0))

Arguments

name 	 character string: name of family

distribution 	 object of class "Distribution": member of the family

distrSymm 	 object of class "DistributionSymmetry": symmetry of distribution.

main 	 numeric vector: main parameter

nuisance 	 numeric vector: nuisance parameter

trafo 	 matrix: transformation of the parameters

param 	 object of class "ParamFamParameter": parameter of the family

props 	 character vector: properties of the family

Details

If name is missing, the default “"parametric family of probability measures"” is used. In case distrSymm is missing it is set to NoSymmetry(). If param is missing, the parameter is created via main, nuisance and trafo as described in ParamFamParameter.

Value

Object of class "ParamFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

ParamFamily-class

Examples


F1 <- ParamFamily()
plot(F1)

## The function is currently defined as
function(name, distribution = Norm(), main = 0, nuisance, 
         trafo, param, props = character(0)){
    if(missing(name)) 
        name <- "parametric family of probability measures"
    if(missing(distrSymm)) distrSymm <- NoSymmetry()
    if(missing(param)) 
        param <- ParamFamParameter(name = paste("parameter of", name), 
                        main = main, nuisance = nuisance, trafo = trafo)
    return(new("ParamFamily", name = name, distribution = distribution, 
               distrSymm = distrSymm, param = param, props = props))
}

[Package ROptEstOld version 0.5.2 Index]