ParamFamily-class(ROptEstOld)	R Documentation

Parametric family of probability measures.

Description

Class of parametric families of probability measures.

Objects from the Class

Objects can be created by calls of the form new("ParamFamily", ...). More frequently they are created via the generating function ParamFamily.

Slots

param:

Object of class "ParamFamParameter": parameter of the family.

name:

Object of class "character": name of the family.

distribution:

Object of class "Distribution": member of the family.

distrSymm:

Object of class "DistributionSymmetry": symmetry of distribution.

props:

Object of class "character": properties of the family.

Extends

Class "ProbFamily", directly.

Methods

main

signature(object = "ParamFamily"): wrapped accessor function for slot main of slot param.

nuisance

signature(object = "ParamFamily"): wrapped accessor function for slot nuisance of slot param.

trafo

signature(object = "ParamFamily"): wrapped accessor function for slot trafo of slot param.

param

signature(object = "ParamFamily"): accessor function for slot param.

plot

signature(x = "ParamFamily"): plot of slot distribution.

show

signature(object = "ParamFamily")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

Distribution-class

Examples


F1 <- new("ParamFamily") # prototype
plot(F1)

[Package ROptEstOld version 0.5.2 Index]