PosDefSymmMatrix-class(ROptEst)	R Documentation

Positive-definite symmetric matrices

Description

The class of positive-definite, symmetric matrices.

Objects from the Class

Objects can be created by calls of the form new("PosDefSymmMatrix", ...). More frequently they are created via the generating function PosDefSymmMatrix.

Slots

.Data:

Object of class "matrix". A numeric matrix with finite entries.

Extends

Class "matrix", from data part.
Class "structure", by class "matrix".
Class "array", by class "matrix".
Class "vector", by class "matrix", with explicit coerce.
Class "vector", by class "matrix", with explicit coerce.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

PosDefSymmMatrix, matrix-class

Examples


new("PosDefSymmMatrix", diag(2))

[Package ROptEst version 0.5.0 Index]