RealRandVariable-class(RandVar)	R Documentation

Real random variable

Description

Class of real random variables.

Objects from the Class

Objects can be created by calls of the form new("RealRandVariable", ...). More frequently they are created via the generating function EuclRandVariable.

Slots

Map:

Object of class "list": list of functions.

Domain:

Object of class "OptionalrSpace": domain of the random variable.

Range:

Object of class "Reals": range of the random variable.

Extends

Class "EuclRandVariable", directly.
Class "RandVariable", by class "EuclRandVariable".

Methods

Range<-

signature(object = "EuclRandVariable"): replacement function for the slot Range.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

EuclRandVariable-class

Examples


new("RealRandVariable", Map=list(function(x){x}), Range = Reals())

[Package RandVar version 0.6.3 Index]