RegTypeFamily(ROptRegTS)	R Documentation

Generating function for RegTypeFamily-class

Description

Generates an object of class "RegTypeFamily".

Usage


RegTypeFamily(name, distribution = LMCondDistribution(), distrSymm, 
        ErrorDistr = Norm(), ErrorSymm, main = 0, nuisance, trafo, 
        param, props = character(0), RegDistr = Norm(), RegSymm, 
        Regressor = RealRandVariable(c(function(x) {x}), Domain = Reals()))

Arguments

name 	 name of the family

distribution 	 conditional distribution (given the regressor)

distrSymm 	 symmetry of distribution

ErrorDistr 	 error distribution

ErrorSymm 	 symmetry of ErrorDistr

main 	 main parameter

nuisance 	 optional nuisance parameter

trafo 	 matrix: optional transformation of the parameter

param 	 parameter of the family

props 	 properties of the family

RegDistr 	 regressor distribution

RegSymm 	 symmetry of RegDistr

Regressor 	 regressor

Details

If name is missing, the default “regression type family” is used. If param is missing, the parameter is created via main, nuisance and trafo as described in ParamFamParameter. In case distrSymm, ErrorSymm or RegSymm is missing, they are set to NoSymmetry().

Value

Object of class "RegTypeFamily"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

ParamFamily-class

Examples


RegTypeFamily()

[Package ROptRegTS version 0.6.1 Index]