RegTypeFamily-class(ROptRegTS)	R Documentation

Parametric regression-type family

Description

Class for parametric regression-type families.

Objects from the Class

Objects can be created by calls of the form new("RegTypeFamily", ...). More frequently they are created via the generating function RegTypeFamily.

Slots

ErrorDistr:

object of class "Distribution": error distribution.

ErrorSymm:

object of class "DistributionSymmetry": symmetry of the error distribution.

RegDistr:

object of class "Distribution": regressor distribution.

RegSymm:

object of class "DistributionSymmetry": symmetry of the regressor distribution.

Regressor:

object of class "EuclRandVariable": regressor.

param:

object of class "ParamFamParameter": parameter of the family.

props:

object of class "character": properties of the family.

name:

object of class "character": name of the family.

distribution:

object of class "CondDistribution": distribution given the regressor.

Extends

Class "ParamFamily", directly.
Class "ProbFamily", by class "ParamFamily".

Methods

ErrorDistr

signature(object = "RegTypeFamily"): accessor function for slot ErrorDistr.

ErrorSymm

signature(object = "RegTypeFamily"): accessor function for slot ErrorSymm.

RegDistr

signature(object = "RegTypeFamily"): accessor function for slot RegDistr.

Regressor

signature(object = "RegTypeFamily"): accessor function for slot Regressor.

RegSymm

signature(object = "RegTypeFamily"): accessor function for slot RegSymm.

show

signature(object = "RegTypeFamily")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

ParamFamily-class

Examples


new("RegTypeFamily")

[Package ROptRegTS version 0.6.1 Index]