RobAStBaseOptions(RobAStBase)	R Documentation

Function to change the global variables of the package ‘RobAStBase’

Description

With RobAStBaseOptions you can inspect and change the global variables of the package RobAStBase.

Usage


RobAStBaseOptions(...)
getRobAStBaseOption(x)

Arguments

... 	 any options can be defined, using name = value or by passing a list of such tagged values.

x 	 a character string holding an option name.

Value

RobAStBaseOptions() returns a list of the global variables.
RobAStBaseOptions(x) returns the global variable x.
getRobAStBaseOption(x) returns the global variable x.
RobAStBaseOptions(x=y) sets the value of the global variable x to y.

Global Options

kStepUseLast:

The default value of argument kStepUseLast is FALSE. Explicitly setting kStepUseLast to TRUE should be done with care as in this situation the influence curve in case of oneStepEstimator and kStepEstimator is re-computed using the value of the one- resp. k-step estimate which may take quite a long time depending on the model.

withUpdateInKer:

if there is a non-trivial trafo in the model with matrix D, shall the parameter be updated on ker(D)? defaults to FALSE.

IC.UpdateInKer:

if there is a non-trivial trafo in the model with matrix D, the IC to be used for this; if NULL the result of getboundedIC(L2Fam,D) is taken; this IC will then be projected onto ker(D); defaults to NULL.

all.verbose:

argument verbose passed on by default to many calls of optIC, radiusminimaxIC, getinfRobIC etc.; well suited for testing purposes.

withPICList:

logical: shall slot pICList of return value of kStepEstimator be filled?

withICList:

logical: shall slot ICList of return value of kStepEstimator be filled?

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

options, getOption

Examples


RobAStBaseOptions()
RobAStBaseOptions("kStepUseLast")
RobAStBaseOptions("kStepUseLast" = TRUE)
# or
RobAStBaseOptions(kStepUseLast = 1e-6)
getRobAStBaseOption("kStepUseLast")

[Package RobAStBase version 0.7 Index]