RobModel-class(RobAStBase)	R Documentation

Robust model

Description

Class of robust models. A robust model consists of family of probability measures center and a neighborhood neighbor about this family.

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

center:

Object of class "ProbFamily"

neighbor:

Object of class "Neighborhood"

Methods

center

signature(object = "RobModel"): accessor function for slot center.

center<-

signature(object = "RobModel"): replacement function for slot center.

neighbor

signature(object = "RobModel"): accessor function for slot neighbor.

neighbor<-

signature(object = "RobModel"): replacement function for slot neighbor.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ProbFamily-class, Neighborhood-class

[Package RobAStBase version 0.2 Index]