SphericalSymmetry-class(ROptEst)	R Documentation

Class for Spherical Symmetric Distributions

Description

Class for spherical symmetric distributions.

Objects from the Class

Objects can be created by calls of the form new("SphericalSymmetry"). More frequently they are created via the generating function SphericalSymmetry. Spherical symmetry for instance leads to a simplification for the computation of optimally robust influence curves.

Slots

type:

Object of class "character": contains “spherical symmetric distribution”

SymmCenter:

Object of class "numeric": center of symmetry

Extends

Class "EllipticalSymmetry", directly.
Class "DistributionSymmetry", by class "EllipticalSymmetry".
Class "Symmetry", by class "EllipticalSymmetry".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

SphericalSymmetry, DistributionSymmetry-class

Examples


new("SphericalSymmetry")

[Package ROptEst version 0.5.0 Index]