TotalVarIC(RobAStBase)	R Documentation

Generating function for TotalVarIC-class

Description

Generates an object of class "TotalVarIC"; i.e., an influence curves eta of the form

eta = max(c, min(A Lambda, d))

with lower clipping bound c, upper clipping bound d and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable parametric family which can be created via CallL2Fam.

Usage


TotalVarIC(name, CallL2Fam = call("L2ParamFamily"), 
           Curve = EuclRandVarList(RealRandVariable(Map = c(function(x) {x}), 
                                                    Domain = Reals())), 
           Risks, Infos, clipLo = -Inf, clipUp = Inf, stand = as.matrix(1), 
           lowerCase = NULL, neighborRadius = 0, w = new("BdStWeight"),
           normtype = NormType(), biastype = symmetricBias(),
           modifyIC = NULL)

Arguments

name 	 object of class "character".

CallL2Fam 	 object of class "call": creates an object of the underlying L2-differentiable parametric family.

Curve 	 object of class "EuclRandVarList".

Risks 	 object of class "list": list of risks; cf. RiskType-class.

Infos 	 matrix of characters with two columns named method and message: additional informations.

clipLo 	 negative real: lower clipping bound.

clipUp 	 positive real: lower clipping bound.

stand 	 matrix: standardizing matrix

w 	 BdStWeight: weight object

lowerCase 	 optional constant for lower case solution.

neighborRadius 	 radius of the corresponding (unconditional) contamination neighborhood.

biastype 	 BiasType: type of the bias

normtype 	 NormType: type of the norm

modifyIC 	 object of class "OptionalFunction": function of two arguments, which are an L2 parametric family and an optional influence curve. Returns an object of class "IC". This function is mainly used for internal computations!

Value

Object of class "TotalVarIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class, ContIC

Examples


IC1 <- TotalVarIC()
plot(IC1)

[Package RobAStBase version 0.1.0 Index]