TotalVarIC-class(RobAStBase)	R Documentation

Influence curve of total variation type

Description

Class of (partial) influence curves of total variation type. i.e., an influence curves eta of the form

eta = max(c, min(A Lambda, d))

with lower clipping bound c, upper clipping bound d and standardizing matrix A. Lambda stands for the L2 derivative of the corresponding L2 differentiable parametric family which can be created via CallL2Fam.

Objects from the Class

Objects can be created by calls of the form new("TotalVarIC", ...). More frequently they are created via the generating function TotalVarIC, respectively via the method generateIC.

Slots

CallL2Fam

object of class "call": creates an object of the underlying L2-differentiable parametric family.

name

object of class "character".

Curve

object of class "EuclRandVarList".

modifyIC

Object of class "OptionalFunction": function of two arguments, which are an L2 parametric family and an optional influence curve. Returns an object of class "IC". This slot is mainly used for internal computations!

Risks

object of class "list": list of risks; cf. RiskType-class.

Infos

object of class "matrix" with two columns named method and message: additional informations.

clipLo

object of class "numeric": lower clipping bound.

clipUp

object of class "numeric": upper clipping bound.

stand

object of class "matrix": standardizing matrix.

weight

object of class "BdStWeight": weight function

biastype

object of class "BiasType": bias type (symmetric/onsided/asymmetric)

normtype

object of class "NormType": norm type (Euclidean, information/self-standardized)

neighborRadius

object of class "numeric": radius of the corresponding (unconditional) contamination neighborhood.

Extends

Class "HampIC", directly.
Class "IC", by class "HampIC".
Class "InfluenceCurve", by class "IC".

Methods

CallL2Fam<-

signature(object = "TotalVarIC"): replacement function for slot CallL2Fam.

clipLo

signature(object = "TotalVarIC"): accessor function for slot clipLo.

clipLo<-

signature(object = "TotalVarIC"): replacement function for slot clipLo.

clipUp

signature(object = "TotalVarIC"): accessor function for slot clipUp.

clipUp<-

signature(object = "TotalVarIC"): replacement function for slot clipUp.

clip

signature(object = "TotalVarIC"): returns clipUp-clipLo.

stand<-

signature(object = "TotalVarIC"): replacement function for slot stand.

lowerCase<-

signature(object = "TotalVarIC"): replacement function for slot lowerCase.

neighbor

signature(object = "TotalVarIC"): generates an object of class "TotalVarNeighborhood" with radius given in slot neighborRadius.

generateIC

signature(neighbor = "TotalVarNeighborhood", L2Fam = "L2ParamFamily"): generate an object of class "TotalVarIC". Rarely called directly.

show

signature(object = "TotalVarIC")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

, ,

Examples


IC1 <- new("TotalVarIC")
plot(IC1)

[Package RobAStBase version 0.7 Index]