asBias-class(ROptEst)	R Documentation

Standardized Asymptotic Bias

Description

Class of standardized asymptotic bias; i.e., the neighborhood radius is omitted respectively, set to 1.

Objects from the Class

Objects can be created by calls of the form new("asBias", ...). More frequently they are created via the generating function asBias.

Slots

type:

Object of class "character": “asymptotic bias”.

Extends

Class "asRisk", directly.
Class "RiskType", by class "asRisk".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class, asBias

Examples


new("asBias")

[Package ROptEst version 0.5.0 Index]