asCov-class(ROptEstOld)	R Documentation

Asymptotic covariance

Description

Class of asymptotic covariance.

Objects from the Class

Objects can be created by calls of the form new("asCov", ...). More frequently they are created via the generating function asCov.

Slots

type:

Object of class "character": “asymptotic covariance”.

Extends

Class "asRisk", directly.
Class "RiskType", by class "asRisk".

Methods

No methods defined with class "asCov" in the signature.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class, asCov

Examples


new("asCov")

[Package ROptEstOld version 0.5.2 Index]