asHampel-class(ROptEstOld)	R Documentation

Asymptotic Hampel risk

Description

Class of asymptotic Hampel risk which is the trace of the asymptotic covariance subject to a given bias bound (bound on gross error sensitivity).

Objects from the Class

Objects can be created by calls of the form new("asHampel", ...). More frequently they are created via the generating function asHampel.

Slots

type:

Object of class "character": “trace of asymptotic covariance for given bias bound”.

bound:

Object of class "numeric": given positive bias bound.

Extends

Class "asRisk", directly.
Class "RiskType", by class "asRisk".

Methods

bound

signature(object = "asHampel"): accessor function for slot bound.

show

signature(object = "asHampel")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class, asHampel

Examples


new("asHampel")

[Package ROptEstOld version 0.5.2 Index]