asMSE-class(ROptEstOld)	R Documentation

Asymptotic mean square error

Description

Class of asymptotic mean square error.

Objects from the Class

Objects can be created by calls of the form new("asMSE", ...). More frequently they are created via the generating function asMSE.

Slots

type:

Object of class "character": “asymptotic mean square error”.

Extends

Class "asGRisk", directly.
Class "asRisk", by class "asGRisk".
Class "RiskType", by class "asGRisk".

Methods

No methods defined with class "asMSE" in the signature.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asGRisk-class, asMSE

Examples


new("asMSE")

[Package ROptEstOld version 0.5.2 Index]