asUnOvShoot-class(ROptEst)	R Documentation

Asymptotic under-/overshoot probability

Description

Class of asymptotic under-/overshoot probability.

Objects from the Class

Objects can be created by calls of the form new("asUnOvShoot", ...). More frequently they are created via the generating function asUnOvShoot.

Slots

type:

Object of class "character": dQuote{asymptotic under-/overshoot probability}.

width:

Object of class "numeric": half the width of given confidence interval.

Extends

Class "asGRisk", directly.
Class "asRisk", by class "asGRisk".
Class "RiskType", by class "asGRisk".

Methods

width

signature(object = "asUnOvShoot"): accessor function for slot width.

show

signature(object = "asUnOvShoot")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asGRisk-class

Examples


new("asUnOvShoot")

[Package ROptEst version 0.5.0 Index]