asymmetricBiasType-class(ROptEst)	R Documentation

asymmetric Bias Type

Description

Class of asymmetric bias types.

Objects from the Class

Objects can be created by calls of the form new("asymmetricBiasType", ...). More frequently they are created via the generating function asymmetricBias.

Slots

name:

Object of class "character".

nu:

Object of class "numeric"; to be in (0,1] x (0,1] with maximum 1; weights for negative and positive bias, respectively

Methods

nu

signature(object = "asymmetricBiasType"): accessor function for slot nu.

nu<-

signature(object = "asymmetricBiasType", value = "numeric"): replacement function for slot nu.

Extends

Class "BiasType", directly.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BiasType-class

Examples


asymmetricBias()
## The function is currently defined as
function(){ new("asymmetricBiasType", name = "asymmetric Bias", nu = c(1,1)) }

aB <- asymmetricBias()
nu(aB)
try(nu(aB) <- -2) ## error
nu(aB) <- c(0.3,1)

[Package ROptEst version 0.5.0 Index]