checkIC(RobAStBase)	R Documentation

Generic Function for Checking ICs

Description

Generic function for checking centering and Fisher consistency of ICs.

Usage


checkIC(IC, L2Fam, ...)

Arguments

IC 	 object of class "IC"

L2Fam 	 L2-differentiable family of probability measures.

... 	 additional parameters

Details

The precisions of the centering and the Fisher consistency are computed.

Value

The maximum deviation from the IC properties is returned.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily-class, IC-class

Examples


IC1 <- new("IC")
checkIC(IC1)

[Package RobAStBase version 0.1.0 Index]