cutoff(RobAStBase)	R Documentation

Generating function(s) for class 'cutoff'

Description

Generating function(s) for class cutoff.

Usage


cutoff(name = "empirical", body.fct0,
       cutoff.quantile  = 0.95,
       norm = NormType(), QF, nsim = 100000)
cutoff.sememp()
cutoff.chisq()

Arguments

name 	 argument for name slot of cutoff object

body.fct0 	 a call generated by code wrapped to substitute resp. quote; the body of the fct slot of the cutoff object

cutoff.quantile 	 numeric; the corresponding slot value for the cutoff object

norm 	 an object of class NormType – the norm/distance by which to produce the cutoff - value.

nsim 	 integer: the sample size used for determining the quantiles of x'Qx^.5 for x multivariate standard normal and Q a corresponding quadratic form

QF 	 a quadratic (positive semidefinite, symmetric) matrix used as quadratic form

Details

cutoff generates a valid object of class "cutoff". As function slot fct may only have a formal argument data, the other arguments to determine the cutoff value, i.e. norm, QF, nsim, cutoff.quantile, nsim have to enter the scope of this function by lexical scoping; now cutoff.quantile, norm, QF are to be taken from the calling environment (not from the defining one), so we have delay evaluation of the function body, which is why we assume it to be given wrapped into substitute resp. quote. body.fct0 is by default (i.e. if argument body.fct0 is missing) set to quote(quantile(slot(norm,"fct")(data), cutoff.quantile)), internally, i.e.; to an empirical quantile of the corresponding norms.

cutoff.sememp() is a helper function generating the theoretical (asymptotic) quantile of (the square root of) a corresponding quadratic form, assuming multivariate normality; to determine this quantile nsim simulations are used.

cutoff.chisq() is a helper function generating the theoretical (asymptotic) quantile of (the square root of) a (self-standardized) quadratic form, assuming multivariate normality; i.e.; a corresponding quantile of a Chi-Square distribution.

Value

Object of class "cutoff".

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

cutoff-class, ddPlot

Examples


cutoff()
cutoff.sememp()
cutoff.chisq()

[Package RobAStBase version 0.2 Index]