cutoff-class(RobAStBase)	R Documentation

Cutoff class for distance-distance plots

Description

Class of methods to determine cutoff point for distance-distance plots; used to derive other cutoff methods later by method dispatch.

Objects from the Class

Objects could in principle be created by calls of the form new("cutoff", ...). More frequently they are created via the generating function cutoff, respectively via the helper functions cutoff.sememp and cutoff.chisq.

Slots

name:

object of class "character"; defaults to "empirical" in prototype;

fct:

an object of of class "function"; for this class layer, this function must only have one argument data (which may but need not be used to determine the cutoff point empirically); in derived classes this restriction could be dropped, if corresponding special methods for ddPlot are derived. Defaults to function(data) quantile(data).

cutoff.quantile:

Object of class "numeric": a quantile (empirical or theoretical) to plot the cutoff line; defaults to 0.95 in prototype;

Methods

cutoff.quantile

signature(object = "cutoff"): accessor function for slot cutoff.quantile.

cutoff.quantile<-

signature(object = "cutoff"): replacement function for slot cutoff.quantile.

fct

signature(object = "cutoff"): accessor function for slot fct.

name

signature(object = "cutoff"): accessor function for slot name.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

ddPlot, outlyingPlotIC cutoff

Examples


cutoff()

[Package RobAStBase version 0.1.2 Index]