ddPlot-methods(RobAStBase)	R Documentation

Methods for Function ddPlot in Package ‘RobAStBase’

Description

ddPlot-methods

Usage


ddPlot(data, dist.x, dist.y, cutoff.x, cutoff.y, ...)
## S4 method for signature 'matrix':
ddPlot(data, dist.x = NormType(), dist.y  = NormType(),
       cutoff.x, cutoff.y, ...,
       cutoff.quantile.x = 0.95, cutoff.quantile.y = cutoff.quantile.x,
       transform.x, transform.y = transform.x,
       id.n, lab.pts, adj, cex.idn,
       col.idn, lty.cutoff, lwd.cutoff, col.cutoff)
## S4 method for signature 'numeric':
ddPlot(data, dist.x = NormType(), dist.y  = NormType(),
       cutoff.x, cutoff.y, ...,
       cutoff.quantile.x = 0.95, cutoff.quantile.y = cutoff.quantile.x,
       transform.x, transform.y = transform.x,
       id.n, lab.pts, adj, cex.idn,
       col.idn, lty.cutoff, lwd.cutoff, col.cutoff)
## S4 method for signature 'data.frame':
ddPlot(data, dist.x = NormType(), dist.y  = NormType(),
       cutoff.x, cutoff.y, ...,
       cutoff.quantile.x = 0.95, cutoff.quantile.y = cutoff.quantile.x,
       transform.x, transform.y = transform.x,
       id.n, lab.pts, adj, cex.idn,
       col.idn, lty.cutoff, lwd.cutoff, col.cutoff)

Arguments

data 	 data coercable to matrix; the data at which to produce the ddPlot.

... 	 further arguments to be passed to plot.default, text, and abline

dist.x 	 object of class NormType; the distance for the x axis.

dist.y 	 object of class NormType; the distance for the y axis.

cutoff.x 	 object of class cutoff; the cutoff information for the x axis (the vertical line discriminating 'good' and 'bad' points).

cutoff.y 	 object of class cutoff; the cutoff information for the y axis (the horizontal line discriminating 'good' and 'bad' points).

cutoff.quantile.x 	 numeric; the cutoff quantile for the x axis.

cutoff.quantile.y 	 numeric; the cutoff quantile for the y axis.

transform.x 	 function; a transformation to be performed before determining the distances of the x axis.

transform.y 	 function; a transformation to be performed before determining the distances of the y axis.

id.n 	 a set of indices (or a corresponding logical vector); to select a subset of the data in argument data.

lab.pts 	 a vector of labels for the (unsubsetted) data.

adj 	 the corresponding argument for text for labelling the outliers.

cex.idn 	 the corresponding cex argument for text for labelling the outliers.

col.idn 	 the corresponding col argument for text for labelling the outliers.

lty.cutoff 	 the corresponding lty argument for abline for drawing the cutoff lines.

lwd.cutoff 	 the corresponding lwd argument for abline for drawing the cutoff lines.

col.cutoff 	 the corresponding col argument for abline for drawing the cutoff lines.

Details

The matrix-method calls .ddPlot.MatNtNtCoCo, the numeric- and data.frame-methods coerce argument data to matrix — the numeric-method by a call to matrix(data, nrow=1), in the data.frame-methods by a call to t(as.matrix(data)).

Value

a list with items

id.x 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff

id.y 	 the indices of (possibly transformed) data (within subset id.n) beyond the y-cutoff

id.xy 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff and the y-cutoff

qtx 	 the quantiles of the distances of the (possibly transformed) data in x direction

qty 	 the quantiles of the distances of the (possibly transformed) data in y direction

cutoff.x.v 	 the cutoff value in x direction

cutoff.y.v 	 the cutoff value in y direction

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

Examples


MX <- matrix(rnorm(1500),nrow=6)
QM <- matrix(rnorm(36),nrow=6); QM <- QM %*% t(QM)
ddPlot(data=MX, dist.y=QFNorm(QuadF=PosSemDefSymmMatrix(QM)))

[Package RobAStBase version 0.2 Index]