evalIC(ROptEst)	R Documentation

Generic function for evaluating ICs

Description

Generic function for evaluating ICs.

Usage


evalIC(IC, x)

Arguments

IC 	 object of class "IC"

x 	 numeric vector or matrix

Details

The list of random variables contained in the slot Curve is evaluated at x.

Value

In case x is numeric a vector and in case x is matrix a matrix is returned.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

[Package ROptEst version 0.5.0 Index]