fiCov-class(ROptEst)	R Documentation

Finite-sample covariance

Description

Class of finite-sample covariance.

Objects from the Class

Objects can be created by calls of the form new("fiCov", ...). More frequently they are created via the generating function fiCov.

Slots

type:

Object of class "character": “finite-sample covariance”.

Extends

Class "fiRisk", directly.
Class "RiskType", by class "fiRisk".

Methods

No methods defined with class "fiCov" in the signature.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Ruckdeschel, P. and Kohl, M. (2005) How to approximate the finite sample risk of M-estimators.

See Also

fiRisk-class, fiCov

Examples


new("fiCov")

[Package ROptEst version 0.5.0 Index]