fiHampel-class(ROptEst)	R Documentation

Finite-sample Hampel risk

Description

Class of finite-sample Hampel risk which is the trace of the finite-sample covariance subject to a given bias bound (bound on gross error sensitivity).

Objects from the Class

Objects can be created by calls of the form new("fiHampel", ...). More frequently they are created via the generating function fiHampel.

Slots

type:

Object of class "character": “trace of finite-sample covariance for given bias bound”.

bound:

Object of class "numeric": given positive bias bound.

Extends

Class "fiRisk", directly.
Class "RiskType", by class "fiRisk".

Methods

bound

signature(object = "fiHampel"): accessor function for slot bound.

show

signature(object = "fiHampel")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Ruckdeschel, P. and Kohl, M. (2005) How to approximate the finite sample risk of M-estimators.

See Also

fiRisk-class, fiHampel

Examples


new("fiHampel")

[Package ROptEst version 0.5.0 Index]