fiUnOvShoot-class(ROptEstOld)	R Documentation

Finite-sample under-/overshoot probability

Description

Class of finite-sample under-/overshoot probability.

Objects from the Class

Objects can be created by calls of the form new("fiUnOvShoot", ...). More frequently they are created via the generating function fiUnOvShoot.

Slots

type:

Object of class "character": dQuote{finite-sample under-/overshoot probability}.

width:

Object of class "numeric": half the width of given confidence interval.

Extends

Class "fiRisk", directly.
Class "RiskType", by class "fiRisk".

Methods

width

signature(object = "fiUnOvShoot"): accessor function for slot width.

show

signature(object = "fiUnOvShoot")

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1989) A finite-sample minimax regression estimator. Statistics 20(2): 211–221.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

Ruckdeschel, P. and Kohl, M. (2005) Computation of the Finite Sample Risk of M-estimators on Neighborhoods.

See Also

fiRisk-class

Examples


new("fiUnOvShoot")

[Package ROptEstOld version 0.5.2 Index]