getAsRisk(ROptEst)	R Documentation

Generic Function for Computation of Asymptotic Risks

Description

Generic function for the computation of asymptotic risks. This function is rarely called directly. It is used by other functions.

Usage


getAsRisk(risk, L2deriv, neighbor, ...)

## S4 method for signature 'asMSE, UnivariateDistribution,
##   Neighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand, trafo)

## S4 method for signature 'asMSE, EuclRandVariable,
##   Neighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   ContNeighborhood':
getAsRisk(risk, L2deriv, neighbor, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   TotalVarNeighborhood':
getAsRisk(risk, L2deriv, neighbor, trafo)

## S4 method for signature 'asBias, RealRandVariable,
##   ContNeighborhood':
getAsRisk(risk, L2deriv, neighbor, Distr, L2derivDistrSymm, trafo, 
             z.start, A.start, maxiter, tol)

## S4 method for signature 'asCov, UnivariateDistribution,
##   ContNeighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand)

## S4 method for signature 'asCov, UnivariateDistribution,
##   TotalVarNeighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand)

## S4 method for signature 'asCov, RealRandVariable,
##   ContNeighborhood':
getAsRisk(risk, L2deriv, neighbor, Distr, clip, cent, stand)

## S4 method for signature 'trAsCov,
##   UnivariateDistribution, UncondNeighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand)

## S4 method for signature 'trAsCov, RealRandVariable,
##   ContNeighborhood':
getAsRisk(risk, L2deriv, neighbor, Distr, clip, cent, stand)

## S4 method for signature 'asUnOvShoot,
##   UnivariateDistribution, UncondNeighborhood':
getAsRisk(risk, L2deriv, neighbor, clip, cent, stand, trafo)

Arguments

risk 	 object of class "asRisk".

L2deriv 	 L2-derivative of some L2-differentiable family of probability distributions.

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

stand 	 standardizing matrix.

trafo 	 matrix: transformation of the parameter.

Distr 	 object of class "Distribution".

L2derivDistrSymm 	 object of class "DistrSymmList".

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

Value

The asymptotic risk is computed.

Methods

risk = "asMSE", L2deriv = "UnivariateDistribution", neighbor = "Neighborhood":

computes asymptotic mean square error in methods for function getInfRobIC.

risk = "asMSE", L2deriv = "EuclRandVariable", neighbor = "Neighborhood":

computes asymptotic mean square error in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asCov", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "asCov", L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "asCov", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "trAsCov", L2deriv = "UnivariateDistribution", neighbor = "UncondNeighborhood":

computes trace of asymptotic covariance in methods for function getInfRobIC.

risk = "trAsCov", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood":

computes trace of asymptotic covariance in methods for function getInfRobIC.

risk = "asUnOvShoot", L2deriv = "UnivariateDistribution", neighbor = "UncondNeighborhood":

computes asymptotic under-/overshoot risk in methods for function getInfRobIC.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions (submitted).

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class

[Package ROptEst version 0.5.0 Index]