getAsRiskRegTS(ROptRegTS)	R Documentation

Generic Function for Computation of Asymptotic Risks in case of Regression-Type Models

Description

Generic function for the computation of asymptotic risks in case of regression-type models. This function is rarely called directly. It is used by other functions.

Usage


getAsRiskRegTS(risk, ErrorL2deriv, Regressor, neighbor, ...)

## S4 method for signature 'asMSE,UnivariateDistribution,Distribution,Neighborhood':
getAsRiskRegTS(risk, ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, trafo)

## S4 method for signature 'asMSE,UnivariateDistribution,Distribution,Av2CondContNeighborhood':
getAsRiskRegTS(risk, ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, trafo)

## S4 method for signature 'asMSE,EuclRandVariable,Distribution,Neighborhood':
getAsRiskRegTS(risk, ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, trafo)

## S4 method for signature 'asBias,UnivariateDistribution,UnivariateDistribution,ContNeighborhood':
getAsRiskRegTS(risk, ErrorL2deriv, 
                Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,UnivariateDistribution,Av1CondContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,UnivariateDistribution,Av1CondTotalVarNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,MultivariateDistribution,ContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,MultivariateDistribution,Av1CondContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,MultivariateDistribution,Av1CondTotalVarNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,UnivariateDistribution,Distribution,Av2CondContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorL2derivDistrSymm, trafo, maxiter, tol)

## S4 method for signature 'asBias,RealRandVariable,Distribution,ContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorDistr, trafo, z.start, A.start, maxiter, tol)

## S4 method for signature 'asBias,RealRandVariable,Distribution,Av1CondContNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, ErrorDistr, trafo, z.start, A.start, maxiter, tol)

## S4 method for signature 'asUnOvShoot,UnivariateDistribution,UnivariateDistribution,UncondNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, clip, cent, stand)

## S4 method for signature 'asUnOvShoot,UnivariateDistribution,UnivariateDistribution,CondNeighborhood':
getAsRiskRegTS(risk, 
                ErrorL2deriv, Regressor, neighbor, clip, cent, stand)

Arguments

risk 	 object of class "asRisk".

ErrorL2deriv 	 L2-derivative of ErrorDistr.

Regressor 	 regressor.

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant/function.

stand 	 standardizing matrix.

trafo 	 matrix: transformation of the parameter.

ErrorDistr 	 error distribution.

ErrorL2derivDistrSymm 	 symmetry of ErrorL2derivDistr.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

z.start 	 initial value for the centering constant/function.

A.start 	 initial value for the standardizing matrix.

Value

The asymptotic risk is computed.

Methods

risk = "asMSE", ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", neighbor = "Neighborhood"

computes asymptotic mean square error in methods for function getInfRobRegTypeIC.

risk = "asMSE", ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", neighbor = "Av2CondContNeighborhood"

computes asymptotic mean square error in methods for function getInfRobRegTypeIC.

risk = "asMSE", ErrorL2deriv = "EuclRandVariable", Regressor = "Distribution", neighbor = "Neighborhood"

computes asymptotic mean square error in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "ContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "ContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", neighbor = "Av2CondContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "ContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asBias", ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "Av1CondContNeighborhood"

computes standardized asymptotic bias in methods for function getInfRobRegTypeIC.

risk = "asUnOvShoot", ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "UncondNeighborhood"

computes asymptotic under-/overshoot risk in methods for function getInfRobRegTypeIC.

risk = "asUnOvShoot", ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "CondNeighborhood"

computes asymptotic under-/overshoot risk in methods for function getInfRobRegTypeIC.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions (submitted).

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class

[Package ROptRegTS version 0.6.1 Index]