getBiasIC(ROptEst)	R Documentation

Generic function for the computation of the asymptotic bias for an IC

Description

Generic function for the computation of the asymptotic bias for an IC.

Usage


getBiasIC(IC, neighbor, ...)

## S4 method for signature 'IC, UncondNeighborhood':
getBiasIC(IC, neighbor, L2Fam, 
                     biastype = symmetricBias(), 
                     tol = .Machine$double.eps^0.25)


Arguments

IC 	 object of class "InfluenceCurve"

neighbor 	 object of class "Neighborhood".

... 	 additional parameters

L2Fam 	 object of class "L2ParamFamily" or missing.

biastype 	 object of class "BiasType".

tol 	 the desired accuracy (convergence tolerance).

Details

To make sure that the results are valid, it is recommended to include an additional check of the IC properties of IC using checkIC.

Value

The asymptotic bias of an IC is computed.

Methods

IC = "IC", neighbor = "UncondNeighborhood":

asymptotic bias of IC in case of unconditional neighborhoods.

Note

This generic function is still under construction.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

See Also

getRiskIC-methods, InfRobModel-class

[Package ROptEst version 0.6.0 Index]