getFixClip(ROptEst)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound in case of robust models with fixed neighborhoods. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getFixClip(clip, Distr, risk, neighbor,  ...)

## S4 method for signature 'numeric, Norm, fiUnOvShoot,
##   ContNeighborhood':
getFixClip(clip, Distr, risk, neighbor)

## S4 method for signature 'numeric, Norm, fiUnOvShoot,
##   TotalVarNeighborhood':
getFixClip(clip, Distr, risk, neighbor)

Arguments

clip 	 positive real: clipping bound

Distr 	 object of class "Distribution".

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

Value

The optimal clipping bound is computed.

Methods

clip = "numeric", Distr = "Norm", risk = "fiUnOvShoot", neighbor = "ContNeighborhood"

optimal clipping bound for finite-sample under-/overshoot risk.

clip = "numeric", Distr = "Norm", risk = "fiUnOvShoot", neighbor = "TotalVarNeighborhood"

optimal clipping bound for finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.6.0 Index]