getFixClipRegTS(ROptRegTS)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound/function. This function is rarely called directly. It is used to compute optimally robust ICs in case of fixed robust models.

Usage


getFixClipRegTS(clip, ErrorDistr, Regressor, risk, neighbor, ...)

Arguments

clip 	 optimal clipping bound.

ErrorDistr 	 error distribution.

Regressor 	 regressor.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

Value

The optimal clipping bound/function is computed.

Methods

clip = "numeric", ErrorDistr = "Norm", Regressor = "UnivariateDistribution", risk = "fiUnOvShoot", neighbor = "ContNeighborhood"

optimal clipping bound for finite-sample under-/overshoot risk.

clip = "numeric", ErrorDistr = "Norm", Regressor = "UnivariateDistribution", risk = "fiUnOvShoot", neighbor = "TotalVarNeighborhood"

optimal clipping bound for finite-sample under-/overshoot risk.

clip = "numeric", ErrorDistr = "Norm", Regressor = "numeric", risk = "fiUnOvShoot", neighbor = "CondContNeighborhood"

optimal clipping function for finite-sample under-/overshoot risk.

clip = "numeric", ErrorDistr = "Norm", Regressor = "numeric", risk = "fiUnOvShoot", neighbor = "CondTotalVarNeighborhood"

optimal clipping function for finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1989) A finite-sample minimax regression estimator. Statistics 20(2): 211–221.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class, , , , ,

[Package ROptRegTS version 0.6.1 Index]