getFixRobIC(ROptEst)	R Documentation

Generic Function for the Computation of Optimally Robust ICs

Description

Generic function for the computation of optimally robust ICs in case of robust models with fixed neighborhoods. This function is rarely called directly.

Usage


getFixRobIC(Distr, risk, neighbor, ...)

## S4 method for signature 'Norm, fiUnOvShoot,
##   UncondNeighborhood':
getFixRobIC(Distr, risk, neighbor, 
          sampleSize, upper, maxiter, tol, warn, Algo, cont)

Arguments

Distr 	 object of class "Distribution".

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

sampleSize 	 integer: sample size.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Algo 	 "A" or "B".

cont 	 "left" or "right".

Value

The optimally robust IC is computed.

Methods

Distr = "Norm", risk = "fiUnOvShoot", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for one-dimensional normal location and finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

FixRobModel-class

[Package ROptEst version 0.6.0 Index]