getFixRobRegTypeIC(ROptRegTS)	R Documentation

Generic Function for the Computation of Optimally Robust Regression-Type ICs

Description

Generic function for the computation of optimally robust regression-type ICs in case of fixed robust models. This function is rarely called directly.

Usage


getFixRobRegTypeIC(ErrorDistr, Regressor, risk, neighbor, ...)

## S4 method for signature 'Norm, UnivariateDistribution,
##   fiUnOvShoot, UncondNeighborhood':
getFixRobRegTypeIC(ErrorDistr, 
            Regressor, risk, neighbor, sampleSize, upper, maxiter, tol, warn, Algo, cont)

## S4 method for signature 'Norm, UnivariateDistribution,
##   fiUnOvShoot, CondNeighborhood':
getFixRobRegTypeIC(ErrorDistr, 
            Regressor, risk, neighbor, sampleSize, upper, maxiter, tol, warn, Algo, cont)

Arguments

ErrorDistr 	 error distribution

Regressor 	 regressor

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

sampleSize 	 integer: sample size.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Algo 	 "A" or "B".

cont 	 "left" or "right".

Value

The optimally robust IC is computed.

Methods

ErrorDistr = "Norm", Regressor = "UnivariateDistribution", risk = "fiUnOvShoot", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for one-dimensional normal regression and finite-sample under-/overshoot risk.

ErrorDistr = "Norm", Regressor = "UnivariateDistribution", risk = "fiUnOvShoot", neighbor = "CondNeighborhood"

computes the optimally robust influence curve for one-dimensional normal regression and finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1989) A finite-sample minimax regression estimator. Statistics 20(2): 211–221.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

FixRobRegTypeModel-class

[Package ROptRegTS version 0.6.1 Index]