getIneffDiff(ROptEst)	R Documentation

Generic Function for the Computation of Inefficiency Differences

Description

Generic function for the computation of inefficiency differencies. This function is rarely called directly. It is used to compute the radius minimax IC and the least favorable radius.

Usage


getIneffDiff(radius, L2Fam, neighbor, risk, biastype, ...)

## S4 method for signature 'numeric, L2ParamFamily,
##   UncondNeighborhood, asMSE, BiasType':
getIneffDiff(
          radius, L2Fam, neighbor, risk, biastype = symmetricBias(), loRad, upRad, 
            loRisk, upRisk, z.start = NULL, A.start = NULL, upper.b, MaxIter, eps, warn)

Arguments

radius 	 neighborhood radius.

L2Fam 	 L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

risk 	 object of class "RiskType".

biastype 	 object of class "BiasType".

... 	 additional parameters

loRad 	 the lower end point of the interval to be searched.

upRad 	 the upper end point of the interval to be searched.

loRisk 	 the risk at the lower end point of the interval.

upRisk 	 the risk at the upper end point of the interval.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper.b 	 upper bound for the optimal clipping bound.

MaxIter 	 the maximum number of iterations

eps 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Value

The inefficieny difference between the left and the right margin of a given radius interval is computed.

Methods

radius = "numeric", L2Fam = "L2ParamFamily", neighbor = "UncondNeighborhood", risk = "asMSE", biastype = "BiasType":

computes difference of asymptotic MSE–inefficiency for the boundaries of a given radius interval.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H., Kohl, M. and Ruckdeschel, P. (2001) The Costs of not Knowing the Radius. Submitted. Appeared as discussion paper Nr. 81. SFB 373 (Quantification and Simulation of Economic Processes), Humboldt University, Berlin; also available under www.uni-bayreuth.de/departments/math/org/mathe7/RIEDER/pubs/RR.pdf

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

radiusMinimaxIC, leastFavorableRadius

[Package ROptEst version 0.6.0 Index]