getIneffDiff-methods(ROptRegTS)	R Documentation

Methods for Function getIneffDiff in Package ‘ROptRegTS’

Description

Methods for function getIneffDiff in package ROptRegTS. These methods are rarely called directly. They are used to compute the radius minimax IC and the least favorable radius.

Methods

radius = "numeric", L2Fam = "L2RegTypeFamily", neighbor = "Neighborhood", risk = "asMSE"

computes difference of asymptotic MSE–inefficiency for the boundaries of a given radius interval.

radius = "numeric", L2Fam = "L2RegTypeFamily", neighbor = "Av2CondContNeighborhood", risk = "asMSE"

computes difference of asymptotic MSE–inefficiency for the boundaries of a given radius interval.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

getIneffDiff

[Package ROptRegTS version 0.6.1 Index]