getInfCent(ROptEstOld)	R Documentation

Generic Function for the Computation of the Optimal Centering Constant/Lower Clipping Bound

Description

Generic function for the computation of the optimal centering constant (contamination neighborhoods) respectively, of the optimal lower clipping bound (total variation neighborhood). This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfCent(L2deriv, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood':
getInfCent(L2deriv, neighbor, clip, cent, tol.z, symm, trafo)

## S4 method for signature 'UnivariateDistribution,
##   TotalVarNeighborhood':
getInfCent(L2deriv, neighbor, clip, cent, tol.z, symm, trafo)

## S4 method for signature 'RealRandVariable,
##   ContNeighborhood':
getInfCent(L2deriv, neighbor, Distr, z.comp, stand, cent, clip)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

Distr 	 distribution of L2-differentiable family.

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

stand 	 standardizing matrix.

tol.z 	 the desired accuracy (convergence tolerance).

symm 	 logical: indicating symmetry of L2deriv.

trafo 	 matrix: transformation of the parameter.

z.comp 	 logical vector: indication which components of the centering constant have to be computed.

Value

The optimal centering constant is computed.

Methods

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood"

computation of optimal centering constant.

L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood"

computation of optimal lower clipping bound.

L2deriv = "RealRandVariable", neighbor = "ContNeighborhood"

computation of optimal centering constant.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEstOld version 0.5.2 Index]