getInfCentRegTS(ROptRegTS)	R Documentation

Generic Function for the Computation of the Optimal Centering Constant/Function resp. Lower Clipping Bound/Function

Description

Generic function for the computation of the optimal centering constant/function (contamination neighborhoods) respectively, of the optimal lower clipping bound/function (total variation neighborhoods). This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfCentRegTS(ErrorL2deriv, Regressor, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,ContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,TotalVarNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, z.comp)

## S4 method for signature 'UnivariateDistribution,numeric,CondTotalVarNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, z.comp)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,Av1CondContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp, x.vec)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,Av1CondTotalVarNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp, x.vec, tol.z)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,ContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,Av1CondContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp, x.vec)

## S4 method for signature 'UnivariateDistribution,Distribution,Av2CondContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, clip, cent, stand, z.comp, tol.z)

## S4 method for signature 'RealRandVariable,Distribution,ContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, ErrorDistr, stand, cent, clip, z.comp)

## S4 method for signature 'RealRandVariable,Distribution,Av1CondContNeighborhood':
getInfCentRegTS(ErrorL2deriv, 
                Regressor, neighbor, ErrorDistr, stand, cent, clip, z.comp, x.vec)

Arguments

ErrorL2deriv 	 L2-derivative of ErrorDistr.

Regressor 	 regressor.

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant/function.

stand 	 standardizing matrix.

z.comp 	 which components of the centering constant/function have to be computed.

x.vec 	 (approximated) support of Regressor.

tol.z 	 the desired accuracy (convergence tolerance).

ErrorDistr 	 error distribution.

Value

The optimal centering constant/function is computed.

Methods

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "ContNeighborhood"

computation of optimal centering constant.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "TotalVarNeighborhood"

computation of lower clipping bound.

ErrorL2deriv = "UnivariateDistribution", Regressor = "numeric", neighbor = "CondTotalVarNeighborhood"

computation of lower clipping bound.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondContNeighborhood"

computation of optimal centering function.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computation of optimal lower clipping function.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "ContNeighborhood"

computation of optimal centering constant.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondContNeighborhood"

computation of optimal centering function.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computation of optimal lower clipping function.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", neighbor = "Av2CondContNeighborhood"

computation of optimal centering constant.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "ContNeighborhood"

computation of optimal centering constant.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "Av1CondContNeighborhood"

computation of optimal centering function.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, , , , ,

[Package ROptRegTS version 0.6.1 Index]