getInfClip(ROptEst)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound in case of infinitesimal robust models. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfClip(clip, L2deriv, risk, neighbor, ...)

## S4 method for signature 'numeric,
##   UnivariateDistribution, asMSE, ContNeighborhood':
getInfClip(clip, L2deriv, risk, neighbor, biastype, cent, symm, trafo)

## S4 method for signature 'numeric,
##   UnivariateDistribution, asMSE, TotalVarNeighborhood':
getInfClip(clip, L2deriv, risk, neighbor, biastype, cent, symm, trafo)

## S4 method for signature 'numeric, EuclRandVariable,
##   asMSE, ContNeighborhood':
getInfClip(clip, L2deriv, risk, neighbor, Distr, stand, biastype, cent, trafo)

## S4 method for signature 'numeric,
##   UnivariateDistribution, asUnOvShoot,
##   UncondNeighborhood':
getInfClip(clip, L2deriv, risk, neighbor, biastype, cent, symm, trafo)

## S4 method for signature 'numeric,
##   UnivariateDistribution, asSemivar, ContNeighborhood':
getInfClip(clip, L2deriv, risk, neighbor, biastype, cent, symm, trafo)

Arguments

clip 	 positive real: clipping bound

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

biastype 	 object of class "BiasType"

cent 	 optimal centering constant.

stand 	 standardizing matrix.

Distr 	 object of class "Distribution".

symm 	 logical: indicating symmetry of L2deriv.

trafo 	 matrix: transformation of the parameter.

Value

The optimal clipping bound is computed.

Methods

clip = "numeric", L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood"

optimal clipping bound for asymtotic mean square error.

clip = "numeric", L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "TotalVarNeighborhood"

optimal clipping bound for asymtotic mean square error.

clip = "numeric", L2deriv = "EuclRandVariable", risk = "asMSE", neighbor = "ContNeighborhood"

optimal clipping bound for asymtotic mean square error.

clip = "numeric", L2deriv = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "UncondNeighborhood"

optimal clipping bound for asymtotic under-/overshoot risk.

clip = "numeric", L2deriv = "UnivariateDistribution", risk = "asSemivar", neighbor = "ContNeighborhood"

optimal clipping bound for asymtotic semivariance.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.6.0 Index]