getInfGamma(ROptEst)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound. This function is rarely called directly. It is called by getInfClip to compute optimally robust ICs.

Usage


getInfGamma(L2deriv, risk, neighbor, biastype, ...)

## S4 method for signature 'UnivariateDistribution, asMSE,
##   ContNeighborhood, BiasType':
getInfGamma(L2deriv, 
     risk, neighbor, biastype, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   asGRisk, TotalVarNeighborhood, BiasType':
getInfGamma(L2deriv, 
     risk, neighbor, biastype, cent, clip)

## S4 method for signature 'RealRandVariable, asMSE,
##   ContNeighborhood, BiasType':
getInfGamma(L2deriv, 
     risk, neighbor, biastype, Distr, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   asUnOvShoot, ContNeighborhood, BiasType':
getInfGamma(L2deriv, 
     risk, neighbor, biastype, cent, clip)

## S4 method for signature 'UnivariateDistribution, asMSE,
##   ContNeighborhood, onesidedBias':
getInfGamma(L2deriv, 
     risk, neighbor, biastype, cent, clip)

## S4 method for signature 'UnivariateDistribution, asMSE,
##   ContNeighborhood, asymmetricBias':
getInfGamma(L2deriv, 
    risk, neighbor, biastype, cent, clip)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType"

... 	 additional parameters

cent 	 optimal centering constant.

clip 	 optimal clipping bound.

stand 	 standardizing matrix.

Distr 	 object of class "Distribution".

Details

The function is used in case of asymptotic G-risks; confer Ruckdeschel and Rieder (2004).

Methods

L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood", biastype = "BiasType"

used by getInfClip for symmetric bias.

L2deriv = "UnivariateDistribution", risk = "asGRisk", neighbor = "TotalVarNeighborhood", biastype = "BiasType"

used by getInfClip for symmetric bias.

L2deriv = "RealRandVariable", risk = "asMSE", neighbor = "ContNeighborhood", biastype = "BiasType"

used by getInfClip for symmetric bias.

L2deriv = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "ContNeighborhood", biastype = "BiasType"

used by getInfClip for symmetric bias.

L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood", biastype = "onesidedBias"

used by getInfClip for onesided bias.

L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood", biastype = "asymmetricBias"

used by getInfClip for asymmetric bias.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions 22, 201-223.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asGRisk-class, asMSE-class, asUnOvShoot-class, ContIC-class, TotalVarIC-class

[Package ROptEst version 0.6.0 Index]