getInfGammaRegTS(ROptRegTS)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound. This function is rarely called directly. It is called by getInfClipRegTS to compute optimally robust ICs.

Usage


getInfGammaRegTS(ErrorL2deriv, Regressor, risk, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution, asMSE, ContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution, asMSE,
##   Av1CondContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution, asMSE,
##   Av1CondTotalVarNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   MultivariateDistribution, asMSE, ContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   MultivariateDistribution, asMSE,
##   Av1CondContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   MultivariateDistribution, asMSE,
##   Av1CondTotalVarNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   Distribution, asMSE, Av2CondContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, z.comp, stand, cent, clip)

## S4 method for signature 'RealRandVariable, Distribution,
##   asMSE, ContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorDistr, stand, cent, clip)

## S4 method for signature 'RealRandVariable, Distribution,
##   asMSE, Av1CondContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorDistr, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution, asUnOvShoot,
##   ContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution, asUnOvShoot,
##   TotalVarNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   numeric, asUnOvShoot, CondContNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, clip)

## S4 method for signature 'UnivariateDistribution,
##   numeric, asUnOvShoot, CondTotalVarNeighborhood':
getInfGammaRegTS(ErrorL2deriv, 
                Regressor, risk, neighbor, clip)

Arguments

ErrorL2deriv 	 L2-derivative of ErrorDistr.

Regressor 	 regressor.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant/function.

stand 	 standardizing matrix.

z.comp 	 which components of the centering constant/function have to be computed.

ErrorDistr 	 error distribution.

Details

The function is used in case of asymptotic G-risks; confer Ruckdeschel and Rieder (2004).

Methods

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asMSE", neighbor = "Av1CondContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asMSE", neighbor = "Av1CondTotalVarNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asMSE", neighbor = "Av1CondContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asMSE", neighbor = "Av1CondTotalVarNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asMSE", neighbor = "Av2CondContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asMSE", neighbor = "ContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asMSE", neighbor = "Av1CondContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "ContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "TotalVarNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "numeric", risk = "asUnOvShoot", neighbor = "CondContNeighborhood"

used by getInfClipRegTS.

ErrorL2deriv = "UnivariateDistribution", Regressor = "numeric", risk = "asUnOvShoot", neighbor = "CondTotalVarNeighborhood"

used by getInfClipRegTS.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions (submitted).

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asMSE-class, asUnOvShoot-class, ContIC-class, Av1CondContIC-class, Av2CondContIC-class, Av1CondTotalVarIC-class, CondContIC-class, CondTotalVarIC-class

[Package ROptRegTS version 0.6.1 Index]