getInfRobIC(ROptEst)	R Documentation

Generic Function for the Computation of Optimally Robust ICs

Description

Generic function for the computation of optimally robust ICs in case of infinitesimal robust models. This function is rarely called directly.

Usage


getInfRobIC(L2deriv, risk, neighbor, ...)

## S4 method for signature 'UnivariateDistribution, asCov,
##   ContNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution, asCov,
##   TotalVarNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, Finfo, trafo)

## S4 method for signature 'RealRandVariable, asCov,
##   ContNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, Distr, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution, asBias,
##   UncondNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, symm, trafo, 
             maxiter, tol)

## S4 method for signature 'RealRandVariable, asBias,
##   ContNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, Distr,  
             L2derivDistrSymm, z.start, A.start, trafo, maxiter, tol)

## S4 method for signature 'UnivariateDistribution,
##   asHampel, UncondNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, symm, Finfo, trafo, 
             upper, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable, asHampel,
##   ContNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, Distr, DistrSymm, L2derivSymm, 
             L2derivDistrSymm, Finfo, z.start, A.start, trafo, upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,
##   asGRisk, UncondNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, symm, Finfo, trafo, 
             upper, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable, asGRisk,
##   ContNeighborhood':
getInfRobIC(L2deriv, risk, neighbor,  Distr, DistrSymm, L2derivSymm, 
             L2derivDistrSymm, Finfo, z.start, A.start, trafo, upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,
##   asUnOvShoot, UncondNeighborhood':
getInfRobIC(L2deriv, risk, neighbor, symm, Finfo, trafo, 
             upper, maxiter, tol, warn)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

Distr 	 object of class "Distribution".

symm 	 logical: indicating symmetry of L2deriv.

DistrSymm 	 object of class "DistributionSymmetry".

L2derivSymm 	 object of class "FunSymmList".

L2derivDistrSymm 	 object of class "DistrSymmList".

Finfo 	 Fisher information matrix.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

trafo 	 matrix: transformation of the parameter.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Value

The optimally robust IC is computed.

Methods

L2deriv = "UnivariateDistribution", risk = "asCov", neighbor = "ContNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "UnivariateDistribution", risk = "asCov", neighbor = "TotalVarNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asCov", neighbor = "ContNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "UnivariateDistribution", risk = "asBias", neighbor = "UncondNeighborhood"

computes the bias optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asBias", neighbor = "ContNeighborhood"

computes the bias optimal influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "UnivariateDistribution", risk = "asHampel", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asHampel", neighbor = "ContNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "UnivariateDistribution", risk = "asGRisk", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asGRisk", neighbor = "ContNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for one-dimensional L2 differentiable parametric families and asymptotic under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106-115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions 22: 201-223.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfRobModel-class

[Package ROptEst version 0.6.0 Index]