getInfRobRegTypeIC(ROptRegTS)	R Documentation

Generic Function for the Computation of Optimally Robust Regression-Type ICs

Description

Generic function for the computation of optimally robust regression-type ICs in case of infinitesimal robust models. This function is rarely called directly.

Usage


getInfRobRegTypeIC(ErrorL2deriv, Regressor, risk, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asBias,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asBias,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asBias,Av1CondTotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,Distribution,asBias,Av2CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asCov,TotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,CondTotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,Av2CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asCov,Av1CondTotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,asGRisk,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,Distribution,asGRisk,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,Distribution,asGRisk,Av2CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,Distribution,asGRisk,Av1CondTotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,asBias,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,asBias,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,asBias,Av1CondTotalVarNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable,Distribution,asBias,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorSymm, RegSymm, ErrorDistr, ErrorL2derivSymm, 
                ErrorL2derivDistrSymm, Finfo, trafo, upper, z.start, A.start, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable,Distribution,asBias,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorSymm, RegSymm, ErrorDistr, ErrorL2derivSymm, 
                ErrorL2derivDistrSymm, Finfo, trafo, upper, z.start, A.start, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable,Distribution,asCov,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorDistr, Finfo, trafo)

## S4 method for signature 'RealRandVariable,Distribution,asCov,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorDistr, Finfo, trafo)

## S4 method for signature 'RealRandVariable,Distribution,asGRisk,ContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorSymm, RegSymm, ErrorDistr, ErrorL2derivSymm, 
                ErrorL2derivDistrSymm, Finfo, trafo, upper, z.start, A.start, maxiter, tol, warn)

## S4 method for signature 'RealRandVariable,Distribution,asGRisk,Av1CondContNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorSymm, RegSymm, ErrorDistr, ErrorL2derivSymm, 
                ErrorL2derivDistrSymm, Finfo, trafo, upper, z.start, A.start, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asUnOvShoot,UncondNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,asUnOvShoot,CondNeighborhood':
getInfRobRegTypeIC(ErrorL2deriv, 
                Regressor, risk, neighbor, ErrorL2derivDistrSymm, RegSymm, Finfo, trafo, 
                upper, maxiter, tol, warn)

Arguments

ErrorL2deriv 	 L2-derivative of ErrorDistr.

Regressor 	 regressor.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

ErrorSymm 	 symmetry of ErrorDistr.

ErrorL2derivDistrSymm 	 symmetry of ErrorL2derivDistr.

RegSymm 	 symmetry of RegDistr.

ErrorDistr 	 error distribution.

ErrorL2derivSymm 	 symmetry of ErrorL2deriv.

Finfo 	 Fisher information matrix.

trafo 	 matrix: transformation of the parameter.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

z.start 	 initial value for the centering constant/function.

A.start 	 initial value for the standardizing matrix.

Value

The optimally robust IC is computed.

Methods

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asBias", neighbor = "ContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asBias", neighbor = "Av1CondContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asBias", neighbor = "Av1CondTotalVarNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asBias", neighbor = "Av2CondContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "ContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asCov", neighbor = "TotalVarNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "CondContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "CondTotalVarNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "Av1CondContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "Av2CondContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asCov", neighbor = "Av1CondTotalVarNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asGRisk", neighbor = "ContNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asGRisk", neighbor = "Av1CondContNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asGRisk", neighbor = "Av2CondContNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", risk = "asGRisk", neighbor = "Av1CondTotalVarNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asBias", neighbor = "ContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asBias", neighbor = "Av1CondContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", risk = "asBias", neighbor = "Av1CondTotalVarNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asBias", neighbor = "ContNeighborhood"

computes the bias optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asCov", neighbor = "ContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asCov", neighbor = "Av1CondContNeighborhood"

computes the classical optimal influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asGRisk", neighbor = "ContNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", risk = "asGRisk", neighbor = "Av1CondContNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "CondNeighborhood"

computes the optimally robust influence curve for L2 differentiable regression-type families.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

[Package ROptRegTS version 0.6.1 Index]