getInfStand(ROptEst)	R Documentation

Generic Function for the Computation of the Standardizing Matrix

Description

Generic function for the computation of the standardizing matrix which takes care of the Fisher consistency of the corresponding IC. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfStand(L2deriv, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood':
getInfStand(L2deriv, neighbor, clip, cent, trafo)

## S4 method for signature 'UnivariateDistribution,
##   TotalVarNeighborhood':
getInfStand(L2deriv, neighbor, clip, cent, trafo)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood':
getInfStand(L2deriv, neighbor, Distr, A.comp, stand, clip, cent, trafo)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood"

... 	 additional parameters

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

stand 	 standardizing matrix.

Distr 	 object of class "Distribution".

trafo 	 matrix: transformation of the parameter.

A.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

Value

The standardizing matrix is computed.

Methods

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood"

computes standardizing matrix.

L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood"

computes standardizing matrix.

L2deriv = "RealRandVariable", neighbor = "ContNeighborhood"

computes standardizing matrix.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.5.0 Index]