getInfStandRegTS(ROptRegTS)	R Documentation

Generic Function for the Computation of the Standardizing Matrix

Description

Generic function for the computation of the standardizing matrix which takes care of the Fisher consistency of the corresponding IC. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfStandRegTS(ErrorL2deriv, Regressor, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,ContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,TotalVarNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, clip, cent)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,CondTotalVarNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, clip, cent)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,Av1CondContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,UnivariateDistribution,Av1CondTotalVarNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,ContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,Av1CondContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,MultivariateDistribution,Av1CondTotalVarNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'UnivariateDistribution,Distribution,Av2CondContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, z.comp, clip, cent, stand, trafo)

## S4 method for signature 'RealRandVariable,Distribution,ContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, ErrorDistr, A.comp, stand, clip, cent, trafo)

## S4 method for signature 'RealRandVariable,Distribution,Av1CondContNeighborhood':
getInfStandRegTS(ErrorL2deriv, 
		Regressor, neighbor, ErrorDistr, A.comp, stand, clip, cent, trafo)

Arguments

ErrorL2deriv 	 L2-derivative of ErrorDistr.

Regressor 	 regressor.

neighbor 	 object of class "Neighborhood".

... 	 additional parameters.

ErrorDistr 	 error distribution.

clip 	 optimal clipping bound/function.

cent 	 optimal centering constant/function.

stand 	 standardizing matrix.

z.comp 	 which components of the centering constant/function have to be computed.

A.comp 	 which components of the standardizing matrix have to be computed.

trafo 	 matrix: transformation of the parameter.

Value

The standardizing matrix is computed.

Methods

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "ContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "TotalVarNeighborhood"

computes standardizing constant.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "CondTotalVarNeighborhood"

computes standardizing constant.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "UnivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "ContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "MultivariateDistribution", neighbor = "Av1CondTotalVarNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "UnivariateDistribution", Regressor = "Distribution", neighbor = "Av2CondContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "ContNeighborhood"

computes standardizing matrix.

ErrorL2deriv = "RealRandVariable", Regressor = "Distribution", neighbor = "Av1CondContNeighborhood"

computes standardizing matrix.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class, , , , ,

[Package ROptRegTS version 0.6.1 Index]