getInfV(ROptEst)	R Documentation

Generic Function for the Computation of the asymptotic Variance of a Hampel type IC

Description

Generic function for the computation of the optimal clipping bound in case of infinitesimal robust models. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfV(L2deriv, neighbor, biastype, ...)
## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, BiasType':
getInfV(L2deriv, 
         neighbor, biastype, clip, cent, stand)
## S4 method for signature 'UnivariateDistribution,
##   TotalVarNeighborhood, BiasType':
getInfV(L2deriv, 
         neighbor, biastype, clip, cent, stand)
## S4 method for signature 'RealRandVariable,
##   ContNeighborhood, BiasType':
getInfV(L2deriv, 
         neighbor, biastype, Distr, V.comp, cent, stand, 
         w)
## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, onesidedBias':
getInfV(L2deriv, 
         neighbor, biastype, clip, cent, stand)
## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, asymmetricBias':
getInfV(L2deriv, 
         neighbor, biastype, clip, cent, stand)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType"

... 	 additional parameters.

clip 	 positive real: clipping bound

cent 	 optimal centering constant.

stand 	 standardizing matrix.

Distr 	 standardizing matrix.

V.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

w 	 object of class RobWeight; current weight

Value

The asymptotic variance of an ALE to IC of Hampel type is computed.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.6.0 Index]