getModifyIC(ROptEst)	R Documentation

Generic Function for the Computation of Functions for Slot modifyIC

Description

This function is used by internal computations and is rarely called directly.

Usage


getModifyIC(L2FamIC, neighbor, risk,...)
## S4 method for signature 'L2ParamFamily,Neighborhood,asRisk':
getModifyIC(L2FamIC,
          neighbor, risk, ...)
## S4 method for signature 'L2LocationFamily,UncondNeighborhood,asGRisk':
getModifyIC(L2FamIC,
          neighbor, risk, ...)
## S4 method for signature 'L2LocationFamily,UncondNeighborhood,fiUnOvShoot':
getModifyIC(L2FamIC,
          neighbor, risk, ...)
## S4 method for signature 'L2ScaleFamily,ContNeighborhood,asGRisk':
getModifyIC(L2FamIC,
          neighbor, risk, ...)
## S4 method for signature 'L2ScaleFamily,TotalVarNeighborhood,asGRisk':
getModifyIC(L2FamIC,
          neighbor, risk, ...)
## S4 method for signature 'L2LocationScaleFamily,ContNeighborhood,asGRisk':
getModifyIC(L2FamIC,
          neighbor, risk, ...)

Arguments

L2FamIC 	 object of class L2ParamFamily.

neighbor 	 object of class "Neighborhood".

risk 	 object of class "RiskType"

... 	 further arguments to be passed over to optIC.

Details

This function is used for internal computations. By setting RobAStBaseOption("all.verbose" = TRUE) somewhere globally, the generated function modifyIC will generate calls to optIC with argument verbose=TRUE.

Value

Function for slot modifyIC of ICs

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

, IC-class

[Package ROptEst version 0.7 Index]