getRiskIC(ROptEstOld)	R Documentation

Generic function for the computation of a risk for an IC

Description

Generic function for the computation of a risk for an IC.

Usage


getRiskIC(IC, risk, neighbor, L2Fam, ...)

## S4 method for signature 'IC, asCov, missing, missing':
getRiskIC(IC, risk, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asCov, missing,
##   L2ParamFamily':
getRiskIC(IC, risk, L2Fam, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, trAsCov, missing, missing':
getRiskIC(IC, risk, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, trAsCov, missing,
##   L2ParamFamily':
getRiskIC(IC, risk, L2Fam, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asBias, ContNeighborhood,
##   missing':
getRiskIC(IC, risk, neighbor, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asBias, ContNeighborhood,
##   L2ParamFamily':
getRiskIC(IC, risk, neighbor, L2Fam, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asBias,
##   TotalVarNeighborhood, missing':
getRiskIC(IC, risk, neighbor, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asBias,
##   TotalVarNeighborhood, L2ParamFamily':
getRiskIC(IC, risk, neighbor, L2Fam, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asMSE, UncondNeighborhood,
##   missing':
getRiskIC(IC, risk, neighbor, tol = .Machine$double.eps^0.25)

## S4 method for signature 'IC, asMSE, UncondNeighborhood,
##   L2ParamFamily':
getRiskIC(IC, risk, neighbor, L2Fam, tol = .Machine$double.eps^0.25)

## S4 method for signature 'TotalVarIC, asUnOvShoot,
##   UncondNeighborhood, missing':
getRiskIC(IC, risk, neighbor)

## S4 method for signature 'IC, fiUnOvShoot,
##   ContNeighborhood, missing':
getRiskIC(IC, risk, neighbor, sampleSize, Algo = "A", cont = "left")

## S4 method for signature 'IC, fiUnOvShoot,
##   TotalVarNeighborhood, missing':
getRiskIC(IC, risk, neighbor, sampleSize, Algo = "A", cont = "left")

Arguments

IC 	 object of class "InfluenceCurve"

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

L2Fam 	 object of class "L2ParamFamily".

... 	 additional parameters

tol 	 the desired accuracy (convergence tolerance).

sampleSize 	 integer: sample size.

Algo 	 "A" or "B".

cont 	 "left" or "right".

Details

To make sure that the results are valid, it is recommended to include an additional check of the IC properties of IC using checkIC.

Value

The risk of an IC is computed.

Methods

IC = "IC", risk = "asCov", neighbor = "missing", L2Fam = "missing"

asymptotic covariance of IC.

IC = "IC", risk = "asCov", neighbor = "missing", L2Fam = "L2ParamFamily"

asymptotic covariance of IC under L2Fam.

IC = "IC", risk = "trAsCov", neighbor = "missing", L2Fam = "missing"

asymptotic covariance of IC.

IC = "IC", risk = "trAsCov", neighbor = "missing", L2Fam = "L2ParamFamily"

asymptotic covariance of IC under L2Fam.

IC = "IC", risk = "asBias", neighbor = "ContNeighborhood", L2Fam = "missing"

asymptotic bias of IC under convex contaminations.

IC = "IC", risk = "asBias", neighbor = "ContNeighborhood", L2Fam = "L2ParamFamily"

asymptotic bias of IC under convex contaminations and L2Fam.

IC = "IC", risk = "asBias", neighbor = "TotalVarNeighborhood", L2Fam = "missing"

asymptotic bias of IC in case of total variation neighborhoods.

IC = "IC", risk = "asBias", neighbor = "TotalVarNeighborhood", L2Fam = "L2ParamFamily"

asymptotic bias of IC under L2Fam in case of total variation neighborhoods.

IC = "IC", risk = "asMSE", neighbor = "UncondNeighborhood", L2Fam = "missing"

asymptotic mean square error of IC.

IC = "IC", risk = "asMSE", neighbor = "UncondNeighborhood", L2Fam = "L2ParamFamily"

asymptotic mean square error of IC under L2Fam.

IC = "TotalVarIC", risk = "asUnOvShoot", neighbor = "UncondNeighborhood", L2Fam = "missing"

asymptotic under-/overshoot risk of IC.

IC = "IC", risk = "fiUnOvShoot", neighbor = "ContNeighborhood", L2Fam = "missing"

finite-sample under-/overshoot risk of IC.

IC = "IC", risk = "fiUnOvShoot", neighbor = "TotalVarNeighborhood", L2Fam = "missing"

finite-sample under-/overshoot risk of IC.

Note

This generic function is still under construction.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

Ruckdeschel, P. and Kohl, M. (2005) Computation of the Finite Sample Risk of M-estimators on Neighborhoods.

See Also

getRiskIC-methods, InfRobModel-class

[Package ROptEstOld version 0.5.2 Index]