getInfLM(ROptEst)	R Documentation

Functions to determine Lagrange multipliers

Description

Functions to determine Lagrange multipliers A and a in a Hampel problem or in a(n) (inner) loop in a MSE problem; can be done either by optimization or by fixed point iteration. These functions are rarely called directly.

Usage


getLagrangeMultByIter(b, L2deriv, risk, trafo,
                      neighbor, biastype, normtype, Distr,
                      a.start, z.start, A.start, w.start, std, z.comp,
                      A.comp, maxiter, tol,
                      verbose = getRobAStBaseOption("all.verbose"),
                      warnit = TRUE)
getLagrangeMultByOptim(b, L2deriv, risk, FI, trafo,
                      neighbor, biastype, normtype, Distr,
                      a.start, z.start, A.start, w.start,  std, z.comp,
                      A.comp, maxiter, tol,
                      verbose = getRobAStBaseOption("all.verbose"), ...)


Arguments

b 	 numeric; (>b_min; clipping bound for which the Lagrange multipliers are searched

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

FI 	 matrix: Fisher information.

trafo 	 matrix: transformation of the parameter.

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType" — the bias type with we work.

normtype 	 object of class "NormType" — the norm type with we work.

Distr 	 object of class "Distribution".

a.start 	 initial value for the centering constant (in p-space).

z.start 	 initial value for the centering constant (in k-space).

A.start 	 initial value for the standardizing matrix.

w.start 	 initial value for the weight function.

std 	 matrix of (or which may coerced to) class PosSemDefSymmMatrix for use of different (standardizing) norm.

z.comp 	 logical vector: indication which components of the centering constant have to be computed.

A.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

verbose 	 logical: if TRUE, some messages are printed.

warnit 	 logical: if TRUE warning is issued if maximal number of iterations is reached.

... 	 additional parameters for optim.

Value

a list with items

A 	 Lagrange multiplier A (standardizing matrix)

a 	 Lagrange multiplier a (centering in p-space)

z 	 Lagrange multiplier z (centering in k-space)

w 	 weight function involving Lagrange multipliers

biastype 	 (possibly modified) bias type biastype from argument

normtype 	 (possibly modified) norm type normtype from argument

normtype.old 	 (possibly modified) norm type normtype before last (internal) update

risk 	 (possibly [norm-]modified) risk risk from argument

std 	 (possibly modified) argument std

iter 	 number of iterations needed

prec 	 precision achieved

b 	 used clippng height b

call 	 call with which either getLagrangeMultByIter or getLagrangeMultByOptim was called

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106-115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions 22: 201-223.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfRobModel-class

[Package ROptEst version 0.7 Index]