getweight-methods(RobAStBase)	R Documentation

Generating weights

Description

Generates weight functions of Hampel / BdSt type for different bias and norm types.

Usage


getweight(Weight, neighbor, biastype, ...)
minbiasweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   BiasType':
getweight(Weight, neighbor, biastype, normW)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   BiasType':
minbiasweight(Weight, neighbor, biastype, normW)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   onesidedBias':
getweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   onesidedBias':
minbiasweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   asymmetricBias':
getweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'HampelWeight, ContNeighborhood,
##   asymmetricBias':
minbiasweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'BdStWeight,
##   TotalVarNeighborhood, BiasType':
getweight(Weight, neighbor, biastype, ...)
## S4 method for signature 'BdStWeight,
##   TotalVarNeighborhood, BiasType':
minbiasweight(Weight, neighbor, biastype, ...)

Arguments

Weight 	 Object of class "RobWeight".

neighbor 	 Object of class "Neighborhood".

biastype 	 Object of class "BiasType".

normW 	 Object of class "NormType" — only for signature HampelWeight,ContNeighborhood,BiasType.

... 	 possibly additional (unused) arguments — like in a call to the less specific methods.

Details

These functions generate the weight function in slot weight in a corresp. object of class RobWeight and descendants.

Value

Object of class "HampelWeight" resp. "BdStWeight"

Methods

getweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "BiasType"
with additional argument biastype of class "BiasType": produces weight slot...

minbiasweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "BiasType" with additional argument biastype of class "BiasType": produces weight slot...

getweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "onesidedBias": produces weight slot...

minbiasweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "onesidedBias": produces weight slot...

getweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "asymmetricBias": produces weight slot...

minbiasweight

signature(Weight = "HampelWeight", neighbor = "ContNeighborhood", biastype = "asymmetricBias": produces weight slot...

getweight

signature(Weight = "BdStWeight", neighbor = "TotalVarNeighborhood", biastype = "BiasType": produces weight slot...

minbiasweight

signature(Weight = "BdStWeight", neighbor = "TotalVarNeighborhood", biastype = "BiasType": produces weight slot...

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Hampel et al. (1986) Robust Statistics. The Approach Based on Influence Functions. New York: Wiley.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BdStWeight-class, HampelWeight-class, IC-class

Examples



[Package RobAStBase version 0.1.0 Index]