internals_for_RobAStBase(RobAStBase)	R Documentation

Internal / Helper functions of package RobAStBase

Description

These functions are used internally by package RobAStBase.

Usage


.eq(x,y = 0*x, tol = 1e-7)
.getDistr(L2Fam)

Arguments

x 	 a (numeric) vector

y 	 a (numeric) vector

tol 	 numeric — tolerance

L2fam 	 object of class L2ParamFamily

Details

.eqchecks equality of two vectors up to a given precision; .getDistr produces a string with the class of the family and its parameter value;

Value

.eq 	 (a vector of) logical.

.getDistr 	 character.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

[Package RobAStBase version 0.2 Index]