internals_for_RobAStBase_ddPlot(RobAStBase)	R Documentation

Internal / Helper functions of package RobAStBase for ddPlot

Description

This function is an internally used helper function for ddPlot in package RobAStBase.

Usage


.ddPlot.MatNtNtCoCo(data, ...,  dist.x = NormType(), dist.y  = NormType(),
       cutoff.x = cutoff(norm = dist.x, cutoff.quantile  = cutoff.quantile.x),
       cutoff.y = cutoff(norm = dist.y, cutoff.quantile  = cutoff.quantile.y),
       cutoff.quantile.x = 0.95, cutoff.quantile.y = cutoff.quantile.x,
       transform.x, transform.y = transform.x,
       id.n, lab.pts, adj =0, cex.idn = 1,
       col.idn = par("col"), lty.cutoff,
       lwd.cutoff, col.cutoff = "red")

Arguments

data 	 data in matrix form (columns are observations; rows are variable dimensions) at which to produce the ddPlot.

... 	 further arguments to be passed to plot.default, text, and abline

dist.x 	 object of class NormType; the distance for the x axis.

dist.y 	 object of class NormType; the distance for the y axis.

cutoff.x 	 object of class cutoff; the cutoff information for the x axis (the vertical line discriminating 'good' and 'bad' points).

cutoff.y 	 object of class cutoff; the cutoff information for the y axis (the horizontal line discriminating 'good' and 'bad' points).

cutoff.quantile.x 	 numeric; the cutoff quantile for the x axis.

cutoff.quantile.y 	 numeric; the cutoff quantile for the y axis.

transform.x 	 function; a transformation to be performed before determining the distances of the x axis.

transform.y 	 function; a transformation to be performed before determining the distances of the y axis.

id.n 	 a set of indices (or a corresponding logical vector); to select a subset of the data in argument data.

lab.pts 	 a vector of labels for the (unsubsetted) data.

adj 	 the corresponding argument for text for labelling the outliers.

cex.idn 	 the corresponding cex argument for text for labelling the outliers.

col.idn 	 the corresponding col argument for text for labelling the outliers.

lty.cutoff 	 the corresponding lty argument for abline for drawing the cutoff lines.

lwd.cutoff 	 the corresponding lwd argument for abline for drawing the cutoff lines.

col.cutoff 	 the corresponding col argument for abline for drawing the cutoff lines.

Details

performs the plotting for ddPlot and outlyingPlotIC; all arguments except for data are optional. In case they are missing default values are used as usual; for those arguments without default arguments, we do

transform.x

defaults to identity, internally
.
id.n

defaults to 1:ncol(data), internally
.
lab.pts

defaults to (1:ncol(data))[id.n], internally
.
lwd.cutoff

defaults to argument lwd, if given, else to par{lwd}, internally
.
lty.cutoff

defaults to argument lty, if given, else to par{lty}, internally
.

Value

a list with items

id.x 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff

id.y 	 the indices of (possibly transformed) data (within subset id.n) beyond the y-cutoff

id.xy 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff and the y-cutoff

qtx 	 the quantiles of the distances of the (possibly transformed) data in x direction

qty 	 the quantiles of the distances of the (possibly transformed) data in y direction

cutoff.x.v 	 the cutoff value in x direction

cutoff.y.v 	 the cutoff value in y direction

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

plot.default, par, ddPlot, outlyingPlotIC

Examples


MX <- matrix(rnorm(1500),nrow=6)
QM <- matrix(rnorm(36),nrow=6); QM <- QM %*% t(QM)
RobAStBase:::.ddPlot.MatNtNtCoCo(data=MX, 
        dist.y=QFNorm(QuadF=PosSemDefSymmMatrix(QM)),
        xlab="Norm.x",ylab="Norm.y", cex.idn = 1.3, offset=0,
        lwd=2, lwd.cutoff=4, lty=2, col.cutoff =2, col.idn="green",
        col = "blue", adj=0.4, pos=4,id.n = sample(1:200,size=100),
        lab.pts=letters,log="x", main="GA", sub="NO",cex.sub=0.2)

[Package RobAStBase version 0.2 Index]