kStepEstimator(RobAStBase)	R Documentation

Generic function for the computation of k-step estimates

Description

Generic function for the computation of k-step estimates.

Usage


kStepEstimator(x, IC, start, ...)

## S4 method for signature 'numeric, IC, numeric':
kStepEstimator(x, IC, start, steps = 1L, 
    useLast = getRobAStBaseOption("kStepUseLast"))
## S4 method for signature 'matrix, IC, numeric':
kStepEstimator(x, IC, start, steps = 1L, 
    useLast = getRobAStBaseOption("kStepUseLast"))
## S4 method for signature 'numeric, IC, Estimate':
kStepEstimator(x, IC, start, steps = 1L, 
    useLast = getRobAStBaseOption("kStepUseLast"))
## S4 method for signature 'matrix, IC, Estimate':
kStepEstimator(x, IC, start, steps = 1L, 
    useLast = getRobAStBaseOption("kStepUseLast"))

Arguments

x 	 sample

IC 	 object of class "IC"

start 	 initial estimate

steps 	 integer: number of steps

useLast 	 which parameter estimate (initial estimate or k-step estimate) shall be used to fill the slots pIC, asvar and asbias of the return value.

... 	 additional parameters

Details

Given an initial estimation start, a sample x and an influence curve IC the corresponding k-step estimator is computed.

The default value of argument useLast is set by the global option kStepUseLast which by default is set to FALSE. In case of general models useLast remains unchanged during the computations. However, if slot CallL2Fam of IC generates an object of class "L2GroupParamFamily" the value of useLast is changed to TRUE. Explicitly setting useLast to TRUE should be done with care as in this situation the influence curve is re-computed using the value of the one-step estimate which may take quite a long time depending on the model.

If useLast is set to TRUE and slot modifyIC of IC is filled with some function (which can be used to re-compute the IC for a different parameter), the computation of asvar, asbias and IC is based on the k-step estimate.

Value

Object of class "kStepEstimate".

Methods

x = "numeric", IC = "IC", start = "numeric"

univariate samples.

x = "matrix", IC = "IC", start = "numeric"

multivariate samples.

x = "matrix", IC = "IC", start = "Estimate"

multivariate samples.

x = "matrix", IC = "IC", start = "Estimate"

multivariate samples.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class, kStepEstimate-class

[Package RobAStBase version 0.1.0 Index]