leastFavorableRadius(ROptEst)	R Documentation

Generic Function for the Computation of Least Favorable Radii

Description

Generic function for the computation of least favorable radii.

Usage


leastFavorableRadius(L2Fam, neighbor, risk, ...)

## S4 method for signature 'L2ParamFamily,
##   UncondNeighborhood, asGRisk':
leastFavorableRadius(
          L2Fam, neighbor, risk, rho, upRad = 1, 
            z.start = NULL, A.start = NULL, upper = 100, maxiter = 100, 
            tol = .Machine$double.eps^0.4, warn = FALSE, verbose = FALSE)

Arguments

L2Fam 	 L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

risk 	 object of class "RiskType".

... 	 additional parameters

upRad 	 the upper end point of the radius interval to be searched.

rho 	 The considered radius interval is: [r*rho, r/rho] with 0 < rho < 1.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

verbose 	 logical: if TRUE, some messages are printed

Value

The least favorable radius and the corresponding inefficiency are computed.

Methods

L2Fam = "L2ParamFamily", neighbor = "UncondNeighborhood", risk = "asGRisk"

computation of the least favorable radius.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H., Kohl, M. and Ruckdeschel, P. (2008) The Costs of not Knowing the Radius. Statistical Methods and Applications 17(1) 13-40.

Rieder, H., Kohl, M. and Ruckdeschel, P. (2001) The Costs of not Knowing the Radius. Submitted. Appeared as discussion paper Nr. 81. SFB 373 (Quantification and Simulation of Economic Processes), Humboldt University, Berlin; also available under www.uni-bayreuth.de/departments/math/org/mathe7/RIEDER/pubs/RR.pdf

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

radiusMinimaxIC

Examples


N <- NormLocationFamily(mean=0, sd=1) 
leastFavorableRadius(L2Fam=N, neighbor=ContNeighborhood(),
                     risk=asMSE(), rho=0.5)

[Package ROptEst version 0.6.0 Index]