locMEstimator(ROptEstOld)	R Documentation

Generic function for the computation of location M estimators

Description

Generic function for the computation of location M estimators.

Usage


locMEstimator(x, IC, ...)

## S4 method for signature 'numeric, InfluenceCurve':
locMEstimator(x, IC, eps = .Machine$double.eps^0.5)

Arguments

x 	 sample

IC 	 object of class "InfluenceCurve"

... 	 additional parameters

eps 	 the desired accuracy (convergence tolerance).

Value

Returns a list with component

loc 	 M estimator of location

Methods

x = "numeric", IC = "InfluenceCurve"

univariate location.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1964) Robust estimation of a location parameter. Ann. Math. Stat. 35: 73–101.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class

[Package ROptEstOld version 0.5.2 Index]