minmaxBias(ROptEst)	R Documentation

Generic Function for the Computation of Bias-Optimally Robust ICs

Description

Generic function for the computation of bias-optimally robust ICs in case of infinitesimal robust models. This function is rarely called directly.

Usage


minmaxBias(L2deriv, neighbor, biastype, ...)

## S4 method for signature 'UnivariateDistribution,ContNeighborhood,BiasType':
minmaxBias(L2deriv, neighbor, biastype, symm, trafo, 
             maxiter, tol, warn, Finfo)

## S4 method for signature 'UnivariateDistribution,ContNeighborhood,asymmetricBias':
minmaxBias(L2deriv, neighbor, biastype, symm, trafo, 
             maxiter, tol, warn, Finfo)

## S4 method for signature 'UnivariateDistribution,ContNeighborhood,onesidedBias':
minmaxBias(L2deriv, neighbor, biastype, symm, trafo, 
             maxiter, tol, warn, Finfo)

## S4 method for signature 'UnivariateDistribution,TotalVarNeighborhood,BiasType':
minmaxBias(L2deriv, neighbor, biastype, symm, trafo, 
             maxiter, tol, warn, Finfo)

## S4 method for signature 'RealRandVariable,ContNeighborhood,BiasType':
minmaxBias(L2deriv, neighbor, biastype, normtype, Distr, 
             z.start, A.start,  z.comp, A.comp, Finfo, trafo, maxiter, tol, verbose = FALSE)

## S4 method for signature 'RealRandVariable,TotalVarNeighborhood,BiasType':
minmaxBias(L2deriv, neighbor, biastype, normtype, Distr,
             z.start, A.start,  z.comp, A.comp, Finfo, trafo, maxiter, tol, verbose = FALSE)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType".

normtype 	 object of class "NormType".

... 	 additional parameters.

Distr 	 object of class "Distribution".

symm 	 logical: indicating symmetry of L2deriv.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

z.comp 	 logical indicator which indices need to be computed and which are 0 due to symmetry.

A.comp 	 matrix of logical indicator which indices need to be computed and which are 0 due to symmetry.

trafo 	 matrix: transformation of the parameter.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Finfo 	 Fisher information matrix.

verbose 	 logical: if TRUE, some messages are printed

Value

The bias-optimally robust IC is computed.

Methods

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "BiasType"

computes the bias optimal influence curve for symmetric bias for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "asymmetricBias"

computes the bias optimal influence curve for asymmetric bias for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood", biastype = "BiasType"

computes the bias optimal influence curve for symmetric bias for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "BiasType"

computes the bias optimal influence curve for symmetric bias for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "RealRandVariable", neighbor = "TotalNeighborhood", biastype = "BiasType"

computes the bias optimal influence curve for symmetric bias for L2 differentiable parametric families in a setting where we are interested in a p=1 dimensional aspect of an unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfRobModel-class

[Package ROptEst version 0.7 Index]