onesidedBiasType-class(ROptEst)	R Documentation

onesided Bias Type

Description

Class of onesided bias types.

Objects from the Class

Objects can be created by calls of the form new("onesidedBiasType", ...). More frequently they are created via the generating function positiveBias or negativeBias.

Slots

name:

Object of class "character".

sign:

Object of class "numeric"; to be {-1,1} — whether bias is to be positive or negative

Methods

sign

signature(object = "onesidedBiasType"): accessor function for slot sign.

sign<-

signature(object = "onesidedBiasType", value = "numeric"): replacement function for slot sign.

Extends

Class "BiasType", directly.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BiasType-class

Examples


positiveBias()
## The function is currently defined as
function(){ new("onesidedBiasType", name = "positive Bias", sign = 1) }

negativeBias()
## The function is currently defined as
function(){ new("onesidedBiasType", name = "negative Bias", sign = -1) }

pB <- positiveBias()
sign(pB)
try(sign(pB) <- -2) ## error
sign(pB) <- -1

[Package ROptEst version 0.5.0 Index]