optIC(RobAStBase)	R Documentation

Generic function for the computation of optimally robust ICs

Description

Generic function for the computation of optimally robust ICs.

Usage


optIC(model, risk, ...)

## S4 method for signature 'L2ParamFamily, asCov':
optIC(model, risk)

Arguments

model 	 probability model.

risk 	 object of class "RiskType".

... 	 additional parameters.

Details

The classical optimal IC which ist optimal in sense of the Cramer-Rao bound is computed.

Value

Some optimally robust IC is computed.

Methods

model = "L2ParamFamily", risk = "asCov"

computes classical optimal influence curve for L2 differentiable parametric families.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, RiskType-class

Examples


B <- BinomFamily(size = 25, prob = 0.25) 

## classical optimal IC
IC0 <- optIC(model = B, risk = asCov())
plot(IC0) # plot IC
checkIC(IC0, B)

[Package RobAStBase version 0.2 Index]