optIC-methods(ROptRegTS)	R Documentation

Methods for Function optIC in Package ‘ROptRegTS’

Description

Methods for function optIC in package ROptRegTS.

Usage


## S4 method for signature 'L2RegTypeFamily, asCov':
optIC(model, risk)

## S4 method for signature 'InfRobRegTypeModel, asRisk':
optIC(model, risk, z.start = NULL, A.start = NULL, upper = 1e4, 
             maxiter = 50, tol = .Machine$double.eps^0.4, warn = TRUE)

## S4 method for signature 'InfRobRegTypeModel,
##   asUnOvShoot':
optIC(model, risk, upper = 1e4, maxiter = 50, 
             tol = .Machine$double.eps^0.4, warn = TRUE)

## S4 method for signature 'FixRobRegTypeModel,
##   fiUnOvShoot':
optIC(model, risk, sampleSize, upper = 1e4, 
             maxiter = 50, tol = .Machine$double.eps^0.4, warn = TRUE, Algo = "A", cont = "left")

Arguments

model 	 probability model.

risk 	 object of class "RiskType".

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

sampleSize 	 integer: sample size.

Algo 	 "A" or "B".

cont 	 "left" or "right".

Details

In case of the finite-sample risk "fiUnOvShoot" one can choose between two algorithms for the computation of this risk where the least favorable contamination is assumed to be “left” or “right” of some boundary curve. For more details we refer to Subsections 12.1.3 and 12.2.3 of Kohl (2005).

Value

Some optimally robust IC is computed.

Methods

model = "L2RegTypeFamily", risk = "asCov"

computes classical optimal influence curve for L2 differentiable regression-type families.

model = "InfRobRegTypeModel", risk = "asRisk"

computes optimally robust influence curve for robust regression-type models with infinitesimal neighborhoods and various asymptotic risks.

model = "InfRobRegTypeModel", risk = "asUnOvShoot"

computes optimally robust influence curve for robust regression-type models with infinitesimal neighborhoods and asymptotic under-/overshoot risk.

model = "FixRobRegTypeModel", risk = "fiUnOvShoot"

computes optimally robust influence curve for robust regression-type models with fixed neighborhoods and finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

optIC

[Package ROptRegTS version 0.6.1 Index]