optRisk(ROptEst)	R Documentation

Generic function for the computation of the minimal risk

Description

Generic function for the computation of the optimal (i.e., minimal) risk for a probability model.

Usage


optRisk(model, risk, ...)

## S4 method for signature 'InfRobModel, asRisk':
optRisk(model, risk, z.start = NULL, A.start = NULL, upper = 1e4, 
             maxiter = 50, tol = .Machine$double.eps^0.4, warn = TRUE)

## S4 method for signature 'FixRobModel, fiUnOvShoot':
optRisk(model, risk, sampleSize, upper = 1e4, maxiter = 50, 
             tol = .Machine$double.eps^0.4, warn = TRUE, Algo = "A", cont = "left")

Arguments

model 	 probability model

risk 	 object of class RiskType

... 	 additional parameters

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

sampleSize 	 integer: sample size.

Algo 	 "A" or "B".

cont 	 "left" or "right".

Details

In case of the finite-sample risk "fiUnOvShoot" one can choose between two algorithms for the computation of this risk where the least favorable contamination is assumed to be left or right of some bound. For more details we refer to Section 11.3 of Kohl (2005).

Value

The minimal risk is computed.

Methods

model = "L2ParamFamily", risk = "asCov"

asymptotic covariance of L2 differentiable parameteric family.

model = "InfRobModel", risk = "asRisk"

asymptotic risk of a infinitesimal robust model.

model = "FixRobModel", risk = "fiUnOvShoot"

finite-sample under-/overshoot risk of a robust model with fixed neighborhood.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

RiskType-class

Examples


optRisk(model = NormLocationScaleFamily(), risk = asCov())

[Package ROptEst version 0.5.0 Index]