outlyingPlotIC(RobAStBase)	R Documentation

Function outlyingPlotIC in Package ‘RobAStBase’

Description

outlyingPlotIC produces an outlyingness plot based on distances applied to ICs

Usage


outlyingPlotIC(data, IC.x, IC.y = IC.x, dist.x = NormType(),
                         dist.y, cutoff.y = cutoff.chisq(), cutoff.x = cutoff.sememp(), ...,
                         cutoff.quantile.x = 0.95,
                         cutoff.quantile.y = cutoff.quantile.x,
                         id.n, lab.pts, adj, cex.idn,
                         col.idn, lty.cutoff, lwd.cutoff, col.cutoff,
                         main = gettext("Outlyingness by means of a distance-distance plot")
                         )

Arguments

data 	 data coercable to matrix; the data at which to produce the ddPlot.

IC.x 	 object of class IC the influence curve to produce the distances for the x axis.

IC.y 	 object of class IC the influence curve to produce the distances for the y axis.

... 	 further arguments to be passed to plot.default, text, and abline

dist.x 	 object of class NormType; the distance for the x axis.

dist.y 	 object of class NormType; the distance for the y axis.

cutoff.x 	 object of class cutoff; the cutoff information for the x axis (the vertical line discriminating 'good' and 'bad' points).

cutoff.y 	 object of class cutoff; the cutoff information for the y axis (the horizontal line discriminating 'good' and 'bad' points).

cutoff.quantile.x 	 numeric; the cutoff quantile for the x axis.

cutoff.quantile.y 	 numeric; the cutoff quantile for the y axis.

id.n 	 a set of indices (or a corresponding logical vector); to select a subset of the data in argument data.

lab.pts 	 a vector of labels for the (unsubsetted) data.

adj 	 the corresponding argument for text for labelling the outliers.

cex.idn 	 the corresponding cex argument for text for labelling the outliers.

col.idn 	 the corresponding col argument for text for labelling the outliers.

lty.cutoff 	 the corresponding lty argument for abline for drawing the cutoff lines.

lwd.cutoff 	 the corresponding lwd argument for abline for drawing the cutoff lines.

col.cutoff 	 the corresponding col argument for abline for drawing the cutoff lines.

main 	 the main title.

Details

calls a corresponding ddPlot method to produce the plot.

Value

a list with items

id.x 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff

id.y 	 the indices of (possibly transformed) data (within subset id.n) beyond the y-cutoff

id.xy 	 the indices of (possibly transformed) data (within subset id.n) beyond the x-cutoff and the y-cutoff

qtx 	 the quantiles of the distances of the (possibly transformed) data in x direction

qty 	 the quantiles of the distances of the (possibly transformed) data in y direction

cutoff.x.v 	 the cutoff value in x direction

cutoff.y.v 	 the cutoff value in y direction

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

Examples


if(require(ROptEst)){
## generates normal location and scale family with mean = -2 and sd = 3
N0 <- NormLocationScaleFamily()
N0.IC0 <- optIC(model = N0, risk = asCov())
N0.Rob1 <- InfRobModel(center = N0, neighbor = ContNeighborhood(radius = 0.5))
N0.IC1 <- optIC(model = N0.Rob1, risk = asMSE())
xn <- c(rnorm(100),rcauchy(20)+20)
outlyingPlotIC(xn, IC.x=N0.IC0)
outlyingPlotIC(xn, IC.x=N0.IC1)
}

[Package RobAStBase version 0.1.2 Index]