plot-methods(RobAStBase)	R Documentation

Methods for Function plot in Package ‘RobAStBase’

Description

plot-methods

Usage

plot(x, y, ...)
## S4 method for signature 'IC, missing':
plot(x, ..., withSweave = getdistrOption("withSweave"),
             main = FALSE, inner = TRUE, sub = FALSE,
             col.inner = par("col.main"), cex.inner = 0.8,
             bmar = par("mar")[1], tmar = par("mar")[3],
             mfColRow = TRUE)

Arguments

x 	 object of class "IC": IC to be plotted

y 	 missing

withSweave 	 logical: if TRUE (for working with Sweave) no extra device is opened

main 	 logical: is a main title to be used? or
just as argument main in plot.default.

inner 	 logical: panels have their own titles? or
character vector of / cast to length number of comparands: main in plot.default)

sub 	 logical: is a sub-title to be used? or
just as argument sub in plot.default.

tmar 	 top margin – useful for non-standard main title sizes

bmar 	 bottom margin – useful for non-standard sub title sizes

cex.inner 	 magnification to be used for inner titles relative to the current setting of cex; as in par

col.inner 	 character or integer code; color for the inner title

mfColRow 	 shall default partition in panels be used — defaults to TRUE

...

{further parameters for plot}

Details

Any parameters of plot.default may be passed on to this particular plot method.

For main-, inner, and subtitles given as arguments main, inner, and sub, top and bottom margins are enlarged to 5 resp. 6 by default but may also be specified by tmar / bmar arguments. If main / inner / sub are logical then if the respective argument is FALSE nothing is done/plotted, but if it is TRUE, we use a default main title taking up the calling arguments in case of main, default inner titles taking up the class and (named) parameter slots of arguments in case of inner, and a "generated on <data>"-tag in case of sub. Of course, if main / inner / sub are character, this is used for the title; in case of inner it is then checked whether it has correct length. In all title arguments, the following patterns are substituted:

"%C"

class of argument object

"%A"

deparsed argument object

"%D"

time/date-string when the plot was generated

Examples


IC1 <- new("IC")
plot(IC1)
plot(IC1, main = TRUE, panel.first= grid(),
     col = "blue", cex.main = 2, cex.inner = 1)

[Package RobAStBase version 0.1.3 Index]