radiusMinimaxIC(ROptEstOld)	R Documentation

Generic function for the computation of the radius minimax IC

Description

Generic function for the computation of the radius minimax IC.

Usage


radiusMinimaxIC(L2Fam, neighbor, risk, ...)

## S4 method for signature 'L2ParamFamily,
##   UncondNeighborhood, asGRisk':
radiusMinimaxIC(L2Fam, neighbor, risk, 
        loRad, upRad, z.start = NULL, A.start = NULL, upper = 1e5, 
        maxiter = 100, tol = .Machine$double.eps^0.4, warn = FALSE)

Arguments

L2Fam 	 L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

risk 	 object of class "RiskType".

... 	 additional parameters.

loRad 	 the lower end point of the interval to be searched.

upRad 	 the upper end point of the interval to be searched.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

Value

The radius minimax IC is computed.

Methods

L2Fam = "L2ParamFamily", neighbor = "UncondNeighborhood", risk = "asGRisk":

computation of the radius minimax IC for an L2 differentiable parametric family.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H., Kohl, M. and Ruckdeschel, P. (2001) The Costs of not Knowing the Radius. Submitted. Appeared as discussion paper Nr. 81. SFB 373 (Quantification and Simulation of Economic Processes), Humboldt University, Berlin; also available under www.uni-bayreuth.de/departments/math/org/mathe7/RIEDER/pubs/RR.pdf

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

radiusMinimaxIC

Examples


N <- NormLocationFamily(mean=0, sd=1) 
radiusMinimaxIC(L2Fam=N, neighbor=ContNeighborhood(), 
                risk=asMSE(), loRad=0.1, upRad=0.5)

[Package ROptEstOld version 0.5.2 Index]