rgsOptIC.ALc(RobRex)	R Documentation

Computation of the optimally robust IC for AL estimators

Description

The function rgsOptIC.ALc computes the optimally robust conditionally centered IC for AL estimators in case of linear regression with unknown scale and average conditional (convex) contamination neighborhoods where the regressor is random; confer Subsubsection 7.2.1.2 of Kohl (2005).

Usage


rgsOptIC.ALc(r, K, theta, scale = 1, A.rg.start, a.sc.start, A.sc.start = 0.5, 
             bUp = 1000, delta = 1e-06, itmax = 50, check = FALSE)

Arguments

r 	 non-negative real: neighborhood radius.

K 	 object of class "DiscreteDistribution" or object of class "DisreteMVDistribution".

theta 	 specified regression parameter.

scale 	 specified error scale.

A.rg.start 	 positive definite and symmetric matrix: starting value for the standardizing matrix of the regression part.

a.sc.start 	 real vector: starting values for centering function of the scale part.

A.sc.start 	 positive real: starting value for the standardizing constant of the scale part.

bUp 	 positive real: the upper end point of the interval to be searched for b.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

check 	 logical. Should constraints be checked.

Details

If theta is missing, it is set to 0. If A.rg.start is missing, the inverse of the second moment matrix of K is used. In case a.sc.start is missing, it is set to a null vector with length of the support of K.

Value

Object of class "Av1CondContIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

Av1CondContIC-class

Examples


K <- DiscreteDistribution(1:5) # = Unif({1,2,3,4,5})
IC1 <- rgsOptIC.ALc(r = 0.1, K = K)
checkIC(IC1)
Risks(IC1)

[Package RobRex version 0.7 Index]