rgsOptIC.ALs(RobRex)	R Documentation

Computation of the optimally robust IC for ALs estimators

Description

The function rgsOptIC.ALs computes the optimally robust IC for ALs estimators in case of linear regression with unknown scale and (convex) contamination neighborhoods where the regressor is random; confer Subsection 7.3.1 of Kohl (2005).

Usage


rgsOptIC.ALs(r, K, A.rg.start, b.rg.Up = 1000, delta = 1e-06, 
             itmax = 50, check = FALSE)

Arguments

r 	 non-negative real: neighborhood radius.

K 	 object of class "Distribution".

A.rg.start 	 positive definite and symmetric matrix: starting value for the standardizing matrix of the regression part.

b.rg.Up 	 positive real: the upper end point of the interval to be searched for b.rg.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

check 	 logical. Should constraints be checked.

Details

If A.rg.start is missing, the inverse of the second moment matrix of K is used.

Value

Object of class "ContIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class

Examples


K <- DiscreteDistribution(1:5) # = Unif({1,2,3,4,5})
IC1 <- rgsOptIC.ALs(r = 0.1, K = K)
checkIC(IC1)
Risks(IC1)
Infos(IC1)

[Package RobRex version 0.6.1 Index]