rgsOptIC.BM(RobRex)	R Documentation

Computation of the optimally robust IC for BM estimators

Description

The function rgsOptIC.BM computes the optimally robust IC for BM estimators in case of linear regression with unknown scale and (convex) contamination neighborhoods where the regressor is random. These estimators were proposed by Bednarski and Mueller (2001); confer also Subsection 7.3.3 of Kohl (2005).

Usage


rgsOptIC.BM(r, K, b.rg.start = 2.5, b.sc.0.x.start, delta = 1e-06, 
            MAX = 100, itmax = 1000)

Arguments

r 	 non-negative real: neighborhood radius.

K 	 object of class "DiscreteDistribution"

b.rg.start 	 positive real: starting value for b_rg.

b.sc.0.x.start 	 positive real: starting value for b_sc,0,x.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

MAX 	 if b_loc or b_sc,0 are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for BM estimators is based on optim where MAX is used to control the constraints on b_rg and b_sc,0,x.

Value

Object of class "CondIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Bednarski, T and Mueller, C.H. (2001) Optimal bounded influence regression and scale M-estimators in the context of experimental design. Statistics, 35(4): 349–369.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class

Examples


## code takes some time
## Not run: 
K <- DiscreteDistribution(1:5) # = Unif({1,2,3,4,5})
IC1 <- rgsOptIC.BM(r = 0.1, K = K)
checkIC(IC1)
Risks(IC1)

## End(Not run)

[Package RobRex version 0.6.1 Index]