rgsOptIC.MK(RobRex)	R Documentation

Computation of the optimally robust IC for MK estimators

Description

The function rgsOptIC.MK computes the optimally robust IC for MK estimators in case of linear regression with unknown scale and (convex) contamination neighborhoods where the regressor is random; confer Subsubsection 7.2.2.1 of Kohl (2005).

Usage


rgsOptIC.MK(r, K, ggLo = 0.5, ggUp = 1, a1.start = -0.25, a3.start = 0.25, 
             B.start, bUp = 1000, delta = 1e-06, itmax = 1000, check = FALSE)

Arguments

r 	 non-negative real: neighborhood radius.

K 	 object of class "Distribution".

ggLo 	 positive real: the lower end point of the interval to be searched for gamma.

ggUp 	 positive real: the upper end point of the interval to be searched for gamma.

a1.start 	 real: starting value for Lagrange multiplier alpha_1.

a3.start 	 real: starting value for Lagrange multiplier alpha_3.

B.start 	 symmetric matrix: starting value for Lagrange multiplier B.

bUp 	 positive real: the upper end point of the interval to be searched for b.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

check 	 logical. Should constraints be checked.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


## code takes some time
## Not run: 
K <- DiscreteDistribution(1:5) # = Unif({1,2,3,4,5})
IC1 <- rgsOptIC.MK(r = 0.1, K = K)
checkIC(IC1)
Risks(IC1)

## End(Not run)

[Package RobRex version 0.6.1 Index]