rgsOptIC.Ms(RobRex)	R Documentation

Computation of the optimally robust IC for Ms estimators

Description

The function rgsOptIC.Ms computes the optimally robust conditionally centered IC for Ms estimators in case of linear regression with unknown scale and average conditional (convex) contamination neighborhoods where the regressor is random; confer Subsection 7.3.2 of Kohl (2005).

Usage


rgsOptIC.Ms(r, K, a1.x.start, a3.start = 0.25, b.sc.start = 1.5, 
             bUp = 1000, ggLo = 0.5, ggUp = 1, delta = 1e-06, 
             itmax = 1000, check = FALSE)

Arguments

r 	 non-negative real: neighborhood radius.

K 	 object of class "DiscreteDistribution"

ggLo 	 positive real: the lower end point of the interval to be searched for gamma.

ggUp 	 positive real: the upper end point of the interval to be searched for gamma.

a1.x.start 	 real: starting value for the Lagrange multiplier function alpha_1(x).

a3.start 	 real: starting value for Lagrange multiplier alpha_3.

b.sc.start 	 positive real: starting value for the clipping bound b_sc.

bUp 	 positive real: the upper end point of the interval to be searched for b.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

check 	 logical. Should constraints be checked.

Value

Object of class "CondIC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

CondIC-class

Examples


## code takes some time
## Not run: 
K <- DiscreteDistribution(1:5) # = Unif({1,2,3,4,5})
IC1 <- rgsOptIC.Ms(r = 0.1, K = K)
checkIC(IC1)
Risks(IC1)

## End(Not run)

[Package RobRex version 0.6.1 Index]