rlOptIC(RobLox)	R Documentation

Computation of the optimally robust IC for AL estimators

Description

The function rlOptIC computes the optimally robust IC for AL estimators in case of normal location and (convex) contamination neighborhoods. The definition of these estimators can be found in Rieder (1994) or Kohl (2005), respectively.

Usage


rlOptIC(r, mean = 0, sd = 1, bUp = 1000, computeIC = TRUE)

Arguments

r 	 non-negative real: neighborhood radius.

mean 	 specified mean.

sd 	 specified standard deviation.

bUp 	 positive real: the upper end point of the interval to be searched for the clipping bound b.

computeIC 	 logical: should IC be computed. See details below.

Details

If 'computeIC' is 'FALSE' only the Lagrange multipliers 'A', 'a', and 'b' contained in the optimally robust IC are computed.

Value

If 'computeIC' is 'TRUE' an object of class "ContIC" is returned, otherwise a list of Lagrane multipliers

A 	 standardizing constant

a 	 centering constant; always '= 0' is this symmetric setup

b 	 optimal clipping bound

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, roblox

Examples


IC1 <- rlOptIC(r = 0.1)
distrExOptions("ErelativeTolerance" = 1e-12)
checkIC(IC1)
distrExOptions("ErelativeTolerance" = .Machine$double.eps^0.25) # default
Risks(IC1)
cent(IC1)
clip(IC1)
stand(IC1)
plot(IC1)

[Package RobLox version 0.6.0 Index]