rlsOptIC.BM(RobLox)	R Documentation

Computation of the optimally robust IC for BM estimators

Description

The function rlsOptIC.BM computes the optimally robust IC for BM estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. These estimators were proposed by Bednarski and Mueller (2001). A definition of these estimators can also be found in Section 8.4 of Kohl (2005).

Usage


rlsOptIC.BM(r, bL.start = 2, bS.start = 1.5, delta = 1e-06, MAX = 100)

Arguments

r 	 non-negative real: neighborhood radius.

bL.start 	 positive real: starting value for b_loc.

bS.start 	 positive real: starting value for b_sc,0.

delta 	 the desired accuracy (convergence tolerance).

MAX 	 if b_loc or b_sc,0 are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for BM estimators is based on optim where MAX is used to control the constraints on b_loc and b_sc,0. The optimal values of the tuning constants b_loc, b_sc,0, alpha and gamma can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Bednarski, T and Mueller, C.H. (2001) Optimal bounded influence regression and scale M-estimators in the context of experimental design. Statistics, 35(4): 349–369.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.BM(r = 0.1)
checkIC(IC1)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.6.0 Index]