rlsOptIC.HaMad(RobLox)	R Documentation

Computation of the optimally robust IC for HuMad estimators

Description

The function rlsOptIC.HuMad computes the optimally robust IC for HuMad estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. These estimators were considered in Andrews et al. (1972). A definition of these estimators can also be found in Subsection 8.5.2 of Kohl (2005).

Usage


rlsOptIC.HaMad(r, a.start = 0.25, b.start = 2.5, c.start = 5, 
               delta = 1e-06, MAX = 100)

Arguments

r 	 non-negative real: neighborhood radius.

a.start 	 positive real: starting value for a.

b.start 	 positive real: starting value for b.

c.start 	 positive real: starting value for c.

delta 	 the desired accuracy (convergence tolerance).

MAX 	 if a or b or c are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for HaMad estimators is based on optim where MAX is used to control the constraints on a, b and c. The optimal values of the tuning constants a, b, and c can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Andrews, D.F., Bickel, P.J., Hampel, F.R., Huber, P.J., Rogers, W.H. and Tukey, J.W. (1972) Robust estimates of location. Princeton University Press.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.HaMad(r = 0.1)
checkIC(IC1)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.6.0 Index]