rlsOptIC.Hu2a(RobLox)	R Documentation

Computation of the optimally robust IC for Hu2a estimators

Description

The function rlsOptIC.Hu2a computes the optimally robust IC for Hu2a estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. These estimators are a simple modification of Huber (1964), Proposal 2 where we, in addition, admit a clipping from below. The definition of these estimators can be found in Subsection 8.5.1 of Kohl (2005).

Usage


rlsOptIC.Hu2a(r, k1.start = 0.25, k2.start = 2.5, delta = 1e-06, MAX = 100)

Arguments

r 	 non-negative real: neighborhood radius.

k1.start 	 positive real: starting value for k1.

k2.start 	 positive real: starting value for k2.

delta 	 the desired accuracy (convergence tolerance).

MAX 	 if k1 or k2 are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for Hu2a estimators is based on optim where MAX is used to control the constraints on k1 and k2. The optimal values of the tuning constants k1 and k2 can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1964) Robust estimation of a location parameter. Ann. Math. Stat. 35: 73–101.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.Hu2a(r = 0.1)
checkIC(IC1)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.6.0 Index]