rlsOptIC.Hu3(RobLox)	R Documentation

Computation of the optimally robust IC for Hu3 estimators

Description

The function rlsOptIC.Hu3 computes the optimally robust IC for Hu3 estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. The definition of these estimators can be found in Subsection 8.5.1 of Kohl (2005).

Usage


rlsOptIC.Hu3(r, k.start = 1, c1.start = 0.1, c2.start = 0.5, 
             delta = 1e-06, MAX = 100)

Arguments

r 	 non-negative real: neighborhood radius.

k.start 	 positive real: starting value for k.

c1.start 	 positive real: starting value for c1.

c2.start 	 positive real: starting value for c2.

delta 	 the desired accuracy (convergence tolerance).

MAX 	 if k or c1 or c2 are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for Hu2 estimators is based on optim where MAX is used to control the constraints on k, c1 and c2. The optimal values of the tuning constants k, c1 and c2 can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.Hu3(r = 0.1)
checkIC(IC1)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.7 Index]