rlsOptIC.M(RobLox)	R Documentation

Computation of the optimally robust IC for M estimators

Description

The function rlsOptIC.M computes the optimally robust IC for M estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. The definition of these estimators can be found in Section 8.3 of Kohl (2005).

Usage


rlsOptIC.M(r, ggLo = 0.5, ggUp = 1.5, a1.start = 0.75, a3.start = 0.25, 
           bUp = 1000, delta = 1e-05, itmax = 100, check = FALSE)

Arguments

r 	 non-negative real: neighborhood radius.

ggLo 	 non-negative real: the lower end point of the interval to be searched for gamma.

ggUp 	 positive real: the upper end point of the interval to be searched for gamma.

a1.start 	 real: starting value for alpha_1.

a3.start 	 real: starting value for alpha_3.

bUp 	 positive real: upper bound used in the computation of the optimal clipping bound b.

delta 	 the desired accuracy (convergence tolerance).

itmax 	 the maximum number of iterations.

check 	 logical. Should constraints be checked.

Details

The optimal values of the tuning constants alpha_1, alpha_3, b and gamma can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1981) Robust Statistics. New York: Wiley.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.M(r = 0.1, check = TRUE)
distrExOptions("ErelativeTolerance" = 1e-12)
checkIC(IC1, NormLocationScaleFamily())
distrExOptions("ErelativeTolerance" = .Machine$double.eps^0.25)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.6.0 Index]