rlsOptIC.MM2(RobLox)	R Documentation

Computation of the optimally robust IC for MM2 estimators

Description

The function rlsOptIC.MM2 computes the optimally robust IC for MM2 estimators in case of normal location with unknown scale and (convex) contamination neighborhoods. These estimators are based on a proposal of Fraiman et al. (2001), p. 206. A definition of these estimators can also be found in Section 8.6 of Kohl (2005).

Usage


rlsOptIC.MM2(r, c.start = 1.5, d.start = 2, delta = 1e-06, MAX = 100)

Arguments

r 	 non-negative real: neighborhood radius.

c.start 	 positive real: starting value for c.

d.start 	 positive real: starting value for d.

delta 	 the desired accuracy (convergence tolerance).

MAX 	 if a or k are beyond the admitted values, MAX is returned.

Details

The computation of the optimally robust IC for MM2 estimators is based on optim where MAX is used to control the constraints on c and d. The optimal values of the tuning constants c and d can be read off from the slot Infos of the resulting IC.

Value

Object of class "IC"

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Fraiman, R., Yohai, V.J. and Zamar, R.H. (2001) Optimal robust M-estimates of location. Ann. Stat. 29(1): 194–223.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

IC-class

Examples


IC1 <- rlsOptIC.MM2(r = 0.1)
checkIC(IC1)
Risks(IC1)
Infos(IC1)
plot(IC1)
infoPlot(IC1)

[Package RobLox version 0.6.0 Index]