symmetricBiasType-class(ROptEst)	R Documentation

symmetric Bias Type

Description

Class of symmetric bias types.

Objects from the Class

Objects can be created by calls of the form new("symmetricBiasType", ...). More frequently they are created via the generating function symmetricBias.

Slots

name:

Object of class "character".

Methods

No methods defined with class "symmetricBiasType" in the signature.

Extends

Class "BiasType", directly.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

BiasType-class

Examples


symmetricBias()
## The function is currently defined as
function(){ new("symmetricBiasType", name = "symmetric Bias") }

[Package ROptEst version 0.5.0 Index]