trAsCov-class(ROptEst)	R Documentation

Trace of asymptotic covariance

Description

Class of trace of asymptotic covariance.

Objects from the Class

Objects can be created by calls of the form new("trAsCov", ...). More frequently they are created via the generating function trAsCov.

Slots

type:

Object of class "character": “trace of asymptotic covariance”.

Extends

Class "asRisk", directly.
Class "RiskType", by class "asRisk".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class, trAsCov

Examples


new("trAsCov")

[Package ROptEst version 0.6.0 Index]