trFiCov-class(ROptEst)	R Documentation

Trace of finite-sample covariance

Description

Class of trace of finite-sample covariance.

Objects from the Class

Objects can be created by calls of the form new("trFiCov", ...). More frequently they are created via the generating function trFiCov.

Slots

type:

Object of class "character": “trace of finite-sample covariance”.

Extends

Class "fiRisk", directly.
Class "RiskType", by class "fiRisk".

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Ruckdeschel, P. and Kohl, M. (2005) How to approximate the finite sample risk of M-estimators.

See Also

fiRisk-class, trFiCov

Examples


new("trFiCov")

[Package ROptEst version 0.5.0 Index]