updateNorm-methods(ROptEst)	R Documentation

Methods for Function updateNorm in Package ‘ROptEst’

Description

updateNorm-methods to update norm in IC-Algo

Usage

updateNorm(normtype, ...)
## S4 method for signature 'SelfNorm':
updateNorm(normtype, L2, neighbor, biastype, Distr, V.comp, 
                                cent, stand,  w)     

Arguments

normtype 	 normtype of class NormType

... 	 further arguments to be passed to specific methods.

L2 	 L2derivative

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType"

cent 	 optimal centering constant.

stand 	 standardizing matrix.

Distr 	 standardizing matrix.

V.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

w 	 object of class RobWeight; current weight

Details

updateNorm is used internally in the opt-IC-algorithm to be able to work with a norm that depends on the current covariance (SelfNorm)

Value

updateNorm 	 an updated object of class NormType.

Methods

updateNorm

signature(normtype = "SelfNorm"): udates the norm in the self-standardized case; just used internally in the opt-IC-Algorithm.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

See Also

NormType-class

[Package ROptEst version 0.6.3 Index]