getAsRisk(ROptEst)	R Documentation

Generic Function for Computation of Asymptotic Risks

Description

Generic function for the computation of asymptotic risks. This function is rarely called directly. It is used by other functions.

Usage


getAsRisk(risk, L2deriv, neighbor, biastype, ...)

## S4 method for signature 'asMSE, UnivariateDistribution,
##   Neighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                      neighbor, biastype, clip, cent, stand, trafo)

## S4 method for signature 'asMSE, EuclRandVariable,
##   Neighborhood, ANY':
getAsRisk(risk, L2deriv, neighbor, 
                                                    biastype, clip, cent, stand, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   ContNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                              neighbor, biastype, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   ContNeighborhood, onesidedBias':
getAsRisk(risk, 
                                                      L2deriv, neighbor, biastype, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   ContNeighborhood, asymmetricBias':
getAsRisk(risk, 
                                                        L2deriv, neighbor, biastype, trafo)

## S4 method for signature 'asBias, UnivariateDistribution,
##   TotalVarNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                                  neighbor, biastype, trafo)

## S4 method for signature 'asBias, RealRandVariable,
##   ContNeighborhood, ANY':
getAsRisk(risk, L2deriv, neighbor, 
                              biastype, Distr, DistrSymm, L2derivSymm, L2derivDistrSymm, 
                              trafo, z.start, A.start, maxiter, tol, warn)

## S4 method for signature 'asCov, UnivariateDistribution,
##   ContNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                  neighbor, biastype, clip, cent, stand)

## S4 method for signature 'asCov, UnivariateDistribution,
##   TotalVarNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                    neighbor, biastype, clip, cent, stand)

## S4 method for signature 'asCov, RealRandVariable,
##   ContNeighborhood, ANY':
getAsRisk(risk, L2deriv, neighbor, 
                           biastype, Distr, cent, stand, 
                           V.comp =  matrix(TRUE, ncol = nrow(stand), nrow = nrow(stand)), w)

## S4 method for signature 'trAsCov,
##   UnivariateDistribution, UncondNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                      neighbor, biastype, clip, cent, stand)

## S4 method for signature 'trAsCov, RealRandVariable,
##   ContNeighborhood, ANY':
getAsRisk(risk, L2deriv, neighbor, 
                                                biastype, Distr, clip, cent, stand, normtype)

## S4 method for signature 'asUnOvShoot,
##   UnivariateDistribution, UncondNeighborhood, ANY':
getAsRisk(risk, L2deriv, 
                                                  neighbor, biastype, clip, cent, stand, trafo)

## S4 method for signature 'asSemivar,
##   UnivariateDistribution, Neighborhood, onesidedBias':
getAsRisk(risk, L2deriv, 
                                                    neighbor, biastype, clip, cent, stand, trafo)

Arguments

risk 	 object of class "asRisk".

L2deriv 	 L2-derivative of some L2-differentiable family of probability distributions.

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "ANY".

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

stand 	 standardizing matrix.

trafo 	 matrix: transformation of the parameter.

Distr 	 object of class "Distribution".

DistrSymm 	 object of class "DistributionSymmetry".

L2derivSymm 	 object of class "FunSymmList".

L2derivDistrSymm 	 object of class "DistrSymmList".

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

maxiter 	 the maximum number of iterations

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

normtype 	 object of class "NormType".

V.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

w 	 object of class RobWeight; current weight

Details

This function is rarely called directly. It is used by other functions/methods.

Value

The asymptotic risk is computed.

Methods

risk = "asMSE", L2deriv = "UnivariateDistribution", neighbor = "Neighborhood", biastype = "ANY":

computes asymptotic mean square error in methods for function getInfRobIC.

risk = "asMSE", L2deriv = "EuclRandVariable", neighbor = "Neighborhood", biastype = "ANY":

computes asymptotic mean square error in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "ANY":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "onesidedBias":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "asymmetricBias":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood", biastype = "ANY":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asBias", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "ANY":

computes standardized asymptotic bias in methods for function getInfRobIC.

risk = "asCov", L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "ANY":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "asCov", L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood", biastype = "ANY":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "asCov", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "ANY":

computes asymptotic covariance in methods for function getInfRobIC.

risk = "trAsCov", L2deriv = "UnivariateDistribution", neighbor = "UncondNeighborhood", biastype = "ANY":

computes trace of asymptotic covariance in methods for function getInfRobIC.

risk = "trAsCov", L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "ANY":

computes trace of asymptotic covariance in methods for function getInfRobIC.

risk = "asUnOvShoot", L2deriv = "UnivariateDistribution", neighbor = "UncondNeighborhood", biastype = "ANY":

computes asymptotic under-/overshoot risk in methods for function getInfRobIC.

risk = "asSemivar", L2deriv = "UnivariateDistribution", neighbor = "Neighborhood", biastype = "onesidedBias":

computes asymptotic semivariance in methods for function getInfRobIC.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions 22, 201-223.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asRisk-class

[Package ROptEst version 0.7 Index]