getBiasIC(RobAStBase)	R Documentation

Generic function for the computation of the asymptotic bias for an IC

Description

Generic function for the computation of the asymptotic bias for an IC.

Usage


getBiasIC(IC, neighbor, ...)

## S4 method for signature 'IC, UncondNeighborhood':
getBiasIC(IC, neighbor, L2Fam, biastype = symmetricBias(),
             normtype = NormType(), tol = .Machine$double.eps^0.25, numbeval = 1e5)

Arguments

IC 	 object of class "InfluenceCurve"

neighbor 	 object of class "Neighborhood".

... 	 additional parameters

L2Fam 	 object of class "L2ParamFamily".

biastype 	 object of class "BiasType"

normtype 	 object of class "NormType"

tol 	 the desired accuracy (convergence tolerance).

numbeval 	 number of evalation points.

Details

Value

The bias of the IC is computed.

Methods

IC = "IC", neighbor = "UncondNeighborhood"

determines the as. bias by random evaluation of the IC; this random evaluation is done by the internal S4-method .evalBiasIC; this latter dispatches according to the signature IC, neighbor, biastype.
For signature IC="IC", neighbor = "ContNeighborhood", biastype = "BiasType", also an argument normtype is used to be able to use self- or information standardizing norms; besides this the signatures IC="IC", neighbor = "TotalVarNeighborhood", biastype = "BiasType", IC="IC", neighbor = "ContNeighborhood", biastype = "onesidedBias", and IC="IC", neighbor = "ContNeighborhood", biastype = "asymmetricBias" are implemented.

Note

This generic function is still under construction.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

Ruckdeschel, P. and Kohl, M. (2005) Computation of the Finite Sample Bias of M-estimators on Neighborhoods.

See Also

getRiskIC-methods, InfRobModel-class

[Package RobAStBase version 0.1.0 Index]