getBiasIC(ROptEst)	R Documentation

Generic function for the computation of the asymptotic bias for an IC

Description

Generic function for the computation of the asymptotic bias for an IC.

Usage


getBiasIC(IC, neighbor, ...)

## S4 method for signature 'HampIC, UncondNeighborhood':
getBiasIC(IC, neighbor, L2Fam, ...)

Arguments

IC 	 object of class "InfluenceCurve"

neighbor 	 object of class "Neighborhood".

L2Fam 	 object of class "L2ParamFamily".

... 	 additional parameters

Details

This function is rarely called directly. It is used by other functions/methods.

Value

The bias of the IC is computed.

Methods

IC = "HampIC", neighbor = "UncondNeighborhood"

reads off the as. bias from the risks-slot of the IC.

IC = "TotalVarIC", neighbor = "UncondNeighborhood"

reads off the as. bias from the risks-slot of the IC, resp. if this is NULL from the corresponding Lagrange Multipliers.

Note

This generic function is still under construction.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

Ruckdeschel, P. and Kohl, M. (2005) Computation of the Finite Sample Bias of M-estimators on Neighborhoods.

See Also

getRiskIC-methods, InfRobModel-class

[Package ROptEst version 0.7 Index]