getInfCent(ROptEst)	R Documentation

Generic Function for the Computation of the Optimal Centering Constant/Lower Clipping Bound

Description

Generic function for the computation of the optimal centering constant (contamination neighborhoods) respectively, of the optimal lower clipping bound (total variation neighborhood). This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfCent(L2deriv, neighbor, biastype, ...)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, BiasType':
getInfCent(L2deriv, 
     neighbor, biastype, clip, cent, tol.z, symm, trafo)

## S4 method for signature 'UnivariateDistribution,
##   TotalVarNeighborhood, BiasType':
getInfCent(L2deriv, 
     neighbor, biastype, clip, cent, tol.z, symm, trafo)

## S4 method for signature 'RealRandVariable,
##   ContNeighborhood, BiasType':
getInfCent(L2deriv, 
     neighbor, biastype, Distr, z.comp, w)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, onesidedBias':
getInfCent(L2deriv, 
     neighbor, biastype, clip, cent, tol.z, symm, trafo)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, asymmetricBias':
getInfCent(L2deriv, 
     neighbor, biastype, clip, cent, tol.z, symm, trafo)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType"

... 	 additional parameters.

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

tol.z 	 the desired accuracy (convergence tolerance).

symm 	 logical: indicating symmetry of L2deriv.

trafo 	 matrix: transformation of the parameter.

Distr 	 object of class Distribution.

z.comp 	 logical vector: indication which components of the centering constant have to be computed.

w 	 object of class RobWeight; current weight

Value

The optimal centering constant is computed.

Methods

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "BiasType"

computation of optimal centering constant for symmetric bias.

L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood", biastype = "BiasType"

computation of optimal lower clipping bound for symmetric bias.

L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "BiasType"

computation of optimal centering constant for symmetric bias.

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "onesidedBias"

computation of optimal centering constant for onesided bias.

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "asymmetricBias"

computation of optimal centering constant for asymmetric bias.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.7 Index]