getInfGamma(ROptEst)	R Documentation

Generic Function for the Computation of the Optimal Clipping Bound

Description

Generic function for the computation of the optimal clipping bound. This function is rarely called directly. It is called by getInfClip to compute optimally robust ICs.

Usage


getInfGamma(L2deriv, risk, neighbor, ...)

## S4 method for signature 'UnivariateDistribution, asMSE,
##   ContNeighborhood':
getInfGamma(L2deriv, risk, neighbor, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   asGRisk, TotalVarNeighborhood':
getInfGamma(L2deriv, risk, neighbor, cent, clip)

## S4 method for signature 'RealRandVariable, asMSE,
##   ContNeighborhood':
getInfGamma(L2deriv, risk, neighbor, Distr, stand, cent, clip)

## S4 method for signature 'UnivariateDistribution,
##   asUnOvShoot, ContNeighborhood':
getInfGamma(L2deriv, risk, neighbor, cent, clip)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters

cent 	 optimal centering constant.

clip 	 optimal clipping bound.

stand 	 standardizing matrix.

Distr 	 object of class "Distribution".

Details

The function is used in case of asymptotic G-risks; confer Ruckdeschel and Rieder (2004).

Methods

L2deriv = "UnivariateDistribution", risk = "asMSE", neighbor = "ContNeighborhood"

used by getInfClip.

L2deriv = "UnivariateDistribution", risk = "asGRisk", neighbor = "TotalVarNeighborhood"

used by getInfClip.

L2deriv = "RealRandVariable", risk = "asMSE", neighbor = "ContNeighborhood"

used by getInfClip.

L2deriv = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "ContNeighborhood"

used by getInfClip.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions (submitted).

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

asGRisk-class, asMSE-class, asUnOvShoot-class, ContIC-class, TotalVarIC-class

[Package ROptEst version 0.5.0 Index]