getInfRobIC(ROptEst)	R Documentation

Generic Function for the Computation of Optimally Robust ICs

Description

Generic function for the computation of optimally robust ICs in case of infinitesimal robust models. This function is rarely called directly.

Usage


getInfRobIC(L2deriv, risk, neighbor, ...)

## S4 method for signature 'UnivariateDistribution,asCov,ContNeighborhood':
getInfRobIC(L2deriv,
                       risk, neighbor, Finfo, trafo,
                       verbose = getRobAStBaseOption("all.verbose"))

## S4 method for signature 'UnivariateDistribution,asCov,TotalVarNeighborhood':
getInfRobIC(L2deriv,
                       risk, neighbor, Finfo, trafo,
                       verbose = getRobAStBaseOption("all.verbose"))

## S4 method for signature 'RealRandVariable,asCov,UncondNeighborhood':
getInfRobIC(L2deriv, risk,
                       neighbor, Distr, Finfo, trafo,
                       QuadForm = diag(nrow(trafo)),
                       verbose = getRobAStBaseOption("all.verbose"))

## S4 method for signature 'UnivariateDistribution,asBias,UncondNeighborhood':
getInfRobIC(L2deriv,
                       risk, neighbor, symm, trafo,
                       maxiter, tol, warn, Finfo,
                       verbose = getRobAStBaseOption("all.verbose"), ...)

## S4 method for signature 'RealRandVariable,asBias,UncondNeighborhood':
getInfRobIC(L2deriv, risk,
                       neighbor, Distr, DistrSymm, L2derivSymm,
                       L2derivDistrSymm, z.start, A.start, Finfo, trafo,
                       maxiter, tol, warn,
                       verbose = getRobAStBaseOption("all.verbose"), ...)

## S4 method for signature 'UnivariateDistribution,asHampel,UncondNeighborhood':
getInfRobIC(L2deriv,
                       risk, neighbor, symm, Finfo, trafo, upper = NULL,
                       lower=NULL, maxiter, tol, warn, noLow = FALSE,
                       verbose = getRobAStBaseOption("all.verbose"),
                       checkBounds = TRUE)

## S4 method for signature 'RealRandVariable,asHampel,UncondNeighborhood':
getInfRobIC(L2deriv, risk,
                       neighbor, Distr, DistrSymm, L2derivSymm,
                       L2derivDistrSymm, Finfo, trafo, onesetLM = FALSE,
                       z.start, A.start, upper = NULL, lower=NULL,
                       OptOrIter = "iterate", maxiter, tol, warn,
                       verbose = getRobAStBaseOption("all.verbose"),
                       checkBounds = TRUE, ...)

## S4 method for signature 'UnivariateDistribution,asGRisk,UncondNeighborhood':
getInfRobIC(L2deriv,
                       risk, neighbor, symm, Finfo, trafo,
                       upper = NULL, lower = NULL,
                       maxiter, tol, warn, noLow = FALSE,
                       verbose = getRobAStBaseOption("all.verbose"))

## S4 method for signature 'RealRandVariable,asGRisk,UncondNeighborhood':
getInfRobIC(L2deriv, risk,
                       neighbor,  Distr, DistrSymm, L2derivSymm,
                       L2derivDistrSymm, Finfo, trafo, onesetLM = FALSE, z.start,
                       A.start, upper = NULL, lower = NULL, OptOrIter = "iterate",
                       maxiter, tol, warn,
                       verbose = getRobAStBaseOption("all.verbose"), ...)

## S4 method for signature 'UnivariateDistribution,asUnOvShoot,UncondNeighborhood':
getInfRobIC(
                       L2deriv, risk, neighbor, symm, Finfo, trafo,
                       upper, lower, maxiter, tol, warn)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

... 	 additional parameters (mainly for optim).

Distr 	 object of class "Distribution".

symm 	 logical: indicating symmetry of L2deriv.

DistrSymm 	 object of class "DistributionSymmetry".

L2derivSymm 	 object of class "FunSymmList".

L2derivDistrSymm 	 object of class "DistrSymmList".

Finfo 	 Fisher information matrix.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

trafo 	 matrix: transformation of the parameter.

upper 	 upper bound for the optimal clipping bound.

lower 	 lower bound for the optimal clipping bound.

OptOrIter 	 character; which method to be used for determining Lagrange multipliers A and a: if (partially) matched to "optimize" (and by default) getLagrangeMultByOptim is used; else if matched to "iterate", (or in case risk is of class "asGRisk" "doubleiterate") getLagrangeMultByIter is used — in case risk is of class "asGRisk" if matched to "iterate" with only one (inner) iteration, if matched to "doubleiterate" maximally Maxiter (inner) iterations.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

noLow 	 logical: is lower case to be computed?

onesetLM 	 logical: use one set of Lagrange multipliers?

QuadForm 	 matrix of (or which may coerced to) class PosSemDefSymmMatrix for use of different (standardizing) norm

verbose 	 logical: if TRUE, some messages are printed

checkBounds 	 logical: if TRUE, minimal and maximal clipping bound are computed to check if a valid bound was specified.

Value

The optimally robust IC is computed.

Methods

L2deriv = "UnivariateDistribution", risk = "asCov", neighbor = "ContNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "UnivariateDistribution", risk = "asCov", neighbor = "TotalVarNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asCov", neighbor = "UncondNeighborhood"

computes the classical optimal influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate; for total variation neighborhoods only is implemented for the case where there is a 1 x k transformation trafo matrix.

L2deriv = "UnivariateDistribution", risk = "asBias", neighbor = "UncondNeighborhood"

computes the bias optimal influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asBias", neighbor = "UncondNeighborhood"

computes the bias optimal influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate.

L2deriv = "UnivariateDistribution", risk = "asHampel", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asHampel", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate; for total variation neighborhoods only is implemented for the case where there is a 1 x k transformation trafo matrix.

L2deriv = "UnivariateDistribution", risk = "asGRisk", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown one-dimensional parameter.

L2deriv = "RealRandVariable", risk = "asGRisk", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for L2 differentiable parametric families with unknown k-dimensional parameter (k > 1) where the underlying distribution is univariate; for total variation neighborhoods only is implemented for the case where there is a 1 x k transformation trafo matrix.

L2deriv = "UnivariateDistribution", risk = "asUnOvShoot", neighbor = "UncondNeighborhood"

computes the optimally robust influence curve for one-dimensional L2 differentiable parametric families and asymptotic under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106-115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. and Rieder, H. (2004) Optimal Influence Curves for General Loss Functions. Statistics & Decisions 22: 201-223.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfRobModel-class

[Package ROptEst version 0.7 Index]