getInfStand(ROptEst)	R Documentation

Generic Function for the Computation of the Standardizing Matrix

Description

Generic function for the computation of the standardizing matrix which takes care of the Fisher consistency of the corresponding IC. This function is rarely called directly. It is used to compute optimally robust ICs.

Usage


getInfStand(L2deriv, neighbor, biastype, ...)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, BiasType':
getInfStand(L2deriv, 
     neighbor, biastype, clip, cent, trafo)

## S4 method for signature 'UnivariateDistribution,
##   TotalVarNeighborhood, BiasType':
getInfStand(L2deriv, 
     neighbor, biastype, clip, cent, trafo)

## S4 method for signature 'RealRandVariable,
##   ContNeighborhood, BiasType':
getInfStand(L2deriv, 
     neighbor, biastype, Distr, A.comp, cent, trafo, w)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, BiasType':
getInfStand(L2deriv, 
     neighbor, biastype, clip, cent, trafo)

## S4 method for signature 'UnivariateDistribution,
##   ContNeighborhood, BiasType':
getInfStand(L2deriv, 
     neighbor, biastype, clip, cent, trafo)

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

neighbor 	 object of class "Neighborhood"

biastype 	 object of class "BiasType"

... 	 additional parameters

clip 	 optimal clipping bound.

cent 	 optimal centering constant.

Distr 	 object of class "Distribution".

trafo 	 matrix: transformation of the parameter.

A.comp 	 matrix: indication which components of the standardizing matrix have to be computed.

w 	 object of class RobWeight; current weight

Value

The standardizing matrix is computed.

Methods

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "BiasType"

computes standardizing matrix for symmetric bias.

L2deriv = "UnivariateDistribution", neighbor = "TotalVarNeighborhood", biastype = "BiasType"

computes standardizing matrix for symmetric bias.

L2deriv = "RealRandVariable", neighbor = "ContNeighborhood", biastype = "BiasType"

computes standardizing matrix for symmetric bias.

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "onesidedBias"

computes standardizing matrix for onesided bias.

L2deriv = "UnivariateDistribution", neighbor = "ContNeighborhood", biastype = "asymmetricBias"

computes standardizing matrix for asymmetric bias.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de, Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Ruckdeschel, P. (2005) Optimally One-Sided Bounded Influence Curves. Mathematical Methods in Statistics 14(1), 105-131.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

ContIC-class, TotalVarIC-class

[Package ROptEst version 0.7 Index]