infoPlot(ROptEstOld)	R Documentation

Plot absolute and relative information

Description

Plot absolute and relative information of influence curves.

Usage


infoPlot(object)

Arguments

object 	 object of class "InfluenceCurve"

Details

Absolute information is defined as the square of the length of an IC. The relative information is defined as the absolute information of one component with respect to the absolute information of the whole IC; confer Section 8.1 of Kohl (2005).

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

L2ParamFamily-class, IC-class

Examples


N <- NormLocationScaleFamily(mean=0, sd=1) 
IC1 <- optIC(model = N, risk = asCov())
infoPlot(IC1)

[Package ROptEstOld version 0.5.2 Index]