internals_for_ROptEst(ROptEst)	R Documentation

Internal / Helper functions of package ROptEst

Description

These functions are used internally by package ROptEst.

Usage


### helper function to check whether given b is in (bmin, bmax)
###        if not returns corresponding upper / lower case solution

.checkUpLow(L2deriv, b, risk, neighbor, biastype, normtype,
                        Distr, Finfo, DistrSymm, L2derivSymm,
                        L2derivDistrSymm, z.start, A.start, trafo, maxiter,
                        tol, QuadForm, verbose, nrvalpts, warn)
                        
### helper function to return the upper case solution if r=0
.getUpperSol(L2deriv, b, radius, risk, neighbor, biastype,
                       normtype, Distr, Finfo, trafo,
                       QuadForm, verbose, warn)

### helper function to return the lower case solution if b-search was not successful
.getLowerSol(L2deriv, risk, neighbor, Distr, DistrSymm,
                         L2derivSymm, L2derivDistrSymm,
                         z.start, A.start, trafo,
                         maxiter, tol, warn, Finfo, verbose)


### helper function to return upper & lower bounds for b for b-search
.getLowUpB(L2deriv, Finfo, Distr, normtype, z, A, radius, iter)

### helper function to check whether (TotalVariation) weight w has already been modified
.isVirginW(w)

### helper function to check whether (intermediate) results give a pIC
.checkPIC(L2deriv, neighbor, Distr, trafo, z, A, w, z.comp, A.comp, ...)

.LowerCaseMultivariate(L2deriv, neighbor, biastype,
             normtype, Distr, Finfo, trafo, z.start,
             A.start, z.comp, A.comp, maxiter, tol,
             verbose = getRobAStBaseOption("all.verbose"))

.LowerCaseMultivariateTV(L2deriv, neighbor, biastype,
             normtype, Distr, Finfo, trafo,
             A.start,  maxiter, tol,
             verbose = getRobAStBaseOption("all.verbose"))

Arguments

L2deriv 	 L2-derivative of some L2-differentiable family of probability measures.

b 	 numeric; clipping bound under consideration.

risk 	 object of class "RiskType".

neighbor 	 object of class "Neighborhood".

biastype 	 object of class "BiasType" — the bias type with we work.

normtype 	 object of class "NormType" — the norm type with we work.

Distr 	 object of class "Distribution".

Finfo 	 Fisher information matrix.

DistrSymm 	 object of class "DistributionSymmetry".

L2derivSymm 	 object of class "FunSymmList".

L2derivDistrSymm 	 object of class "DistrSymmList".

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

trafo 	 matrix: transformation of the parameter.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

QuadForm 	 matrix of (or which may coerced to) class PosSemDefSymmMatrix for use of different (standardizing) norm

verbose 	 logical: if TRUE, some messages are printed.

nrvalpts 	 integer: number of evaluation points.

warn 	 logical: print warnings.

radius 	 radius of the neighborhood.

z 	 centering constant (in k-space)

A 	 standardizing matrix.

w 	 a weight of class "BdStWeight"

z.comp 	 logical vector: indicator which components of z need to be computed

A.comp 	 logical matrix: indicator which components of A need to be computed

iter 	 the number of iterations computed so far; used for specifying a different value of the clipping component of the weight in total variation case in the very first iteration.

... 	 further arguments to be passed on E().

Details

.checkUpLow checks whether the given clipping height b lies in (b_min,b_max); .getUpperSol determines the upper case/classical solution and computes corresponding risks .getLowerSol determines the lower case (minimax bias) solution and computes corresponding risks .getLowUpB determines a search interval for b to given radius r, i.e., lower and upper bounds for (b_min,b_max) .isVirginW checks whether the (total variation) weight w in the argument has already been modified since creation (TRUE if not) .checkPIC checks whether (intermediate) results give a pIC .LowerCaseMultivariatefunction determines the Lagrange multipliers for the multivariate lower case solution for convex contamination by solving a corresponding dual problem (Rieder[94],p.199 eq.(18)). .LowerCaseMultivariatefunctionTV determines the Lagrange multipliers for the multivariate lower case solution for total variation in dimension p=1 and k>1 by solving a corresponding dual problem (Rieder[94],p.205 eq.(58)).

Value

.checkUpLow 	 a list with items up (logical) low (logical) and res a return list for getInfRobIC or NULL: if up is TRUE, res contains the upper case solution, if low is TRUE the lower case solution and if both low and up are FALSE NULL.

.getUpperSol 	 a return list for getInfRobIC

.getLowerSol 	 a return list for getInfRobIC

.checkUpLow 	 a list with items lower and upper (both numeric).

.isVirginW 	 TRUE or FALSE

.checkPIC 	 nothing is returned; precision values are issued.

.LowerCaseMultivariatefunction 	 a list with elements erg the return value of the corresponding call to optim, w the corresponding weight (of class HampelWeight), normtype the norm —possibly modified, z.comp the indices of non-null entries in z.

.LowerCaseMultivariatefunctionTV 	 a list with elements A the standardization matrix, b the minimax bias, w the corresponding weight (of class BdStWeight), a the corresponding infimum of the [p]IC.

Author(s)

Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

[Package ROptEst version 0.7 Index]