kStepEstimate-class(RobAStBase)	R Documentation

kStepEstimate-class.

Description

Class of asymptotically linear estimates.

Objects from the Class

Objects can be created by calls of the form new("kStepEstimate", ...). More frequently they are created via the generating function kStepEstimator.

Slots

name

Object of class "character": name of the estimator.

estimate

Object of class "ANY": estimate.

estimate.call

Object of class "call": call by which estimate was produced.

samplesize

object of class "numeric" — the samplesize (only complete cases are counted) at which the estimate was evaluated.

completecases:

object of class "logical" — complete cases at which the estimate was evaluated.

asvar

object of class "OptionalNumericOrMatrix" which may contain the asymptotic (co)variance of the estimator.

asbias

Optional object of class "numeric": asymptotic bias.

pIC

Optional object of class InfluenceCurve: influence curve.

nuis.idx

object of class "OptionalNumeric": indices of estimate belonging to the nuisance part.

fixed

object of class "OptionalNumeric": the fixed and known part of the parameter.

steps

Object of class "integer": number of steps.

Infos

object of class "matrix" with two columns named method and message: additional informations.

trafo

object of class "list": a list with components fct and mat (see below).

untransformed.estimate:

Object of class "ANY": untransformed estimate.

untransformed.asvar:

object of class "OptionalNumericOrMatrix" which may contain the asymptotic (co)variance of the untransformed estimator.

pICList

Optional object of class "OptionalpICList": the list of (intermediate) (partial) influence curves used; only filled when called from kStepEstimator with argument withPICList==TRUE.

ICList

Optional object of class "OptionalpICList": the list of (intermediate) (total) influence curves used; only filled when called from kStepEstimator with argument withICList==TRUE.

start

The argument start — of class "StartClass" used in call to kStepEstimator.

startval

Object of class matrix: the starting value with which the k-step Estimator was initialized (in p-space / transformed).

ustartval

Object of class matrix: the starting value with which the k-step Estimator was initialized (in k-space / untransformed).

ksteps

Object of class "OptionalMatrix": the intermediate estimates (in p-space) for the parameter; only filled when called from kStepEstimator.

uksteps

Object of class "OptionalMatrix": the intermediate estimates (in k-space) for the parameter; only filled when called from kStepEstimator.

Extends

Class "ALEstimate", directly.
Class "Estimate", by class "ALEstimate"

Methods

steps

signature(object = "kStepEstimate"): accessor function for slot steps.

ksteps

signature(object = "kStepEstimate"): accessor function for slot ksteps; has additional argument diff, defaulting to FALSE; if the latter is TRUE, the starting value from slot startval is prepended as first column; otherwise we return the corresponding increments in each step.

uksteps

signature(object = "kStepEstimate"): accessor function for slot uksteps; has additional argument diff, defaulting to FALSE; if the latter is TRUE, the starting value from slot ustartval is prepended as first column; otherwise we return the corresponding increments in each step.

start

signature(object = "kStepEstimate"): accessor function for slot start.

startval

signature(object = "kStepEstimate"): accessor function for slot startval.

ustartval

signature(object = "kStepEstimate"): accessor function for slot startval.

ICList

signature(object = "kStepEstimate"): accessor function for slot ICList.

pICList

signature(object = "kStepEstimate"): accessor function for slot pICList.

show

signature(object = "kStepEstimate"): a show method;

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

[Package RobAStBase version 0.7 Index]