kStepEstimate-class(RobAStBase)	R Documentation

kStepEstimate-class.

Description

Class of asymptotically linear estimates.

Objects from the Class

Objects can be created by calls of the form new("kStepEstimate", ...). More frequently they are created via the generating function kStepEstimator.

Slots

name

Object of class "character": name of the estimator.

estimate

Object of class "ANY": estimate.

estimate.call

Object of class "call": call by which estimate was produced.

samplesize

object of class "numeric" — the samplesize (only complete cases are counted) at which the estimate was evaluated.

completecases:

object of class "logical" — complete cases at which the estimate was evaluated.

asvar

object of class "OptionalNumericOrMatrix" which may contain the asymptotic (co)variance of the estimator.

asbias

Optional object of class "numeric": asymptotic bias.

pIC

Optional object of class InfluenceCurve: influence curve.

nuis.idx

object of class "OptionalNumeric": indices of estimate belonging to the nuisance part.

fixed

object of class "OptionalNumeric": the fixed and known part of the parameter.

steps

Object of class "integer": number of steps.

Infos

object of class "matrix" with two columns named method and message: additional informations.

trafo

object of class "list": a list with components fct and mat (see below).

untransformed.estimate:

Object of class "ANY": untransformed estimate.

untransformed.asvar:

object of class "OptionalNumericOrMatrix" which may contain the asymptotic (co)variance of the untransformed estimator.

Extends

Class "ALEstimate", directly.
Class "Estimate", by class "ALEstimate"

Methods

steps

signature(object = "kStepEstimate"): accessor function for slot steps.

show

signature(object = "kStepEstimate"): a show method;

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

See Also

[Package RobAStBase version 0.7 Index]