locMEstimator(RobAStBase)	R Documentation

Generic function for the computation of location M estimates

Description

Generic function for the computation of location M estimates.

Usage


locMEstimator(x, IC, ...)

## S4 method for signature 'numeric, InfluenceCurve':
locMEstimator(x, IC, eps = .Machine$double.eps^0.5)

Arguments

x 	 sample

IC 	 object of class "InfluenceCurve"

... 	 additional parameters

eps 	 the desired accuracy (convergence tolerance).

Details

Given some sample x and some influence curve IC an M estimate is computed by solving the corresponding M equation.

Value

Object of class "MEstimate"

Methods

x = "numeric", IC = "InfluenceCurve"

univariate location.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1964) Robust estimation of a location parameter. Ann. Math. Stat. 35: 73–101.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, MEstimate-class

[Package RobAStBase version 0.1.0 Index]