oneStepEstimator(RobAStBase)	R Documentation

Generic function for the computation of one-step estimates

Description

Generic function for the computation of one-step estimates.

Usage


oneStepEstimator(x, IC, start, ...)

## S4 method for signature 'numeric, InfluenceCurve, ANY':
oneStepEstimator(x, IC, start, 
    useLast = getRobAStBaseOption("kStepUseLast"))
## S4 method for signature 'matrix, InfluenceCurve, ANY':
oneStepEstimator(x, IC, start, 
    useLast = getRobAStBaseOption("kStepUseLast"))

Arguments

x 	 sample

IC 	 object of class "InfluenceCurve"

start 	 initial estimate

useLast 	 which parameter estimate (initial estimate or one-step estimate) shall be used to fill the slots pIC, asvar and asbias of the return value.

... 	 additional arguments

Details

Given an initial estimation start, a sample x and an influence curve IC the corresponding one-step estimator is computed.

In case IC is an object of class "IC" the slots asvar and asbias of the return value are filled (based on the initial estimate).

The default value of argument useLast is set by the global option kStepUseLast which by default is set to FALSE. In case of general models useLast remains unchanged during the computations. However, if slot CallL2Fam of IC generates an object of class "L2GroupParamFamily" the value of useLast is changed to TRUE. Explicitly setting useLast to TRUE should be done with care as in this situation the influence curve is re-computed using the value of the one-step estimate which may take quite a long time depending on the model.

If useLast is set to TRUE and slot modifyIC of IC is filled with some function (which can be used to re-compute the IC for a different parameter), the computation of asvar, asbias and IC is based on the one-step estimate.

Value

Object of class "kStepEstimate"

Methods

x = "numeric", IC = "InfluenceCurve", start = "ANY"

univariate samples.

x = "matrix", IC = "InfluenceCurve", start = "ANY"

multivariate samples.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, kStepEstimate-class

[Package RobAStBase version 0.7 Index]