optIC(ROptEst)	R Documentation

Generic function for the computation of optimally robust ICs

Description

Generic function for the computation of optimally robust ICs.

Usage


optIC(model, risk, ...)

## S4 method for signature 'InfRobModel, asRisk':
optIC(model, risk, z.start = NULL, A.start = NULL,
                                     upper = 1e4, maxiter = 50, 
                                     tol = .Machine$double.eps^0.4, warn = TRUE, 
                                     noLow = FALSE, verbose = FALSE, ...)

## S4 method for signature 'InfRobModel, asUnOvShoot':
optIC(model, risk, upper = 1e4, maxiter = 50, 
                                          tol = .Machine$double.eps^0.4, warn = TRUE)

## S4 method for signature 'FixRobModel, fiUnOvShoot':
optIC(model, risk, sampleSize, upper = 1e4, 
                                          maxiter = 50, tol = .Machine$double.eps^0.4, 
                                          warn = TRUE, Algo = "A", cont = "left",
                                          verbose = FALSE)

Arguments

model 	 probability model.

risk 	 object of class "RiskType".

... 	 additional parameters.

z.start 	 initial value for the centering constant.

A.start 	 initial value for the standardizing matrix.

upper 	 upper bound for the optimal clipping bound.

maxiter 	 the maximum number of iterations.

tol 	 the desired accuracy (convergence tolerance).

warn 	 logical: print warnings.

sampleSize 	 integer: sample size.

Algo 	 "A" or "B".

cont 	 "left" or "right".

noLow 	 logical: is lower case to be computed?

verbose 	 logical: if TRUE, some messages are printed

Details

In case of the finite-sample risk "fiUnOvShoot" one can choose between two algorithms for the computation of this risk where the least favorable contamination is assumed to be left or right of some bound. For more details we refer to Section 11.3 of Kohl (2005).

Value

Some optimally robust IC is computed.

Methods

model = "InfRobModel", risk = "asRisk"

computes optimally robust influence curve for robust models with infinitesimal neighborhoods and various asymptotic risks.

model = "InfRobModel", risk = "asUnOvShoot"

computes optimally robust influence curve for robust models with infinitesimal neighborhoods and asymptotic under-/overshoot risk.

model = "FixRobModel", risk = "fiUnOvShoot"

computes optimally robust influence curve for robust models with fixed neighborhoods and finite-sample under-/overshoot risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Huber, P.J. (1968) Robust Confidence Limits. Z. Wahrscheinlichkeitstheor. Verw. Geb. 10:269–278.

Rieder, H. (1980) Estimates derived from robust tests. Ann. Stats. 8: 106–115.

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

Kohl, M. (2005) Numerical Contributions to the Asymptotic Theory of Robustness. Bayreuth: Dissertation.

See Also

InfluenceCurve-class, RiskType-class

Examples


B <- BinomFamily(size = 25, prob = 0.25) 

## classical optimal IC
IC0 <- optIC(model = B, risk = asCov())
plot(IC0) # plot IC
checkIC(IC0, B)

[Package ROptEst version 0.7 Index]