plot-methods(RobAStBase)	R Documentation

Methods for Function plot in Package ‘RobAStBase’

Description

plot-methods

Usage

plot(x, y, ...)
## S4 method for signature 'IC,missing':
plot(x, ..., withSweave = getdistrOption("withSweave"),
             main = FALSE, inner = TRUE, sub = FALSE,
             col.inner = par("col.main"), cex.inner = 0.8,
             bmar = par("mar")[1], tmar = par("mar")[3],
             mfColRow = TRUE, to.draw.arg = NULL)
## S4 method for signature 'IC,numeric':
plot(x, y, ..., cex.pts = 1,
                            col.pts = par("col"), pch.pts = 1, jitter.fac = 1,
                            with.lab = FALSE, lab.pts = NULL, lab.font = NULL,
             which.lbs = NULL, which.Order  = NULL, return.Order = FALSE)

Arguments

x 	 object of class "IC": IC to be plotted

y 	 missing or numeric (a dataset, e.g.)

withSweave 	 logical: if TRUE (for working with Sweave) no extra device is opened

main 	 logical: is a main title to be used? or
just as argument main in plot.default.

inner 	 logical: do panels have their own titles? or
character vector of / cast to length 'number of plotted dimensions'; if argument to.draw.arg is used, this refers to a vector of length length(to.draw.arg), the actually plotted dimensions. For further information, see also description of argument main in plot.default.

sub 	 logical: is a sub-title to be used? or
just as argument sub in plot.default.

tmar 	 top margin – useful for non-standard main title sizes

bmar 	 bottom margin – useful for non-standard sub title sizes

cex.inner 	 magnification to be used for inner titles relative to the current setting of cex; as in par

col.inner 	 character or integer code; color for the inner title

mfColRow 	 shall default partition in panels be used — defaults to TRUE

to.draw.arg 	 Either NULL (default; everything is plotted) or a vector of either integers (the indices of the subplots to be drawn) or characters — the names of the subplots to be drawn: these names are to be chosen either among the row names of the trafo matrix rownames(trafo(eval(x@CallL2Fam)@param)) or if the last expression is NULL a vector "dim<dimnr>", dimnr running through the number of rows of the trafo matrix.

cex.pts 	 size of the points of the second argument plotted

col.pts 	 color of the points of the second argument plotted

pch.pts 	 symbol of the points of the second argument plotted

with.lab 	 logical; shall labels be plotted to the observations?

lab.pts 	 character or NULL; labels to be plotted to the observations; if NULL observation indices;

lab.font 	 font to be used for labels

jitter.fac 	 jittering factor used in case of a DiscreteDistribution for plotting points of the second argument in a jittered fashion.

which.lbs 	 either an integer vector with the indices of the observations to be plotted into graph or NULL — then no observation is excluded

which.Order 	 we order the observations according to the norm given by normtype(object); then which.Order either is an integer vector with the indices of the ordered observations (remaining after a possible reduction by argument which.lbs) to be plotted into graph or NULL — then no (further) observation is excluded.

return.Order 	 logical; if TRUE, an order vector is returned; more specifically, the order of the (remaining) observations given by their original index is returned (remaining means: after a possible reduction by argument which.lbs, and ordering is according to the norm given by normtype(object)); othervise we return invisible() as usual.

... 	 further parameters for plot

Details

Any parameters of plot.default may be passed on to this particular plot method.

We start describing the IC,missing-method: For main-, inner, and subtitles given as arguments main, inner, and sub, top and bottom margins are enlarged to 5 resp. 6 by default but may also be specified by tmar / bmar arguments. If main / inner / sub are logical then if the respective argument is FALSE nothing is done/plotted, but if it is TRUE, we use a default main title taking up the calling arguments in case of main, default inner titles taking up the class and (named) parameter slots of arguments in case of inner, and a "generated on <data>"-tag in case of sub. Of course, if main / inner / sub are character, this is used for the title; in case of inner it is then checked whether it has correct length. In all title arguments, the following patterns are substituted:

"%C"

class of argument object

"%A"

deparsed argument object

"%D"

time/date-string when the plot was generated

If argument ... contains argument ylim, this may either be as in plot.default (i.e. a vector of length 2) or a vector of length 2*(number of plotted dimensions + 2), where the first two elements are the values for ylim in panel "d", the first two are for ylim resp. xlim for panels "p" and "q", and the last 2*(number of plotted dimensions) are the values for ylim for the plotted dimensions of the L2derivative, one pair for each dimension.

The IC,numeric-method calls the IC,missing-method but in addition plots the values of a dataset into the IC.

Examples


IC1 <- new("IC")
plot(IC1)
plot(IC1, main = TRUE, panel.first= grid(),
     col = "blue", cex.main = 2, cex.inner = 1)

### selection of subpanels for plotting
N <- NormLocationScaleFamily(mean=0, sd=1) 
IC2 <- optIC(model = N, risk = asCov())
par(mfrow=c(1,1))
plot(IC2, main = TRUE, panel.first= grid(),
     col = "blue", cex.main = 2, cex.inner = 0.6,
     mfColRow = FALSE, to.draw.arg=c("sd"))

## xlim and ylim arguments
plot(IC2, main = TRUE, panel.first= grid(), 
     ylim=c(-3,3), xlim=c(-2,3))
plot(IC2, main = TRUE, panel.first= grid(), 
     ylim=c(-3,3,-1,3), xlim=c(-2,3))

data <- r(N)(30)
plot(IC2, data, panel.first= grid(),
     ylim = c(-3,3,-1,3), xlim=c(-2,3),
     cex.pts = 3, pch.pts = 1:2, col.pts="green",
     with.lab = TRUE, which.lbs = c(1:4,15:20),
     which.Order = 1:6, return.Order = TRUE)

[Package RobAStBase version 0.7 Index]